aidemo refactoring.
