+++ package/ai/code/retinaface_mb_320/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(SRC main.cc anchors_320.cc retinaface.cc ../common/k510_drm.c cv2_utils.cc ../common/v4l2.c)
-set(bin retinaface_mb_320_opencv)
-set(script ../shell/retinaface_mb_320_opencv_bf16.sh ../shell/retinaface_mb_320_opencv_uint8.sh)
+set(bin retinaface_mb_320)
+set(script ../shell/retinaface_mb_320_bf16.sh ../shell/retinaface_mb_320_uint8.sh)
 set(input ../imx219_0.conf ../imx219_1.conf ../video_192x320.conf)
 
 add_definitions(-DFULL_SCREEN)
