--- package/ai/code/retinaface_mb_320_v4l2/anchors_320.cc
@@ -1,1 +1,0 @@
-float anchors320[4200][4] = {{0.012500,0.012500,0.050000,0.050000},{0.012500,0.012500,0.100000,0.100000},{0.037500,0.012500,0.050000,0.050000},{0.037500,0.012500,0.100000,0.100000},{0.062500,0.012500,0.050000,0.050000},{0.062500,0.012500,0.100000,0.100000},{0.087500,0.012500,0.050000,0.050000},{0.087500,0.012500,0.100000,0.100000},{0.112500,0.012500,0.050000,0.050000},{0.112500,0.012500,0.100000,0.100000},{0.137500,0.012500,0.050000,0.050000},{0.137500,0.012500,0.100000,0.100000},{0.162500,0.012500,0.050000,0.050000},{0.162500,0.012500,0.100000,0.100000},{0.187500,0.012500,0.050000,0.050000},{0.187500,0.012500,0.100000,0.100000},{0.212500,0.012500,0.050000,0.050000},{0.212500,0.012500,0.100000,0.100000},{0.237500,0.012500,0.050000,0.050000},{0.237500,0.012500,0.100000,0.100000},{0.262500,0.012500,0.050000,0.050000},{0.262500,0.012500,0.100000,0.100000},{0.287500,0.012500,0.050000,0.050000},{0.287500,0.012500,0.100000,0.100000},{0.312500,0.012500,0.050000,0.050000},{0.312500,0.012500,0.100000,0.100000},{0.337500,0.012500,0.050000,0.050000},{0.337500,0.012500,0.100000,0.100000},{0.362500,0.012500,0.050000,0.050000},{0.362500,0.012500,0.100000,0.100000},{0.387500,0.012500,0.050000,0.050000},{0.387500,0.012500,0.100000,0.100000},{0.412500,0.012500,0.050000,0.050000},{0.412500,0.012500,0.100000,0.100000},{0.437500,0.012500,0.050000,0.050000},{0.437500,0.012500,0.100000,0.100000},{0.462500,0.012500,0.050000,0.050000},{0.462500,0.012500,0.100000,0.100000},{0.487500,0.012500,0.050000,0.050000},{0.487500,0.012500,0.100000,0.100000},{0.512500,0.012500,0.050000,0.050000},{0.512500,0.012500,0.100000,0.100000},{0.537500,0.012500,0.050000,0.050000},{0.537500,0.012500,0.100000,0.100000},{0.562500,0.012500,0.050000,0.050000},{0.562500,0.012500,0.100000,0.100000},{0.587500,0.012500,0.050000,0.050000},{0.587500,0.012500,0.100000,0.100000},{0.612500,0.012500,0.050000,0.050000},{0.612500,0.012500,0.100000,0.100000},{0.637500,0.012500,0.050000,0.050000},{0.637500,0.012500,0.100000,0.100000},{0.662500,0.012500,0.050000,0.050000},{0.662500,0.012500,0.100000,0.100000},{0.687500,0.012500,0.050000,0.050000},{0.687500,0.012500,0.100000,0.100000},{0.712500,0.012500,0.050000,0.050000},{0.712500,0.012500,0.100000,0.100000},{0.737500,0.012500,0.050000,0.050000},{0.737500,0.012500,0.100000,0.100000},{0.762500,0.012500,0.050000,0.050000},{0.762500,0.012500,0.100000,0.100000},{0.787500,0.012500,0.050000,0.050000},{0.787500,0.012500,0.100000,0.100000},{0.812500,0.012500,0.050000,0.050000},{0.812500,0.012500,0.100000,0.100000},{0.837500,0.012500,0.050000,0.050000},{0.837500,0.012500,0.100000,0.100000},{0.862500,0.012500,0.050000,0.050000},{0.862500,0.012500,0.100000,0.100000},{0.887500,0.012500,0.050000,0.050000},{0.887500,0.012500,0.100000,0.100000},{0.912500,0.012500,0.050000,0.050000},{0.912500,0.012500,0.100000,0.100000},{0.937500,0.012500,0.050000,0.050000},{0.937500,0.012500,0.100000,0.100000},{0.962500,0.012500,0.050000,0.050000},{0.962500,0.012500,0.100000,0.100000},{0.987500,0.012500,0.050000,0.050000},{0.987500,0.012500,0.100000,0.100000},{0.012500,0.037500,0.050000,0.050000},{0.012500,0.037500,0.100000,0.100000},{0.037500,0.037500,0.050000,0.050000},{0.037500,0.037500,0.100000,0.100000},{0.062500,0.037500,0.050000,0.050000},{0.062500,0.037500,0.100000,0.100000},{0.087500,0.037500,0.050000,0.050000},{0.087500,0.037500,0.100000,0.100000},{0.112500,0.037500,0.050000,0.050000},{0.112500,0.037500,0.100000,0.100000},{0.137500,0.037500,0.050000,0.050000},{0.137500,0.037500,0.100000,0.100000},{0.162500,0.037500,0.050000,0.050000},{0.162500,0.037500,0.100000,0.100000},{0.187500,0.037500,0.050000,0.050000},{0.187500,0.037500,0.100000,0.100000},{0.212500,0.037500,0.050000,0.050000},{0.212500,0.037500,0.100000,0.100000},{0.237500,0.037500,0.050000,0.050000},{0.237500,0.037500,0.100000,0.100000},{0.262500,0.037500,0.050000,0.050000},{0.262500,0.037500,0.100000,0.100000},{0.287500,0.037500,0.050000,0.050000},{0.287500,0.037500,0.100000,0.100000},{0.312500,0.037500,0.050000,0.050000},{0.312500,0.037500,0.100000,0.100000},{0.337500,0.037500,0.050000,0.050000},{0.337500,0.037500,0.100000,0.100000},{0.362500,0.037500,0.050000,0.050000},{0.362500,0.037500,0.100000,0.100000},{0.387500,0.037500,0.050000,0.050000},{0.387500,0.037500,0.100000,0.100000},{0.412500,0.037500,0.050000,0.050000},{0.412500,0.037500,0.100000,0.100000},{0.437500,0.037500,0.050000,0.050000},{0.437500,0.037500,0.100000,0.100000},{0.462500,0.037500,0.050000,0.050000},{0.462500,0.037500,0.100000,0.100000},{0.487500,0.037500,0.050000,0.050000},{0.487500,0.037500,0.100000,0.100000},{0.512500,0.037500,0.050000,0.050000},{0.512500,0.037500,0.100000,0.100000},{0.537500,0.037500,0.050000,0.050000},{0.537500,0.037500,0.100000,0.100000},{0.562500,0.037500,0.050000,0.050000},{0.562500,0.037500,0.100000,0.100000},{0.587500,0.037500,0.050000,0.050000},{0.587500,0.037500,0.100000,0.100000},{0.612500,0.037500,0.050000,0.050000},{0.612500,0.037500,0.100000,0.100000},{0.637500,0.037500,0.050000,0.050000},{0.637500,0.037500,0.100000,0.100000},{0.662500,0.037500,0.050000,0.050000},{0.662500,0.037500,0.100000,0.100000},{0.687500,0.037500,0.050000,0.050000},{0.687500,0.037500,0.100000,0.100000},{0.712500,0.037500,0.050000,0.050000},{0.712500,0.037500,0.100000,0.100000},{0.737500,0.037500,0.050000,0.050000},{0.737500,0.037500,0.100000,0.100000},{0.762500,0.037500,0.050000,0.050000},{0.762500,0.037500,0.100000,0.100000},{0.787500,0.037500,0.050000,0.050000},{0.787500,0.037500,0.100000,0.100000},{0.812500,0.037500,0.050000,0.050000},{0.812500,0.037500,0.100000,0.100000},{0.837500,0.037500,0.050000,0.050000},{0.837500,0.037500,0.100000,0.100000},{0.862500,0.037500,0.050000,0.050000},{0.862500,0.037500,0.100000,0.100000},{0.887500,0.037500,0.050000,0.050000},{0.887500,0.037500,0.100000,0.100000},{0.912500,0.037500,0.050000,0.050000},{0.912500,0.037500,0.100000,0.100000},{0.937500,0.037500,0.050000,0.050000},{0.937500,0.037500,0.100000,0.100000},{0.962500,0.037500,0.050000,0.050000},{0.962500,0.037500,0.100000,0.100000},{0.987500,0.037500,0.050000,0.050000},{0.987500,0.037500,0.100000,0.100000},{0.012500,0.062500,0.050000,0.050000},{0.012500,0.062500,0.100000,0.100000},{0.037500,0.062500,0.050000,0.050000},{0.037500,0.062500,0.100000,0.100000},{0.062500,0.062500,0.050000,0.050000},{0.062500,0.062500,0.100000,0.100000},{0.087500,0.062500,0.050000,0.050000},{0.087500,0.062500,0.100000,0.100000},{0.112500,0.062500,0.050000,0.050000},{0.112500,0.062500,0.100000,0.100000},{0.137500,0.062500,0.050000,0.050000},{0.137500,0.062500,0.100000,0.100000},{0.162500,0.062500,0.050000,0.050000},{0.162500,0.062500,0.100000,0.100000},{0.187500,0.062500,0.050000,0.050000},{0.187500,0.062500,0.100000,0.100000},{0.212500,0.062500,0.050000,0.050000},{0.212500,0.062500,0.100000,0.100000},{0.237500,0.062500,0.050000,0.050000},{0.237500,0.062500,0.100000,0.100000},{0.262500,0.062500,0.050000,0.050000},{0.262500,0.062500,0.100000,0.100000},{0.287500,0.062500,0.050000,0.050000},{0.287500,0.062500,0.100000,0.100000},{0.312500,0.062500,0.050000,0.050000},{0.312500,0.062500,0.100000,0.100000},{0.337500,0.062500,0.050000,0.050000},{0.337500,0.062500,0.100000,0.100000},{0.362500,0.062500,0.050000,0.050000},{0.362500,0.062500,0.100000,0.100000},{0.387500,0.062500,0.050000,0.050000},{0.387500,0.062500,0.100000,0.100000},{0.412500,0.062500,0.050000,0.050000},{0.412500,0.062500,0.100000,0.100000},{0.437500,0.062500,0.050000,0.050000},{0.437500,0.062500,0.100000,0.100000},{0.462500,0.062500,0.050000,0.050000},{0.462500,0.062500,0.100000,0.100000},{0.487500,0.062500,0.050000,0.050000},{0.487500,0.062500,0.100000,0.100000},{0.512500,0.062500,0.050000,0.050000},{0.512500,0.062500,0.100000,0.100000},{0.537500,0.062500,0.050000,0.050000},{0.537500,0.062500,0.100000,0.100000},{0.562500,0.062500,0.050000,0.050000},{0.562500,0.062500,0.100000,0.100000},{0.587500,0.062500,0.050000,0.050000},{0.587500,0.062500,0.100000,0.100000},{0.612500,0.062500,0.050000,0.050000},{0.612500,0.062500,0.100000,0.100000},{0.637500,0.062500,0.050000,0.050000},{0.637500,0.062500,0.100000,0.100000},{0.662500,0.062500,0.050000,0.050000},{0.662500,0.062500,0.100000,0.100000},{0.687500,0.062500,0.050000,0.050000},{0.687500,0.062500,0.100000,0.100000},{0.712500,0.062500,0.050000,0.050000},{0.712500,0.062500,0.100000,0.100000},{0.737500,0.062500,0.050000,0.050000},{0.737500,0.062500,0.100000,0.100000},{0.762500,0.062500,0.050000,0.050000},{0.762500,0.062500,0.100000,0.100000},{0.787500,0.062500,0.050000,0.050000},{0.787500,0.062500,0.100000,0.100000},{0.812500,0.062500,0.050000,0.050000},{0.812500,0.062500,0.100000,0.100000},{0.837500,0.062500,0.050000,0.050000},{0.837500,0.062500,0.100000,0.100000},{0.862500,0.062500,0.050000,0.050000},{0.862500,0.062500,0.100000,0.100000},{0.887500,0.062500,0.050000,0.050000},{0.887500,0.062500,0.100000,0.100000},{0.912500,0.062500,0.050000,0.050000},{0.912500,0.062500,0.100000,0.100000},{0.937500,0.062500,0.050000,0.050000},{0.937500,0.062500,0.100000,0.100000},{0.962500,0.062500,0.050000,0.050000},{0.962500,0.062500,0.100000,0.100000},{0.987500,0.062500,0.050000,0.050000},{0.987500,0.062500,0.100000,0.100000},{0.012500,0.087500,0.050000,0.050000},{0.012500,0.087500,0.100000,0.100000},{0.037500,0.087500,0.050000,0.050000},{0.037500,0.087500,0.100000,0.100000},{0.062500,0.087500,0.050000,0.050000},{0.062500,0.087500,0.100000,0.100000},{0.087500,0.087500,0.050000,0.050000},{0.087500,0.087500,0.100000,0.100000},{0.112500,0.087500,0.050000,0.050000},{0.112500,0.087500,0.100000,0.100000},{0.137500,0.087500,0.050000,0.050000},{0.137500,0.087500,0.100000,0.100000},{0.162500,0.087500,0.050000,0.050000},{0.162500,0.087500,0.100000,0.100000},{0.187500,0.087500,0.050000,0.050000},{0.187500,0.087500,0.100000,0.100000},{0.212500,0.087500,0.050000,0.050000},{0.212500,0.087500,0.100000,0.100000},{0.237500,0.087500,0.050000,0.050000},{0.237500,0.087500,0.100000,0.100000},{0.262500,0.087500,0.050000,0.050000},{0.262500,0.087500,0.100000,0.100000},{0.287500,0.087500,0.050000,0.050000},{0.287500,0.087500,0.100000,0.100000},{0.312500,0.087500,0.050000,0.050000},{0.312500,0.087500,0.100000,0.100000},{0.337500,0.087500,0.050000,0.050000},{0.337500,0.087500,0.100000,0.100000},{0.362500,0.087500,0.050000,0.050000},{0.362500,0.087500,0.100000,0.100000},{0.387500,0.087500,0.050000,0.050000},{0.387500,0.087500,0.100000,0.100000},{0.412500,0.087500,0.050000,0.050000},{0.412500,0.087500,0.100000,0.100000},{0.437500,0.087500,0.050000,0.050000},{0.437500,0.087500,0.100000,0.100000},{0.462500,0.087500,0.050000,0.050000},{0.462500,0.087500,0.100000,0.100000},{0.487500,0.087500,0.050000,0.050000},{0.487500,0.087500,0.100000,0.100000},{0.512500,0.087500,0.050000,0.050000},{0.512500,0.087500,0.100000,0.100000},{0.537500,0.087500,0.050000,0.050000},{0.537500,0.087500,0.100000,0.100000},{0.562500,0.087500,0.050000,0.050000},{0.562500,0.087500,0.100000,0.100000},{0.587500,0.087500,0.050000,0.050000},{0.587500,0.087500,0.100000,0.100000},{0.612500,0.087500,0.050000,0.050000},{0.612500,0.087500,0.100000,0.100000},{0.637500,0.087500,0.050000,0.050000},{0.637500,0.087500,0.100000,0.100000},{0.662500,0.087500,0.050000,0.050000},{0.662500,0.087500,0.100000,0.100000},{0.687500,0.087500,0.050000,0.050000},{0.687500,0.087500,0.100000,0.100000},{0.712500,0.087500,0.050000,0.050000},{0.712500,0.087500,0.100000,0.100000},{0.737500,0.087500,0.050000,0.050000},{0.737500,0.087500,0.100000,0.100000},{0.762500,0.087500,0.050000,0.050000},{0.762500,0.087500,0.100000,0.100000},{0.787500,0.087500,0.050000,0.050000},{0.787500,0.087500,0.100000,0.100000},{0.812500,0.087500,0.050000,0.050000},{0.812500,0.087500,0.100000,0.100000},{0.837500,0.087500,0.050000,0.050000},{0.837500,0.087500,0.100000,0.100000},{0.862500,0.087500,0.050000,0.050000},{0.862500,0.087500,0.100000,0.100000},{0.887500,0.087500,0.050000,0.050000},{0.887500,0.087500,0.100000,0.100000},{0.912500,0.087500,0.050000,0.050000},{0.912500,0.087500,0.100000,0.100000},{0.937500,0.087500,0.050000,0.050000},{0.937500,0.087500,0.100000,0.100000},{0.962500,0.087500,0.050000,0.050000},{0.962500,0.087500,0.100000,0.100000},{0.987500,0.087500,0.050000,0.050000},{0.987500,0.087500,0.100000,0.100000},{0.012500,0.112500,0.050000,0.050000},{0.012500,0.112500,0.100000,0.100000},{0.037500,0.112500,0.050000,0.050000},{0.037500,0.112500,0.100000,0.100000},{0.062500,0.112500,0.050000,0.050000},{0.062500,0.112500,0.100000,0.100000},{0.087500,0.112500,0.050000,0.050000},{0.087500,0.112500,0.100000,0.100000},{0.112500,0.112500,0.050000,0.050000},{0.112500,0.112500,0.100000,0.100000},{0.137500,0.112500,0.050000,0.050000},{0.137500,0.112500,0.100000,0.100000},{0.162500,0.112500,0.050000,0.050000},{0.162500,0.112500,0.100000,0.100000},{0.187500,0.112500,0.050000,0.050000},{0.187500,0.112500,0.100000,0.100000},{0.212500,0.112500,0.050000,0.050000},{0.212500,0.112500,0.100000,0.100000},{0.237500,0.112500,0.050000,0.050000},{0.237500,0.112500,0.100000,0.100000},{0.262500,0.112500,0.050000,0.050000},{0.262500,0.112500,0.100000,0.100000},{0.287500,0.112500,0.050000,0.050000},{0.287500,0.112500,0.100000,0.100000},{0.312500,0.112500,0.050000,0.050000},{0.312500,0.112500,0.100000,0.100000},{0.337500,0.112500,0.050000,0.050000},{0.337500,0.112500,0.100000,0.100000},{0.362500,0.112500,0.050000,0.050000},{0.362500,0.112500,0.100000,0.100000},{0.387500,0.112500,0.050000,0.050000},{0.387500,0.112500,0.100000,0.100000},{0.412500,0.112500,0.050000,0.050000},{0.412500,0.112500,0.100000,0.100000},{0.437500,0.112500,0.050000,0.050000},{0.437500,0.112500,0.100000,0.100000},{0.462500,0.112500,0.050000,0.050000},{0.462500,0.112500,0.100000,0.100000},{0.487500,0.112500,0.050000,0.050000},{0.487500,0.112500,0.100000,0.100000},{0.512500,0.112500,0.050000,0.050000},{0.512500,0.112500,0.100000,0.100000},{0.537500,0.112500,0.050000,0.050000},{0.537500,0.112500,0.100000,0.100000},{0.562500,0.112500,0.050000,0.050000},{0.562500,0.112500,0.100000,0.100000},{0.587500,0.112500,0.050000,0.050000},{0.587500,0.112500,0.100000,0.100000},{0.612500,0.112500,0.050000,0.050000},{0.612500,0.112500,0.100000,0.100000},{0.637500,0.112500,0.050000,0.050000},{0.637500,0.112500,0.100000,0.100000},{0.662500,0.112500,0.050000,0.050000},{0.662500,0.112500,0.100000,0.100000},{0.687500,0.112500,0.050000,0.050000},{0.687500,0.112500,0.100000,0.100000},{0.712500,0.112500,0.050000,0.050000},{0.712500,0.112500,0.100000,0.100000},{0.737500,0.112500,0.050000,0.050000},{0.737500,0.112500,0.100000,0.100000},{0.762500,0.112500,0.050000,0.050000},{0.762500,0.112500,0.100000,0.100000},{0.787500,0.112500,0.050000,0.050000},{0.787500,0.112500,0.100000,0.100000},{0.812500,0.112500,0.050000,0.050000},{0.812500,0.112500,0.100000,0.100000},{0.837500,0.112500,0.050000,0.050000},{0.837500,0.112500,0.100000,0.100000},{0.862500,0.112500,0.050000,0.050000},{0.862500,0.112500,0.100000,0.100000},{0.887500,0.112500,0.050000,0.050000},{0.887500,0.112500,0.100000,0.100000},{0.912500,0.112500,0.050000,0.050000},{0.912500,0.112500,0.100000,0.100000},{0.937500,0.112500,0.050000,0.050000},{0.937500,0.112500,0.100000,0.100000},{0.962500,0.112500,0.050000,0.050000},{0.962500,0.112500,0.100000,0.100000},{0.987500,0.112500,0.050000,0.050000},{0.987500,0.112500,0.100000,0.100000},{0.012500,0.137500,0.050000,0.050000},{0.012500,0.137500,0.100000,0.100000},{0.037500,0.137500,0.050000,0.050000},{0.037500,0.137500,0.100000,0.100000},{0.062500,0.137500,0.050000,0.050000},{0.062500,0.137500,0.100000,0.100000},{0.087500,0.137500,0.050000,0.050000},{0.087500,0.137500,0.100000,0.100000},{0.112500,0.137500,0.050000,0.050000},{0.112500,0.137500,0.100000,0.100000},{0.137500,0.137500,0.050000,0.050000},{0.137500,0.137500,0.100000,0.100000},{0.162500,0.137500,0.050000,0.050000},{0.162500,0.137500,0.100000,0.100000},{0.187500,0.137500,0.050000,0.050000},{0.187500,0.137500,0.100000,0.100000},{0.212500,0.137500,0.050000,0.050000},{0.212500,0.137500,0.100000,0.100000},{0.237500,0.137500,0.050000,0.050000},{0.237500,0.137500,0.100000,0.100000},{0.262500,0.137500,0.050000,0.050000},{0.262500,0.137500,0.100000,0.100000},{0.287500,0.137500,0.050000,0.050000},{0.287500,0.137500,0.100000,0.100000},{0.312500,0.137500,0.050000,0.050000},{0.312500,0.137500,0.100000,0.100000},{0.337500,0.137500,0.050000,0.050000},{0.337500,0.137500,0.100000,0.100000},{0.362500,0.137500,0.050000,0.050000},{0.362500,0.137500,0.100000,0.100000},{0.387500,0.137500,0.050000,0.050000},{0.387500,0.137500,0.100000,0.100000},{0.412500,0.137500,0.050000,0.050000},{0.412500,0.137500,0.100000,0.100000},{0.437500,0.137500,0.050000,0.050000},{0.437500,0.137500,0.100000,0.100000},{0.462500,0.137500,0.050000,0.050000},{0.462500,0.137500,0.100000,0.100000},{0.487500,0.137500,0.050000,0.050000},{0.487500,0.137500,0.100000,0.100000},{0.512500,0.137500,0.050000,0.050000},{0.512500,0.137500,0.100000,0.100000},{0.537500,0.137500,0.050000,0.050000},{0.537500,0.137500,0.100000,0.100000},{0.562500,0.137500,0.050000,0.050000},{0.562500,0.137500,0.100000,0.100000},{0.587500,0.137500,0.050000,0.050000},{0.587500,0.137500,0.100000,0.100000},{0.612500,0.137500,0.050000,0.050000},{0.612500,0.137500,0.100000,0.100000},{0.637500,0.137500,0.050000,0.050000},{0.637500,0.137500,0.100000,0.100000},{0.662500,0.137500,0.050000,0.050000},{0.662500,0.137500,0.100000,0.100000},{0.687500,0.137500,0.050000,0.050000},{0.687500,0.137500,0.100000,0.100000},{0.712500,0.137500,0.050000,0.050000},{0.712500,0.137500,0.100000,0.100000},{0.737500,0.137500,0.050000,0.050000},{0.737500,0.137500,0.100000,0.100000},{0.762500,0.137500,0.050000,0.050000},{0.762500,0.137500,0.100000,0.100000},{0.787500,0.137500,0.050000,0.050000},{0.787500,0.137500,0.100000,0.100000},{0.812500,0.137500,0.050000,0.050000},{0.812500,0.137500,0.100000,0.100000},{0.837500,0.137500,0.050000,0.050000},{0.837500,0.137500,0.100000,0.100000},{0.862500,0.137500,0.050000,0.050000},{0.862500,0.137500,0.100000,0.100000},{0.887500,0.137500,0.050000,0.050000},{0.887500,0.137500,0.100000,0.100000},{0.912500,0.137500,0.050000,0.050000},{0.912500,0.137500,0.100000,0.100000},{0.937500,0.137500,0.050000,0.050000},{0.937500,0.137500,0.100000,0.100000},{0.962500,0.137500,0.050000,0.050000},{0.962500,0.137500,0.100000,0.100000},{0.987500,0.137500,0.050000,0.050000},{0.987500,0.137500,0.100000,0.100000},{0.012500,0.162500,0.050000,0.050000},{0.012500,0.162500,0.100000,0.100000},{0.037500,0.162500,0.050000,0.050000},{0.037500,0.162500,0.100000,0.100000},{0.062500,0.162500,0.050000,0.050000},{0.062500,0.162500,0.100000,0.100000},{0.087500,0.162500,0.050000,0.050000},{0.087500,0.162500,0.100000,0.100000},{0.112500,0.162500,0.050000,0.050000},{0.112500,0.162500,0.100000,0.100000},{0.137500,0.162500,0.050000,0.050000},{0.137500,0.162500,0.100000,0.100000},{0.162500,0.162500,0.050000,0.050000},{0.162500,0.162500,0.100000,0.100000},{0.187500,0.162500,0.050000,0.050000},{0.187500,0.162500,0.100000,0.100000},{0.212500,0.162500,0.050000,0.050000},{0.212500,0.162500,0.100000,0.100000},{0.237500,0.162500,0.050000,0.050000},{0.237500,0.162500,0.100000,0.100000},{0.262500,0.162500,0.050000,0.050000},{0.262500,0.162500,0.100000,0.100000},{0.287500,0.162500,0.050000,0.050000},{0.287500,0.162500,0.100000,0.100000},{0.312500,0.162500,0.050000,0.050000},{0.312500,0.162500,0.100000,0.100000},{0.337500,0.162500,0.050000,0.050000},{0.337500,0.162500,0.100000,0.100000},{0.362500,0.162500,0.050000,0.050000},{0.362500,0.162500,0.100000,0.100000},{0.387500,0.162500,0.050000,0.050000},{0.387500,0.162500,0.100000,0.100000},{0.412500,0.162500,0.050000,0.050000},{0.412500,0.162500,0.100000,0.100000},{0.437500,0.162500,0.050000,0.050000},{0.437500,0.162500,0.100000,0.100000},{0.462500,0.162500,0.050000,0.050000},{0.462500,0.162500,0.100000,0.100000},{0.487500,0.162500,0.050000,0.050000},{0.487500,0.162500,0.100000,0.100000},{0.512500,0.162500,0.050000,0.050000},{0.512500,0.162500,0.100000,0.100000},{0.537500,0.162500,0.050000,0.050000},{0.537500,0.162500,0.100000,0.100000},{0.562500,0.162500,0.050000,0.050000},{0.562500,0.162500,0.100000,0.100000},{0.587500,0.162500,0.050000,0.050000},{0.587500,0.162500,0.100000,0.100000},{0.612500,0.162500,0.050000,0.050000},{0.612500,0.162500,0.100000,0.100000},{0.637500,0.162500,0.050000,0.050000},{0.637500,0.162500,0.100000,0.100000},{0.662500,0.162500,0.050000,0.050000},{0.662500,0.162500,0.100000,0.100000},{0.687500,0.162500,0.050000,0.050000},{0.687500,0.162500,0.100000,0.100000},{0.712500,0.162500,0.050000,0.050000},{0.712500,0.162500,0.100000,0.100000},{0.737500,0.162500,0.050000,0.050000},{0.737500,0.162500,0.100000,0.100000},{0.762500,0.162500,0.050000,0.050000},{0.762500,0.162500,0.100000,0.100000},{0.787500,0.162500,0.050000,0.050000},{0.787500,0.162500,0.100000,0.100000},{0.812500,0.162500,0.050000,0.050000},{0.812500,0.162500,0.100000,0.100000},{0.837500,0.162500,0.050000,0.050000},{0.837500,0.162500,0.100000,0.100000},{0.862500,0.162500,0.050000,0.050000},{0.862500,0.162500,0.100000,0.100000},{0.887500,0.162500,0.050000,0.050000},{0.887500,0.162500,0.100000,0.100000},{0.912500,0.162500,0.050000,0.050000},{0.912500,0.162500,0.100000,0.100000},{0.937500,0.162500,0.050000,0.050000},{0.937500,0.162500,0.100000,0.100000},{0.962500,0.162500,0.050000,0.050000},{0.962500,0.162500,0.100000,0.100000},{0.987500,0.162500,0.050000,0.050000},{0.987500,0.162500,0.100000,0.100000},{0.012500,0.187500,0.050000,0.050000},{0.012500,0.187500,0.100000,0.100000},{0.037500,0.187500,0.050000,0.050000},{0.037500,0.187500,0.100000,0.100000},{0.062500,0.187500,0.050000,0.050000},{0.062500,0.187500,0.100000,0.100000},{0.087500,0.187500,0.050000,0.050000},{0.087500,0.187500,0.100000,0.100000},{0.112500,0.187500,0.050000,0.050000},{0.112500,0.187500,0.100000,0.100000},{0.137500,0.187500,0.050000,0.050000},{0.137500,0.187500,0.100000,0.100000},{0.162500,0.187500,0.050000,0.050000},{0.162500,0.187500,0.100000,0.100000},{0.187500,0.187500,0.050000,0.050000},{0.187500,0.187500,0.100000,0.100000},{0.212500,0.187500,0.050000,0.050000},{0.212500,0.187500,0.100000,0.100000},{0.237500,0.187500,0.050000,0.050000},{0.237500,0.187500,0.100000,0.100000},{0.262500,0.187500,0.050000,0.050000},{0.262500,0.187500,0.100000,0.100000},{0.287500,0.187500,0.050000,0.050000},{0.287500,0.187500,0.100000,0.100000},{0.312500,0.187500,0.050000,0.050000},{0.312500,0.187500,0.100000,0.100000},{0.337500,0.187500,0.050000,0.050000},{0.337500,0.187500,0.100000,0.100000},{0.362500,0.187500,0.050000,0.050000},{0.362500,0.187500,0.100000,0.100000},{0.387500,0.187500,0.050000,0.050000},{0.387500,0.187500,0.100000,0.100000},{0.412500,0.187500,0.050000,0.050000},{0.412500,0.187500,0.100000,0.100000},{0.437500,0.187500,0.050000,0.050000},{0.437500,0.187500,0.100000,0.100000},{0.462500,0.187500,0.050000,0.050000},{0.462500,0.187500,0.100000,0.100000},{0.487500,0.187500,0.050000,0.050000},{0.487500,0.187500,0.100000,0.100000},{0.512500,0.187500,0.050000,0.050000},{0.512500,0.187500,0.100000,0.100000},{0.537500,0.187500,0.050000,0.050000},{0.537500,0.187500,0.100000,0.100000},{0.562500,0.187500,0.050000,0.050000},{0.562500,0.187500,0.100000,0.100000},{0.587500,0.187500,0.050000,0.050000},{0.587500,0.187500,0.100000,0.100000},{0.612500,0.187500,0.050000,0.050000},{0.612500,0.187500,0.100000,0.100000},{0.637500,0.187500,0.050000,0.050000},{0.637500,0.187500,0.100000,0.100000},{0.662500,0.187500,0.050000,0.050000},{0.662500,0.187500,0.100000,0.100000},{0.687500,0.187500,0.050000,0.050000},{0.687500,0.187500,0.100000,0.100000},{0.712500,0.187500,0.050000,0.050000},{0.712500,0.187500,0.100000,0.100000},{0.737500,0.187500,0.050000,0.050000},{0.737500,0.187500,0.100000,0.100000},{0.762500,0.187500,0.050000,0.050000},{0.762500,0.187500,0.100000,0.100000},{0.787500,0.187500,0.050000,0.050000},{0.787500,0.187500,0.100000,0.100000},{0.812500,0.187500,0.050000,0.050000},{0.812500,0.187500,0.100000,0.100000},{0.837500,0.187500,0.050000,0.050000},{0.837500,0.187500,0.100000,0.100000},{0.862500,0.187500,0.050000,0.050000},{0.862500,0.187500,0.100000,0.100000},{0.887500,0.187500,0.050000,0.050000},{0.887500,0.187500,0.100000,0.100000},{0.912500,0.187500,0.050000,0.050000},{0.912500,0.187500,0.100000,0.100000},{0.937500,0.187500,0.050000,0.050000},{0.937500,0.187500,0.100000,0.100000},{0.962500,0.187500,0.050000,0.050000},{0.962500,0.187500,0.100000,0.100000},{0.987500,0.187500,0.050000,0.050000},{0.987500,0.187500,0.100000,0.100000},{0.012500,0.212500,0.050000,0.050000},{0.012500,0.212500,0.100000,0.100000},{0.037500,0.212500,0.050000,0.050000},{0.037500,0.212500,0.100000,0.100000},{0.062500,0.212500,0.050000,0.050000},{0.062500,0.212500,0.100000,0.100000},{0.087500,0.212500,0.050000,0.050000},{0.087500,0.212500,0.100000,0.100000},{0.112500,0.212500,0.050000,0.050000},{0.112500,0.212500,0.100000,0.100000},{0.137500,0.212500,0.050000,0.050000},{0.137500,0.212500,0.100000,0.100000},{0.162500,0.212500,0.050000,0.050000},{0.162500,0.212500,0.100000,0.100000},{0.187500,0.212500,0.050000,0.050000},{0.187500,0.212500,0.100000,0.100000},{0.212500,0.212500,0.050000,0.050000},{0.212500,0.212500,0.100000,0.100000},{0.237500,0.212500,0.050000,0.050000},{0.237500,0.212500,0.100000,0.100000},{0.262500,0.212500,0.050000,0.050000},{0.262500,0.212500,0.100000,0.100000},{0.287500,0.212500,0.050000,0.050000},{0.287500,0.212500,0.100000,0.100000},{0.312500,0.212500,0.050000,0.050000},{0.312500,0.212500,0.100000,0.100000},{0.337500,0.212500,0.050000,0.050000},{0.337500,0.212500,0.100000,0.100000},{0.362500,0.212500,0.050000,0.050000},{0.362500,0.212500,0.100000,0.100000},{0.387500,0.212500,0.050000,0.050000},{0.387500,0.212500,0.100000,0.100000},{0.412500,0.212500,0.050000,0.050000},{0.412500,0.212500,0.100000,0.100000},{0.437500,0.212500,0.050000,0.050000},{0.437500,0.212500,0.100000,0.100000},{0.462500,0.212500,0.050000,0.050000},{0.462500,0.212500,0.100000,0.100000},{0.487500,0.212500,0.050000,0.050000},{0.487500,0.212500,0.100000,0.100000},{0.512500,0.212500,0.050000,0.050000},{0.512500,0.212500,0.100000,0.100000},{0.537500,0.212500,0.050000,0.050000},{0.537500,0.212500,0.100000,0.100000},{0.562500,0.212500,0.050000,0.050000},{0.562500,0.212500,0.100000,0.100000},{0.587500,0.212500,0.050000,0.050000},{0.587500,0.212500,0.100000,0.100000},{0.612500,0.212500,0.050000,0.050000},{0.612500,0.212500,0.100000,0.100000},{0.637500,0.212500,0.050000,0.050000},{0.637500,0.212500,0.100000,0.100000},{0.662500,0.212500,0.050000,0.050000},{0.662500,0.212500,0.100000,0.100000},{0.687500,0.212500,0.050000,0.050000},{0.687500,0.212500,0.100000,0.100000},{0.712500,0.212500,0.050000,0.050000},{0.712500,0.212500,0.100000,0.100000},{0.737500,0.212500,0.050000,0.050000},{0.737500,0.212500,0.100000,0.100000},{0.762500,0.212500,0.050000,0.050000},{0.762500,0.212500,0.100000,0.100000},{0.787500,0.212500,0.050000,0.050000},{0.787500,0.212500,0.100000,0.100000},{0.812500,0.212500,0.050000,0.050000},{0.812500,0.212500,0.100000,0.100000},{0.837500,0.212500,0.050000,0.050000},{0.837500,0.212500,0.100000,0.100000},{0.862500,0.212500,0.050000,0.050000},{0.862500,0.212500,0.100000,0.100000},{0.887500,0.212500,0.050000,0.050000},{0.887500,0.212500,0.100000,0.100000},{0.912500,0.212500,0.050000,0.050000},{0.912500,0.212500,0.100000,0.100000},{0.937500,0.212500,0.050000,0.050000},{0.937500,0.212500,0.100000,0.100000},{0.962500,0.212500,0.050000,0.050000},{0.962500,0.212500,0.100000,0.100000},{0.987500,0.212500,0.050000,0.050000},{0.987500,0.212500,0.100000,0.100000},{0.012500,0.237500,0.050000,0.050000},{0.012500,0.237500,0.100000,0.100000},{0.037500,0.237500,0.050000,0.050000},{0.037500,0.237500,0.100000,0.100000},{0.062500,0.237500,0.050000,0.050000},{0.062500,0.237500,0.100000,0.100000},{0.087500,0.237500,0.050000,0.050000},{0.087500,0.237500,0.100000,0.100000},{0.112500,0.237500,0.050000,0.050000},{0.112500,0.237500,0.100000,0.100000},{0.137500,0.237500,0.050000,0.050000},{0.137500,0.237500,0.100000,0.100000},{0.162500,0.237500,0.050000,0.050000},{0.162500,0.237500,0.100000,0.100000},{0.187500,0.237500,0.050000,0.050000},{0.187500,0.237500,0.100000,0.100000},{0.212500,0.237500,0.050000,0.050000},{0.212500,0.237500,0.100000,0.100000},{0.237500,0.237500,0.050000,0.050000},{0.237500,0.237500,0.100000,0.100000},{0.262500,0.237500,0.050000,0.050000},{0.262500,0.237500,0.100000,0.100000},{0.287500,0.237500,0.050000,0.050000},{0.287500,0.237500,0.100000,0.100000},{0.312500,0.237500,0.050000,0.050000},{0.312500,0.237500,0.100000,0.100000},{0.337500,0.237500,0.050000,0.050000},{0.337500,0.237500,0.100000,0.100000},{0.362500,0.237500,0.050000,0.050000},{0.362500,0.237500,0.100000,0.100000},{0.387500,0.237500,0.050000,0.050000},{0.387500,0.237500,0.100000,0.100000},{0.412500,0.237500,0.050000,0.050000},{0.412500,0.237500,0.100000,0.100000},{0.437500,0.237500,0.050000,0.050000},{0.437500,0.237500,0.100000,0.100000},{0.462500,0.237500,0.050000,0.050000},{0.462500,0.237500,0.100000,0.100000},{0.487500,0.237500,0.050000,0.050000},{0.487500,0.237500,0.100000,0.100000},{0.512500,0.237500,0.050000,0.050000},{0.512500,0.237500,0.100000,0.100000},{0.537500,0.237500,0.050000,0.050000},{0.537500,0.237500,0.100000,0.100000},{0.562500,0.237500,0.050000,0.050000},{0.562500,0.237500,0.100000,0.100000},{0.587500,0.237500,0.050000,0.050000},{0.587500,0.237500,0.100000,0.100000},{0.612500,0.237500,0.050000,0.050000},{0.612500,0.237500,0.100000,0.100000},{0.637500,0.237500,0.050000,0.050000},{0.637500,0.237500,0.100000,0.100000},{0.662500,0.237500,0.050000,0.050000},{0.662500,0.237500,0.100000,0.100000},{0.687500,0.237500,0.050000,0.050000},{0.687500,0.237500,0.100000,0.100000},{0.712500,0.237500,0.050000,0.050000},{0.712500,0.237500,0.100000,0.100000},{0.737500,0.237500,0.050000,0.050000},{0.737500,0.237500,0.100000,0.100000},{0.762500,0.237500,0.050000,0.050000},{0.762500,0.237500,0.100000,0.100000},{0.787500,0.237500,0.050000,0.050000},{0.787500,0.237500,0.100000,0.100000},{0.812500,0.237500,0.050000,0.050000},{0.812500,0.237500,0.100000,0.100000},{0.837500,0.237500,0.050000,0.050000},{0.837500,0.237500,0.100000,0.100000},{0.862500,0.237500,0.050000,0.050000},{0.862500,0.237500,0.100000,0.100000},{0.887500,0.237500,0.050000,0.050000},{0.887500,0.237500,0.100000,0.100000},{0.912500,0.237500,0.050000,0.050000},{0.912500,0.237500,0.100000,0.100000},{0.937500,0.237500,0.050000,0.050000},{0.937500,0.237500,0.100000,0.100000},{0.962500,0.237500,0.050000,0.050000},{0.962500,0.237500,0.100000,0.100000},{0.987500,0.237500,0.050000,0.050000},{0.987500,0.237500,0.100000,0.100000},{0.012500,0.262500,0.050000,0.050000},{0.012500,0.262500,0.100000,0.100000},{0.037500,0.262500,0.050000,0.050000},{0.037500,0.262500,0.100000,0.100000},{0.062500,0.262500,0.050000,0.050000},{0.062500,0.262500,0.100000,0.100000},{0.087500,0.262500,0.050000,0.050000},{0.087500,0.262500,0.100000,0.100000},{0.112500,0.262500,0.050000,0.050000},{0.112500,0.262500,0.100000,0.100000},{0.137500,0.262500,0.050000,0.050000},{0.137500,0.262500,0.100000,0.100000},{0.162500,0.262500,0.050000,0.050000},{0.162500,0.262500,0.100000,0.100000},{0.187500,0.262500,0.050000,0.050000},{0.187500,0.262500,0.100000,0.100000},{0.212500,0.262500,0.050000,0.050000},{0.212500,0.262500,0.100000,0.100000},{0.237500,0.262500,0.050000,0.050000},{0.237500,0.262500,0.100000,0.100000},{0.262500,0.262500,0.050000,0.050000},{0.262500,0.262500,0.100000,0.100000},{0.287500,0.262500,0.050000,0.050000},{0.287500,0.262500,0.100000,0.100000},{0.312500,0.262500,0.050000,0.050000},{0.312500,0.262500,0.100000,0.100000},{0.337500,0.262500,0.050000,0.050000},{0.337500,0.262500,0.100000,0.100000},{0.362500,0.262500,0.050000,0.050000},{0.362500,0.262500,0.100000,0.100000},{0.387500,0.262500,0.050000,0.050000},{0.387500,0.262500,0.100000,0.100000},{0.412500,0.262500,0.050000,0.050000},{0.412500,0.262500,0.100000,0.100000},{0.437500,0.262500,0.050000,0.050000},{0.437500,0.262500,0.100000,0.100000},{0.462500,0.262500,0.050000,0.050000},{0.462500,0.262500,0.100000,0.100000},{0.487500,0.262500,0.050000,0.050000},{0.487500,0.262500,0.100000,0.100000},{0.512500,0.262500,0.050000,0.050000},{0.512500,0.262500,0.100000,0.100000},{0.537500,0.262500,0.050000,0.050000},{0.537500,0.262500,0.100000,0.100000},{0.562500,0.262500,0.050000,0.050000},{0.562500,0.262500,0.100000,0.100000},{0.587500,0.262500,0.050000,0.050000},{0.587500,0.262500,0.100000,0.100000},{0.612500,0.262500,0.050000,0.050000},{0.612500,0.262500,0.100000,0.100000},{0.637500,0.262500,0.050000,0.050000},{0.637500,0.262500,0.100000,0.100000},{0.662500,0.262500,0.050000,0.050000},{0.662500,0.262500,0.100000,0.100000},{0.687500,0.262500,0.050000,0.050000},{0.687500,0.262500,0.100000,0.100000},{0.712500,0.262500,0.050000,0.050000},{0.712500,0.262500,0.100000,0.100000},{0.737500,0.262500,0.050000,0.050000},{0.737500,0.262500,0.100000,0.100000},{0.762500,0.262500,0.050000,0.050000},{0.762500,0.262500,0.100000,0.100000},{0.787500,0.262500,0.050000,0.050000},{0.787500,0.262500,0.100000,0.100000},{0.812500,0.262500,0.050000,0.050000},{0.812500,0.262500,0.100000,0.100000},{0.837500,0.262500,0.050000,0.050000},{0.837500,0.262500,0.100000,0.100000},{0.862500,0.262500,0.050000,0.050000},{0.862500,0.262500,0.100000,0.100000},{0.887500,0.262500,0.050000,0.050000},{0.887500,0.262500,0.100000,0.100000},{0.912500,0.262500,0.050000,0.050000},{0.912500,0.262500,0.100000,0.100000},{0.937500,0.262500,0.050000,0.050000},{0.937500,0.262500,0.100000,0.100000},{0.962500,0.262500,0.050000,0.050000},{0.962500,0.262500,0.100000,0.100000},{0.987500,0.262500,0.050000,0.050000},{0.987500,0.262500,0.100000,0.100000},{0.012500,0.287500,0.050000,0.050000},{0.012500,0.287500,0.100000,0.100000},{0.037500,0.287500,0.050000,0.050000},{0.037500,0.287500,0.100000,0.100000},{0.062500,0.287500,0.050000,0.050000},{0.062500,0.287500,0.100000,0.100000},{0.087500,0.287500,0.050000,0.050000},{0.087500,0.287500,0.100000,0.100000},{0.112500,0.287500,0.050000,0.050000},{0.112500,0.287500,0.100000,0.100000},{0.137500,0.287500,0.050000,0.050000},{0.137500,0.287500,0.100000,0.100000},{0.162500,0.287500,0.050000,0.050000},{0.162500,0.287500,0.100000,0.100000},{0.187500,0.287500,0.050000,0.050000},{0.187500,0.287500,0.100000,0.100000},{0.212500,0.287500,0.050000,0.050000},{0.212500,0.287500,0.100000,0.100000},{0.237500,0.287500,0.050000,0.050000},{0.237500,0.287500,0.100000,0.100000},{0.262500,0.287500,0.050000,0.050000},{0.262500,0.287500,0.100000,0.100000},{0.287500,0.287500,0.050000,0.050000},{0.287500,0.287500,0.100000,0.100000},{0.312500,0.287500,0.050000,0.050000},{0.312500,0.287500,0.100000,0.100000},{0.337500,0.287500,0.050000,0.050000},{0.337500,0.287500,0.100000,0.100000},{0.362500,0.287500,0.050000,0.050000},{0.362500,0.287500,0.100000,0.100000},{0.387500,0.287500,0.050000,0.050000},{0.387500,0.287500,0.100000,0.100000},{0.412500,0.287500,0.050000,0.050000},{0.412500,0.287500,0.100000,0.100000},{0.437500,0.287500,0.050000,0.050000},{0.437500,0.287500,0.100000,0.100000},{0.462500,0.287500,0.050000,0.050000},{0.462500,0.287500,0.100000,0.100000},{0.487500,0.287500,0.050000,0.050000},{0.487500,0.287500,0.100000,0.100000},{0.512500,0.287500,0.050000,0.050000},{0.512500,0.287500,0.100000,0.100000},{0.537500,0.287500,0.050000,0.050000},{0.537500,0.287500,0.100000,0.100000},{0.562500,0.287500,0.050000,0.050000},{0.562500,0.287500,0.100000,0.100000},{0.587500,0.287500,0.050000,0.050000},{0.587500,0.287500,0.100000,0.100000},{0.612500,0.287500,0.050000,0.050000},{0.612500,0.287500,0.100000,0.100000},{0.637500,0.287500,0.050000,0.050000},{0.637500,0.287500,0.100000,0.100000},{0.662500,0.287500,0.050000,0.050000},{0.662500,0.287500,0.100000,0.100000},{0.687500,0.287500,0.050000,0.050000},{0.687500,0.287500,0.100000,0.100000},{0.712500,0.287500,0.050000,0.050000},{0.712500,0.287500,0.100000,0.100000},{0.737500,0.287500,0.050000,0.050000},{0.737500,0.287500,0.100000,0.100000},{0.762500,0.287500,0.050000,0.050000},{0.762500,0.287500,0.100000,0.100000},{0.787500,0.287500,0.050000,0.050000},{0.787500,0.287500,0.100000,0.100000},{0.812500,0.287500,0.050000,0.050000},{0.812500,0.287500,0.100000,0.100000},{0.837500,0.287500,0.050000,0.050000},{0.837500,0.287500,0.100000,0.100000},{0.862500,0.287500,0.050000,0.050000},{0.862500,0.287500,0.100000,0.100000},{0.887500,0.287500,0.050000,0.050000},{0.887500,0.287500,0.100000,0.100000},{0.912500,0.287500,0.050000,0.050000},{0.912500,0.287500,0.100000,0.100000},{0.937500,0.287500,0.050000,0.050000},{0.937500,0.287500,0.100000,0.100000},{0.962500,0.287500,0.050000,0.050000},{0.962500,0.287500,0.100000,0.100000},{0.987500,0.287500,0.050000,0.050000},{0.987500,0.287500,0.100000,0.100000},{0.012500,0.312500,0.050000,0.050000},{0.012500,0.312500,0.100000,0.100000},{0.037500,0.312500,0.050000,0.050000},{0.037500,0.312500,0.100000,0.100000},{0.062500,0.312500,0.050000,0.050000},{0.062500,0.312500,0.100000,0.100000},{0.087500,0.312500,0.050000,0.050000},{0.087500,0.312500,0.100000,0.100000},{0.112500,0.312500,0.050000,0.050000},{0.112500,0.312500,0.100000,0.100000},{0.137500,0.312500,0.050000,0.050000},{0.137500,0.312500,0.100000,0.100000},{0.162500,0.312500,0.050000,0.050000},{0.162500,0.312500,0.100000,0.100000},{0.187500,0.312500,0.050000,0.050000},{0.187500,0.312500,0.100000,0.100000},{0.212500,0.312500,0.050000,0.050000},{0.212500,0.312500,0.100000,0.100000},{0.237500,0.312500,0.050000,0.050000},{0.237500,0.312500,0.100000,0.100000},{0.262500,0.312500,0.050000,0.050000},{0.262500,0.312500,0.100000,0.100000},{0.287500,0.312500,0.050000,0.050000},{0.287500,0.312500,0.100000,0.100000},{0.312500,0.312500,0.050000,0.050000},{0.312500,0.312500,0.100000,0.100000},{0.337500,0.312500,0.050000,0.050000},{0.337500,0.312500,0.100000,0.100000},{0.362500,0.312500,0.050000,0.050000},{0.362500,0.312500,0.100000,0.100000},{0.387500,0.312500,0.050000,0.050000},{0.387500,0.312500,0.100000,0.100000},{0.412500,0.312500,0.050000,0.050000},{0.412500,0.312500,0.100000,0.100000},{0.437500,0.312500,0.050000,0.050000},{0.437500,0.312500,0.100000,0.100000},{0.462500,0.312500,0.050000,0.050000},{0.462500,0.312500,0.100000,0.100000},{0.487500,0.312500,0.050000,0.050000},{0.487500,0.312500,0.100000,0.100000},{0.512500,0.312500,0.050000,0.050000},{0.512500,0.312500,0.100000,0.100000},{0.537500,0.312500,0.050000,0.050000},{0.537500,0.312500,0.100000,0.100000},{0.562500,0.312500,0.050000,0.050000},{0.562500,0.312500,0.100000,0.100000},{0.587500,0.312500,0.050000,0.050000},{0.587500,0.312500,0.100000,0.100000},{0.612500,0.312500,0.050000,0.050000},{0.612500,0.312500,0.100000,0.100000},{0.637500,0.312500,0.050000,0.050000},{0.637500,0.312500,0.100000,0.100000},{0.662500,0.312500,0.050000,0.050000},{0.662500,0.312500,0.100000,0.100000},{0.687500,0.312500,0.050000,0.050000},{0.687500,0.312500,0.100000,0.100000},{0.712500,0.312500,0.050000,0.050000},{0.712500,0.312500,0.100000,0.100000},{0.737500,0.312500,0.050000,0.050000},{0.737500,0.312500,0.100000,0.100000},{0.762500,0.312500,0.050000,0.050000},{0.762500,0.312500,0.100000,0.100000},{0.787500,0.312500,0.050000,0.050000},{0.787500,0.312500,0.100000,0.100000},{0.812500,0.312500,0.050000,0.050000},{0.812500,0.312500,0.100000,0.100000},{0.837500,0.312500,0.050000,0.050000},{0.837500,0.312500,0.100000,0.100000},{0.862500,0.312500,0.050000,0.050000},{0.862500,0.312500,0.100000,0.100000},{0.887500,0.312500,0.050000,0.050000},{0.887500,0.312500,0.100000,0.100000},{0.912500,0.312500,0.050000,0.050000},{0.912500,0.312500,0.100000,0.100000},{0.937500,0.312500,0.050000,0.050000},{0.937500,0.312500,0.100000,0.100000},{0.962500,0.312500,0.050000,0.050000},{0.962500,0.312500,0.100000,0.100000},{0.987500,0.312500,0.050000,0.050000},{0.987500,0.312500,0.100000,0.100000},{0.012500,0.337500,0.050000,0.050000},{0.012500,0.337500,0.100000,0.100000},{0.037500,0.337500,0.050000,0.050000},{0.037500,0.337500,0.100000,0.100000},{0.062500,0.337500,0.050000,0.050000},{0.062500,0.337500,0.100000,0.100000},{0.087500,0.337500,0.050000,0.050000},{0.087500,0.337500,0.100000,0.100000},{0.112500,0.337500,0.050000,0.050000},{0.112500,0.337500,0.100000,0.100000},{0.137500,0.337500,0.050000,0.050000},{0.137500,0.337500,0.100000,0.100000},{0.162500,0.337500,0.050000,0.050000},{0.162500,0.337500,0.100000,0.100000},{0.187500,0.337500,0.050000,0.050000},{0.187500,0.337500,0.100000,0.100000},{0.212500,0.337500,0.050000,0.050000},{0.212500,0.337500,0.100000,0.100000},{0.237500,0.337500,0.050000,0.050000},{0.237500,0.337500,0.100000,0.100000},{0.262500,0.337500,0.050000,0.050000},{0.262500,0.337500,0.100000,0.100000},{0.287500,0.337500,0.050000,0.050000},{0.287500,0.337500,0.100000,0.100000},{0.312500,0.337500,0.050000,0.050000},{0.312500,0.337500,0.100000,0.100000},{0.337500,0.337500,0.050000,0.050000},{0.337500,0.337500,0.100000,0.100000},{0.362500,0.337500,0.050000,0.050000},{0.362500,0.337500,0.100000,0.100000},{0.387500,0.337500,0.050000,0.050000},{0.387500,0.337500,0.100000,0.100000},{0.412500,0.337500,0.050000,0.050000},{0.412500,0.337500,0.100000,0.100000},{0.437500,0.337500,0.050000,0.050000},{0.437500,0.337500,0.100000,0.100000},{0.462500,0.337500,0.050000,0.050000},{0.462500,0.337500,0.100000,0.100000},{0.487500,0.337500,0.050000,0.050000},{0.487500,0.337500,0.100000,0.100000},{0.512500,0.337500,0.050000,0.050000},{0.512500,0.337500,0.100000,0.100000},{0.537500,0.337500,0.050000,0.050000},{0.537500,0.337500,0.100000,0.100000},{0.562500,0.337500,0.050000,0.050000},{0.562500,0.337500,0.100000,0.100000},{0.587500,0.337500,0.050000,0.050000},{0.587500,0.337500,0.100000,0.100000},{0.612500,0.337500,0.050000,0.050000},{0.612500,0.337500,0.100000,0.100000},{0.637500,0.337500,0.050000,0.050000},{0.637500,0.337500,0.100000,0.100000},{0.662500,0.337500,0.050000,0.050000},{0.662500,0.337500,0.100000,0.100000},{0.687500,0.337500,0.050000,0.050000},{0.687500,0.337500,0.100000,0.100000},{0.712500,0.337500,0.050000,0.050000},{0.712500,0.337500,0.100000,0.100000},{0.737500,0.337500,0.050000,0.050000},{0.737500,0.337500,0.100000,0.100000},{0.762500,0.337500,0.050000,0.050000},{0.762500,0.337500,0.100000,0.100000},{0.787500,0.337500,0.050000,0.050000},{0.787500,0.337500,0.100000,0.100000},{0.812500,0.337500,0.050000,0.050000},{0.812500,0.337500,0.100000,0.100000},{0.837500,0.337500,0.050000,0.050000},{0.837500,0.337500,0.100000,0.100000},{0.862500,0.337500,0.050000,0.050000},{0.862500,0.337500,0.100000,0.100000},{0.887500,0.337500,0.050000,0.050000},{0.887500,0.337500,0.100000,0.100000},{0.912500,0.337500,0.050000,0.050000},{0.912500,0.337500,0.100000,0.100000},{0.937500,0.337500,0.050000,0.050000},{0.937500,0.337500,0.100000,0.100000},{0.962500,0.337500,0.050000,0.050000},{0.962500,0.337500,0.100000,0.100000},{0.987500,0.337500,0.050000,0.050000},{0.987500,0.337500,0.100000,0.100000},{0.012500,0.362500,0.050000,0.050000},{0.012500,0.362500,0.100000,0.100000},{0.037500,0.362500,0.050000,0.050000},{0.037500,0.362500,0.100000,0.100000},{0.062500,0.362500,0.050000,0.050000},{0.062500,0.362500,0.100000,0.100000},{0.087500,0.362500,0.050000,0.050000},{0.087500,0.362500,0.100000,0.100000},{0.112500,0.362500,0.050000,0.050000},{0.112500,0.362500,0.100000,0.100000},{0.137500,0.362500,0.050000,0.050000},{0.137500,0.362500,0.100000,0.100000},{0.162500,0.362500,0.050000,0.050000},{0.162500,0.362500,0.100000,0.100000},{0.187500,0.362500,0.050000,0.050000},{0.187500,0.362500,0.100000,0.100000},{0.212500,0.362500,0.050000,0.050000},{0.212500,0.362500,0.100000,0.100000},{0.237500,0.362500,0.050000,0.050000},{0.237500,0.362500,0.100000,0.100000},{0.262500,0.362500,0.050000,0.050000},{0.262500,0.362500,0.100000,0.100000},{0.287500,0.362500,0.050000,0.050000},{0.287500,0.362500,0.100000,0.100000},{0.312500,0.362500,0.050000,0.050000},{0.312500,0.362500,0.100000,0.100000},{0.337500,0.362500,0.050000,0.050000},{0.337500,0.362500,0.100000,0.100000},{0.362500,0.362500,0.050000,0.050000},{0.362500,0.362500,0.100000,0.100000},{0.387500,0.362500,0.050000,0.050000},{0.387500,0.362500,0.100000,0.100000},{0.412500,0.362500,0.050000,0.050000},{0.412500,0.362500,0.100000,0.100000},{0.437500,0.362500,0.050000,0.050000},{0.437500,0.362500,0.100000,0.100000},{0.462500,0.362500,0.050000,0.050000},{0.462500,0.362500,0.100000,0.100000},{0.487500,0.362500,0.050000,0.050000},{0.487500,0.362500,0.100000,0.100000},{0.512500,0.362500,0.050000,0.050000},{0.512500,0.362500,0.100000,0.100000},{0.537500,0.362500,0.050000,0.050000},{0.537500,0.362500,0.100000,0.100000},{0.562500,0.362500,0.050000,0.050000},{0.562500,0.362500,0.100000,0.100000},{0.587500,0.362500,0.050000,0.050000},{0.587500,0.362500,0.100000,0.100000},{0.612500,0.362500,0.050000,0.050000},{0.612500,0.362500,0.100000,0.100000},{0.637500,0.362500,0.050000,0.050000},{0.637500,0.362500,0.100000,0.100000},{0.662500,0.362500,0.050000,0.050000},{0.662500,0.362500,0.100000,0.100000},{0.687500,0.362500,0.050000,0.050000},{0.687500,0.362500,0.100000,0.100000},{0.712500,0.362500,0.050000,0.050000},{0.712500,0.362500,0.100000,0.100000},{0.737500,0.362500,0.050000,0.050000},{0.737500,0.362500,0.100000,0.100000},{0.762500,0.362500,0.050000,0.050000},{0.762500,0.362500,0.100000,0.100000},{0.787500,0.362500,0.050000,0.050000},{0.787500,0.362500,0.100000,0.100000},{0.812500,0.362500,0.050000,0.050000},{0.812500,0.362500,0.100000,0.100000},{0.837500,0.362500,0.050000,0.050000},{0.837500,0.362500,0.100000,0.100000},{0.862500,0.362500,0.050000,0.050000},{0.862500,0.362500,0.100000,0.100000},{0.887500,0.362500,0.050000,0.050000},{0.887500,0.362500,0.100000,0.100000},{0.912500,0.362500,0.050000,0.050000},{0.912500,0.362500,0.100000,0.100000},{0.937500,0.362500,0.050000,0.050000},{0.937500,0.362500,0.100000,0.100000},{0.962500,0.362500,0.050000,0.050000},{0.962500,0.362500,0.100000,0.100000},{0.987500,0.362500,0.050000,0.050000},{0.987500,0.362500,0.100000,0.100000},{0.012500,0.387500,0.050000,0.050000},{0.012500,0.387500,0.100000,0.100000},{0.037500,0.387500,0.050000,0.050000},{0.037500,0.387500,0.100000,0.100000},{0.062500,0.387500,0.050000,0.050000},{0.062500,0.387500,0.100000,0.100000},{0.087500,0.387500,0.050000,0.050000},{0.087500,0.387500,0.100000,0.100000},{0.112500,0.387500,0.050000,0.050000},{0.112500,0.387500,0.100000,0.100000},{0.137500,0.387500,0.050000,0.050000},{0.137500,0.387500,0.100000,0.100000},{0.162500,0.387500,0.050000,0.050000},{0.162500,0.387500,0.100000,0.100000},{0.187500,0.387500,0.050000,0.050000},{0.187500,0.387500,0.100000,0.100000},{0.212500,0.387500,0.050000,0.050000},{0.212500,0.387500,0.100000,0.100000},{0.237500,0.387500,0.050000,0.050000},{0.237500,0.387500,0.100000,0.100000},{0.262500,0.387500,0.050000,0.050000},{0.262500,0.387500,0.100000,0.100000},{0.287500,0.387500,0.050000,0.050000},{0.287500,0.387500,0.100000,0.100000},{0.312500,0.387500,0.050000,0.050000},{0.312500,0.387500,0.100000,0.100000},{0.337500,0.387500,0.050000,0.050000},{0.337500,0.387500,0.100000,0.100000},{0.362500,0.387500,0.050000,0.050000},{0.362500,0.387500,0.100000,0.100000},{0.387500,0.387500,0.050000,0.050000},{0.387500,0.387500,0.100000,0.100000},{0.412500,0.387500,0.050000,0.050000},{0.412500,0.387500,0.100000,0.100000},{0.437500,0.387500,0.050000,0.050000},{0.437500,0.387500,0.100000,0.100000},{0.462500,0.387500,0.050000,0.050000},{0.462500,0.387500,0.100000,0.100000},{0.487500,0.387500,0.050000,0.050000},{0.487500,0.387500,0.100000,0.100000},{0.512500,0.387500,0.050000,0.050000},{0.512500,0.387500,0.100000,0.100000},{0.537500,0.387500,0.050000,0.050000},{0.537500,0.387500,0.100000,0.100000},{0.562500,0.387500,0.050000,0.050000},{0.562500,0.387500,0.100000,0.100000},{0.587500,0.387500,0.050000,0.050000},{0.587500,0.387500,0.100000,0.100000},{0.612500,0.387500,0.050000,0.050000},{0.612500,0.387500,0.100000,0.100000},{0.637500,0.387500,0.050000,0.050000},{0.637500,0.387500,0.100000,0.100000},{0.662500,0.387500,0.050000,0.050000},{0.662500,0.387500,0.100000,0.100000},{0.687500,0.387500,0.050000,0.050000},{0.687500,0.387500,0.100000,0.100000},{0.712500,0.387500,0.050000,0.050000},{0.712500,0.387500,0.100000,0.100000},{0.737500,0.387500,0.050000,0.050000},{0.737500,0.387500,0.100000,0.100000},{0.762500,0.387500,0.050000,0.050000},{0.762500,0.387500,0.100000,0.100000},{0.787500,0.387500,0.050000,0.050000},{0.787500,0.387500,0.100000,0.100000},{0.812500,0.387500,0.050000,0.050000},{0.812500,0.387500,0.100000,0.100000},{0.837500,0.387500,0.050000,0.050000},{0.837500,0.387500,0.100000,0.100000},{0.862500,0.387500,0.050000,0.050000},{0.862500,0.387500,0.100000,0.100000},{0.887500,0.387500,0.050000,0.050000},{0.887500,0.387500,0.100000,0.100000},{0.912500,0.387500,0.050000,0.050000},{0.912500,0.387500,0.100000,0.100000},{0.937500,0.387500,0.050000,0.050000},{0.937500,0.387500,0.100000,0.100000},{0.962500,0.387500,0.050000,0.050000},{0.962500,0.387500,0.100000,0.100000},{0.987500,0.387500,0.050000,0.050000},{0.987500,0.387500,0.100000,0.100000},{0.012500,0.412500,0.050000,0.050000},{0.012500,0.412500,0.100000,0.100000},{0.037500,0.412500,0.050000,0.050000},{0.037500,0.412500,0.100000,0.100000},{0.062500,0.412500,0.050000,0.050000},{0.062500,0.412500,0.100000,0.100000},{0.087500,0.412500,0.050000,0.050000},{0.087500,0.412500,0.100000,0.100000},{0.112500,0.412500,0.050000,0.050000},{0.112500,0.412500,0.100000,0.100000},{0.137500,0.412500,0.050000,0.050000},{0.137500,0.412500,0.100000,0.100000},{0.162500,0.412500,0.050000,0.050000},{0.162500,0.412500,0.100000,0.100000},{0.187500,0.412500,0.050000,0.050000},{0.187500,0.412500,0.100000,0.100000},{0.212500,0.412500,0.050000,0.050000},{0.212500,0.412500,0.100000,0.100000},{0.237500,0.412500,0.050000,0.050000},{0.237500,0.412500,0.100000,0.100000},{0.262500,0.412500,0.050000,0.050000},{0.262500,0.412500,0.100000,0.100000},{0.287500,0.412500,0.050000,0.050000},{0.287500,0.412500,0.100000,0.100000},{0.312500,0.412500,0.050000,0.050000},{0.312500,0.412500,0.100000,0.100000},{0.337500,0.412500,0.050000,0.050000},{0.337500,0.412500,0.100000,0.100000},{0.362500,0.412500,0.050000,0.050000},{0.362500,0.412500,0.100000,0.100000},{0.387500,0.412500,0.050000,0.050000},{0.387500,0.412500,0.100000,0.100000},{0.412500,0.412500,0.050000,0.050000},{0.412500,0.412500,0.100000,0.100000},{0.437500,0.412500,0.050000,0.050000},{0.437500,0.412500,0.100000,0.100000},{0.462500,0.412500,0.050000,0.050000},{0.462500,0.412500,0.100000,0.100000},{0.487500,0.412500,0.050000,0.050000},{0.487500,0.412500,0.100000,0.100000},{0.512500,0.412500,0.050000,0.050000},{0.512500,0.412500,0.100000,0.100000},{0.537500,0.412500,0.050000,0.050000},{0.537500,0.412500,0.100000,0.100000},{0.562500,0.412500,0.050000,0.050000},{0.562500,0.412500,0.100000,0.100000},{0.587500,0.412500,0.050000,0.050000},{0.587500,0.412500,0.100000,0.100000},{0.612500,0.412500,0.050000,0.050000},{0.612500,0.412500,0.100000,0.100000},{0.637500,0.412500,0.050000,0.050000},{0.637500,0.412500,0.100000,0.100000},{0.662500,0.412500,0.050000,0.050000},{0.662500,0.412500,0.100000,0.100000},{0.687500,0.412500,0.050000,0.050000},{0.687500,0.412500,0.100000,0.100000},{0.712500,0.412500,0.050000,0.050000},{0.712500,0.412500,0.100000,0.100000},{0.737500,0.412500,0.050000,0.050000},{0.737500,0.412500,0.100000,0.100000},{0.762500,0.412500,0.050000,0.050000},{0.762500,0.412500,0.100000,0.100000},{0.787500,0.412500,0.050000,0.050000},{0.787500,0.412500,0.100000,0.100000},{0.812500,0.412500,0.050000,0.050000},{0.812500,0.412500,0.100000,0.100000},{0.837500,0.412500,0.050000,0.050000},{0.837500,0.412500,0.100000,0.100000},{0.862500,0.412500,0.050000,0.050000},{0.862500,0.412500,0.100000,0.100000},{0.887500,0.412500,0.050000,0.050000},{0.887500,0.412500,0.100000,0.100000},{0.912500,0.412500,0.050000,0.050000},{0.912500,0.412500,0.100000,0.100000},{0.937500,0.412500,0.050000,0.050000},{0.937500,0.412500,0.100000,0.100000},{0.962500,0.412500,0.050000,0.050000},{0.962500,0.412500,0.100000,0.100000},{0.987500,0.412500,0.050000,0.050000},{0.987500,0.412500,0.100000,0.100000},{0.012500,0.437500,0.050000,0.050000},{0.012500,0.437500,0.100000,0.100000},{0.037500,0.437500,0.050000,0.050000},{0.037500,0.437500,0.100000,0.100000},{0.062500,0.437500,0.050000,0.050000},{0.062500,0.437500,0.100000,0.100000},{0.087500,0.437500,0.050000,0.050000},{0.087500,0.437500,0.100000,0.100000},{0.112500,0.437500,0.050000,0.050000},{0.112500,0.437500,0.100000,0.100000},{0.137500,0.437500,0.050000,0.050000},{0.137500,0.437500,0.100000,0.100000},{0.162500,0.437500,0.050000,0.050000},{0.162500,0.437500,0.100000,0.100000},{0.187500,0.437500,0.050000,0.050000},{0.187500,0.437500,0.100000,0.100000},{0.212500,0.437500,0.050000,0.050000},{0.212500,0.437500,0.100000,0.100000},{0.237500,0.437500,0.050000,0.050000},{0.237500,0.437500,0.100000,0.100000},{0.262500,0.437500,0.050000,0.050000},{0.262500,0.437500,0.100000,0.100000},{0.287500,0.437500,0.050000,0.050000},{0.287500,0.437500,0.100000,0.100000},{0.312500,0.437500,0.050000,0.050000},{0.312500,0.437500,0.100000,0.100000},{0.337500,0.437500,0.050000,0.050000},{0.337500,0.437500,0.100000,0.100000},{0.362500,0.437500,0.050000,0.050000},{0.362500,0.437500,0.100000,0.100000},{0.387500,0.437500,0.050000,0.050000},{0.387500,0.437500,0.100000,0.100000},{0.412500,0.437500,0.050000,0.050000},{0.412500,0.437500,0.100000,0.100000},{0.437500,0.437500,0.050000,0.050000},{0.437500,0.437500,0.100000,0.100000},{0.462500,0.437500,0.050000,0.050000},{0.462500,0.437500,0.100000,0.100000},{0.487500,0.437500,0.050000,0.050000},{0.487500,0.437500,0.100000,0.100000},{0.512500,0.437500,0.050000,0.050000},{0.512500,0.437500,0.100000,0.100000},{0.537500,0.437500,0.050000,0.050000},{0.537500,0.437500,0.100000,0.100000},{0.562500,0.437500,0.050000,0.050000},{0.562500,0.437500,0.100000,0.100000},{0.587500,0.437500,0.050000,0.050000},{0.587500,0.437500,0.100000,0.100000},{0.612500,0.437500,0.050000,0.050000},{0.612500,0.437500,0.100000,0.100000},{0.637500,0.437500,0.050000,0.050000},{0.637500,0.437500,0.100000,0.100000},{0.662500,0.437500,0.050000,0.050000},{0.662500,0.437500,0.100000,0.100000},{0.687500,0.437500,0.050000,0.050000},{0.687500,0.437500,0.100000,0.100000},{0.712500,0.437500,0.050000,0.050000},{0.712500,0.437500,0.100000,0.100000},{0.737500,0.437500,0.050000,0.050000},{0.737500,0.437500,0.100000,0.100000},{0.762500,0.437500,0.050000,0.050000},{0.762500,0.437500,0.100000,0.100000},{0.787500,0.437500,0.050000,0.050000},{0.787500,0.437500,0.100000,0.100000},{0.812500,0.437500,0.050000,0.050000},{0.812500,0.437500,0.100000,0.100000},{0.837500,0.437500,0.050000,0.050000},{0.837500,0.437500,0.100000,0.100000},{0.862500,0.437500,0.050000,0.050000},{0.862500,0.437500,0.100000,0.100000},{0.887500,0.437500,0.050000,0.050000},{0.887500,0.437500,0.100000,0.100000},{0.912500,0.437500,0.050000,0.050000},{0.912500,0.437500,0.100000,0.100000},{0.937500,0.437500,0.050000,0.050000},{0.937500,0.437500,0.100000,0.100000},{0.962500,0.437500,0.050000,0.050000},{0.962500,0.437500,0.100000,0.100000},{0.987500,0.437500,0.050000,0.050000},{0.987500,0.437500,0.100000,0.100000},{0.012500,0.462500,0.050000,0.050000},{0.012500,0.462500,0.100000,0.100000},{0.037500,0.462500,0.050000,0.050000},{0.037500,0.462500,0.100000,0.100000},{0.062500,0.462500,0.050000,0.050000},{0.062500,0.462500,0.100000,0.100000},{0.087500,0.462500,0.050000,0.050000},{0.087500,0.462500,0.100000,0.100000},{0.112500,0.462500,0.050000,0.050000},{0.112500,0.462500,0.100000,0.100000},{0.137500,0.462500,0.050000,0.050000},{0.137500,0.462500,0.100000,0.100000},{0.162500,0.462500,0.050000,0.050000},{0.162500,0.462500,0.100000,0.100000},{0.187500,0.462500,0.050000,0.050000},{0.187500,0.462500,0.100000,0.100000},{0.212500,0.462500,0.050000,0.050000},{0.212500,0.462500,0.100000,0.100000},{0.237500,0.462500,0.050000,0.050000},{0.237500,0.462500,0.100000,0.100000},{0.262500,0.462500,0.050000,0.050000},{0.262500,0.462500,0.100000,0.100000},{0.287500,0.462500,0.050000,0.050000},{0.287500,0.462500,0.100000,0.100000},{0.312500,0.462500,0.050000,0.050000},{0.312500,0.462500,0.100000,0.100000},{0.337500,0.462500,0.050000,0.050000},{0.337500,0.462500,0.100000,0.100000},{0.362500,0.462500,0.050000,0.050000},{0.362500,0.462500,0.100000,0.100000},{0.387500,0.462500,0.050000,0.050000},{0.387500,0.462500,0.100000,0.100000},{0.412500,0.462500,0.050000,0.050000},{0.412500,0.462500,0.100000,0.100000},{0.437500,0.462500,0.050000,0.050000},{0.437500,0.462500,0.100000,0.100000},{0.462500,0.462500,0.050000,0.050000},{0.462500,0.462500,0.100000,0.100000},{0.487500,0.462500,0.050000,0.050000},{0.487500,0.462500,0.100000,0.100000},{0.512500,0.462500,0.050000,0.050000},{0.512500,0.462500,0.100000,0.100000},{0.537500,0.462500,0.050000,0.050000},{0.537500,0.462500,0.100000,0.100000},{0.562500,0.462500,0.050000,0.050000},{0.562500,0.462500,0.100000,0.100000},{0.587500,0.462500,0.050000,0.050000},{0.587500,0.462500,0.100000,0.100000},{0.612500,0.462500,0.050000,0.050000},{0.612500,0.462500,0.100000,0.100000},{0.637500,0.462500,0.050000,0.050000},{0.637500,0.462500,0.100000,0.100000},{0.662500,0.462500,0.050000,0.050000},{0.662500,0.462500,0.100000,0.100000},{0.687500,0.462500,0.050000,0.050000},{0.687500,0.462500,0.100000,0.100000},{0.712500,0.462500,0.050000,0.050000},{0.712500,0.462500,0.100000,0.100000},{0.737500,0.462500,0.050000,0.050000},{0.737500,0.462500,0.100000,0.100000},{0.762500,0.462500,0.050000,0.050000},{0.762500,0.462500,0.100000,0.100000},{0.787500,0.462500,0.050000,0.050000},{0.787500,0.462500,0.100000,0.100000},{0.812500,0.462500,0.050000,0.050000},{0.812500,0.462500,0.100000,0.100000},{0.837500,0.462500,0.050000,0.050000},{0.837500,0.462500,0.100000,0.100000},{0.862500,0.462500,0.050000,0.050000},{0.862500,0.462500,0.100000,0.100000},{0.887500,0.462500,0.050000,0.050000},{0.887500,0.462500,0.100000,0.100000},{0.912500,0.462500,0.050000,0.050000},{0.912500,0.462500,0.100000,0.100000},{0.937500,0.462500,0.050000,0.050000},{0.937500,0.462500,0.100000,0.100000},{0.962500,0.462500,0.050000,0.050000},{0.962500,0.462500,0.100000,0.100000},{0.987500,0.462500,0.050000,0.050000},{0.987500,0.462500,0.100000,0.100000},{0.012500,0.487500,0.050000,0.050000},{0.012500,0.487500,0.100000,0.100000},{0.037500,0.487500,0.050000,0.050000},{0.037500,0.487500,0.100000,0.100000},{0.062500,0.487500,0.050000,0.050000},{0.062500,0.487500,0.100000,0.100000},{0.087500,0.487500,0.050000,0.050000},{0.087500,0.487500,0.100000,0.100000},{0.112500,0.487500,0.050000,0.050000},{0.112500,0.487500,0.100000,0.100000},{0.137500,0.487500,0.050000,0.050000},{0.137500,0.487500,0.100000,0.100000},{0.162500,0.487500,0.050000,0.050000},{0.162500,0.487500,0.100000,0.100000},{0.187500,0.487500,0.050000,0.050000},{0.187500,0.487500,0.100000,0.100000},{0.212500,0.487500,0.050000,0.050000},{0.212500,0.487500,0.100000,0.100000},{0.237500,0.487500,0.050000,0.050000},{0.237500,0.487500,0.100000,0.100000},{0.262500,0.487500,0.050000,0.050000},{0.262500,0.487500,0.100000,0.100000},{0.287500,0.487500,0.050000,0.050000},{0.287500,0.487500,0.100000,0.100000},{0.312500,0.487500,0.050000,0.050000},{0.312500,0.487500,0.100000,0.100000},{0.337500,0.487500,0.050000,0.050000},{0.337500,0.487500,0.100000,0.100000},{0.362500,0.487500,0.050000,0.050000},{0.362500,0.487500,0.100000,0.100000},{0.387500,0.487500,0.050000,0.050000},{0.387500,0.487500,0.100000,0.100000},{0.412500,0.487500,0.050000,0.050000},{0.412500,0.487500,0.100000,0.100000},{0.437500,0.487500,0.050000,0.050000},{0.437500,0.487500,0.100000,0.100000},{0.462500,0.487500,0.050000,0.050000},{0.462500,0.487500,0.100000,0.100000},{0.487500,0.487500,0.050000,0.050000},{0.487500,0.487500,0.100000,0.100000},{0.512500,0.487500,0.050000,0.050000},{0.512500,0.487500,0.100000,0.100000},{0.537500,0.487500,0.050000,0.050000},{0.537500,0.487500,0.100000,0.100000},{0.562500,0.487500,0.050000,0.050000},{0.562500,0.487500,0.100000,0.100000},{0.587500,0.487500,0.050000,0.050000},{0.587500,0.487500,0.100000,0.100000},{0.612500,0.487500,0.050000,0.050000},{0.612500,0.487500,0.100000,0.100000},{0.637500,0.487500,0.050000,0.050000},{0.637500,0.487500,0.100000,0.100000},{0.662500,0.487500,0.050000,0.050000},{0.662500,0.487500,0.100000,0.100000},{0.687500,0.487500,0.050000,0.050000},{0.687500,0.487500,0.100000,0.100000},{0.712500,0.487500,0.050000,0.050000},{0.712500,0.487500,0.100000,0.100000},{0.737500,0.487500,0.050000,0.050000},{0.737500,0.487500,0.100000,0.100000},{0.762500,0.487500,0.050000,0.050000},{0.762500,0.487500,0.100000,0.100000},{0.787500,0.487500,0.050000,0.050000},{0.787500,0.487500,0.100000,0.100000},{0.812500,0.487500,0.050000,0.050000},{0.812500,0.487500,0.100000,0.100000},{0.837500,0.487500,0.050000,0.050000},{0.837500,0.487500,0.100000,0.100000},{0.862500,0.487500,0.050000,0.050000},{0.862500,0.487500,0.100000,0.100000},{0.887500,0.487500,0.050000,0.050000},{0.887500,0.487500,0.100000,0.100000},{0.912500,0.487500,0.050000,0.050000},{0.912500,0.487500,0.100000,0.100000},{0.937500,0.487500,0.050000,0.050000},{0.937500,0.487500,0.100000,0.100000},{0.962500,0.487500,0.050000,0.050000},{0.962500,0.487500,0.100000,0.100000},{0.987500,0.487500,0.050000,0.050000},{0.987500,0.487500,0.100000,0.100000},{0.012500,0.512500,0.050000,0.050000},{0.012500,0.512500,0.100000,0.100000},{0.037500,0.512500,0.050000,0.050000},{0.037500,0.512500,0.100000,0.100000},{0.062500,0.512500,0.050000,0.050000},{0.062500,0.512500,0.100000,0.100000},{0.087500,0.512500,0.050000,0.050000},{0.087500,0.512500,0.100000,0.100000},{0.112500,0.512500,0.050000,0.050000},{0.112500,0.512500,0.100000,0.100000},{0.137500,0.512500,0.050000,0.050000},{0.137500,0.512500,0.100000,0.100000},{0.162500,0.512500,0.050000,0.050000},{0.162500,0.512500,0.100000,0.100000},{0.187500,0.512500,0.050000,0.050000},{0.187500,0.512500,0.100000,0.100000},{0.212500,0.512500,0.050000,0.050000},{0.212500,0.512500,0.100000,0.100000},{0.237500,0.512500,0.050000,0.050000},{0.237500,0.512500,0.100000,0.100000},{0.262500,0.512500,0.050000,0.050000},{0.262500,0.512500,0.100000,0.100000},{0.287500,0.512500,0.050000,0.050000},{0.287500,0.512500,0.100000,0.100000},{0.312500,0.512500,0.050000,0.050000},{0.312500,0.512500,0.100000,0.100000},{0.337500,0.512500,0.050000,0.050000},{0.337500,0.512500,0.100000,0.100000},{0.362500,0.512500,0.050000,0.050000},{0.362500,0.512500,0.100000,0.100000},{0.387500,0.512500,0.050000,0.050000},{0.387500,0.512500,0.100000,0.100000},{0.412500,0.512500,0.050000,0.050000},{0.412500,0.512500,0.100000,0.100000},{0.437500,0.512500,0.050000,0.050000},{0.437500,0.512500,0.100000,0.100000},{0.462500,0.512500,0.050000,0.050000},{0.462500,0.512500,0.100000,0.100000},{0.487500,0.512500,0.050000,0.050000},{0.487500,0.512500,0.100000,0.100000},{0.512500,0.512500,0.050000,0.050000},{0.512500,0.512500,0.100000,0.100000},{0.537500,0.512500,0.050000,0.050000},{0.537500,0.512500,0.100000,0.100000},{0.562500,0.512500,0.050000,0.050000},{0.562500,0.512500,0.100000,0.100000},{0.587500,0.512500,0.050000,0.050000},{0.587500,0.512500,0.100000,0.100000},{0.612500,0.512500,0.050000,0.050000},{0.612500,0.512500,0.100000,0.100000},{0.637500,0.512500,0.050000,0.050000},{0.637500,0.512500,0.100000,0.100000},{0.662500,0.512500,0.050000,0.050000},{0.662500,0.512500,0.100000,0.100000},{0.687500,0.512500,0.050000,0.050000},{0.687500,0.512500,0.100000,0.100000},{0.712500,0.512500,0.050000,0.050000},{0.712500,0.512500,0.100000,0.100000},{0.737500,0.512500,0.050000,0.050000},{0.737500,0.512500,0.100000,0.100000},{0.762500,0.512500,0.050000,0.050000},{0.762500,0.512500,0.100000,0.100000},{0.787500,0.512500,0.050000,0.050000},{0.787500,0.512500,0.100000,0.100000},{0.812500,0.512500,0.050000,0.050000},{0.812500,0.512500,0.100000,0.100000},{0.837500,0.512500,0.050000,0.050000},{0.837500,0.512500,0.100000,0.100000},{0.862500,0.512500,0.050000,0.050000},{0.862500,0.512500,0.100000,0.100000},{0.887500,0.512500,0.050000,0.050000},{0.887500,0.512500,0.100000,0.100000},{0.912500,0.512500,0.050000,0.050000},{0.912500,0.512500,0.100000,0.100000},{0.937500,0.512500,0.050000,0.050000},{0.937500,0.512500,0.100000,0.100000},{0.962500,0.512500,0.050000,0.050000},{0.962500,0.512500,0.100000,0.100000},{0.987500,0.512500,0.050000,0.050000},{0.987500,0.512500,0.100000,0.100000},{0.012500,0.537500,0.050000,0.050000},{0.012500,0.537500,0.100000,0.100000},{0.037500,0.537500,0.050000,0.050000},{0.037500,0.537500,0.100000,0.100000},{0.062500,0.537500,0.050000,0.050000},{0.062500,0.537500,0.100000,0.100000},{0.087500,0.537500,0.050000,0.050000},{0.087500,0.537500,0.100000,0.100000},{0.112500,0.537500,0.050000,0.050000},{0.112500,0.537500,0.100000,0.100000},{0.137500,0.537500,0.050000,0.050000},{0.137500,0.537500,0.100000,0.100000},{0.162500,0.537500,0.050000,0.050000},{0.162500,0.537500,0.100000,0.100000},{0.187500,0.537500,0.050000,0.050000},{0.187500,0.537500,0.100000,0.100000},{0.212500,0.537500,0.050000,0.050000},{0.212500,0.537500,0.100000,0.100000},{0.237500,0.537500,0.050000,0.050000},{0.237500,0.537500,0.100000,0.100000},{0.262500,0.537500,0.050000,0.050000},{0.262500,0.537500,0.100000,0.100000},{0.287500,0.537500,0.050000,0.050000},{0.287500,0.537500,0.100000,0.100000},{0.312500,0.537500,0.050000,0.050000},{0.312500,0.537500,0.100000,0.100000},{0.337500,0.537500,0.050000,0.050000},{0.337500,0.537500,0.100000,0.100000},{0.362500,0.537500,0.050000,0.050000},{0.362500,0.537500,0.100000,0.100000},{0.387500,0.537500,0.050000,0.050000},{0.387500,0.537500,0.100000,0.100000},{0.412500,0.537500,0.050000,0.050000},{0.412500,0.537500,0.100000,0.100000},{0.437500,0.537500,0.050000,0.050000},{0.437500,0.537500,0.100000,0.100000},{0.462500,0.537500,0.050000,0.050000},{0.462500,0.537500,0.100000,0.100000},{0.487500,0.537500,0.050000,0.050000},{0.487500,0.537500,0.100000,0.100000},{0.512500,0.537500,0.050000,0.050000},{0.512500,0.537500,0.100000,0.100000},{0.537500,0.537500,0.050000,0.050000},{0.537500,0.537500,0.100000,0.100000},{0.562500,0.537500,0.050000,0.050000},{0.562500,0.537500,0.100000,0.100000},{0.587500,0.537500,0.050000,0.050000},{0.587500,0.537500,0.100000,0.100000},{0.612500,0.537500,0.050000,0.050000},{0.612500,0.537500,0.100000,0.100000},{0.637500,0.537500,0.050000,0.050000},{0.637500,0.537500,0.100000,0.100000},{0.662500,0.537500,0.050000,0.050000},{0.662500,0.537500,0.100000,0.100000},{0.687500,0.537500,0.050000,0.050000},{0.687500,0.537500,0.100000,0.100000},{0.712500,0.537500,0.050000,0.050000},{0.712500,0.537500,0.100000,0.100000},{0.737500,0.537500,0.050000,0.050000},{0.737500,0.537500,0.100000,0.100000},{0.762500,0.537500,0.050000,0.050000},{0.762500,0.537500,0.100000,0.100000},{0.787500,0.537500,0.050000,0.050000},{0.787500,0.537500,0.100000,0.100000},{0.812500,0.537500,0.050000,0.050000},{0.812500,0.537500,0.100000,0.100000},{0.837500,0.537500,0.050000,0.050000},{0.837500,0.537500,0.100000,0.100000},{0.862500,0.537500,0.050000,0.050000},{0.862500,0.537500,0.100000,0.100000},{0.887500,0.537500,0.050000,0.050000},{0.887500,0.537500,0.100000,0.100000},{0.912500,0.537500,0.050000,0.050000},{0.912500,0.537500,0.100000,0.100000},{0.937500,0.537500,0.050000,0.050000},{0.937500,0.537500,0.100000,0.100000},{0.962500,0.537500,0.050000,0.050000},{0.962500,0.537500,0.100000,0.100000},{0.987500,0.537500,0.050000,0.050000},{0.987500,0.537500,0.100000,0.100000},{0.012500,0.562500,0.050000,0.050000},{0.012500,0.562500,0.100000,0.100000},{0.037500,0.562500,0.050000,0.050000},{0.037500,0.562500,0.100000,0.100000},{0.062500,0.562500,0.050000,0.050000},{0.062500,0.562500,0.100000,0.100000},{0.087500,0.562500,0.050000,0.050000},{0.087500,0.562500,0.100000,0.100000},{0.112500,0.562500,0.050000,0.050000},{0.112500,0.562500,0.100000,0.100000},{0.137500,0.562500,0.050000,0.050000},{0.137500,0.562500,0.100000,0.100000},{0.162500,0.562500,0.050000,0.050000},{0.162500,0.562500,0.100000,0.100000},{0.187500,0.562500,0.050000,0.050000},{0.187500,0.562500,0.100000,0.100000},{0.212500,0.562500,0.050000,0.050000},{0.212500,0.562500,0.100000,0.100000},{0.237500,0.562500,0.050000,0.050000},{0.237500,0.562500,0.100000,0.100000},{0.262500,0.562500,0.050000,0.050000},{0.262500,0.562500,0.100000,0.100000},{0.287500,0.562500,0.050000,0.050000},{0.287500,0.562500,0.100000,0.100000},{0.312500,0.562500,0.050000,0.050000},{0.312500,0.562500,0.100000,0.100000},{0.337500,0.562500,0.050000,0.050000},{0.337500,0.562500,0.100000,0.100000},{0.362500,0.562500,0.050000,0.050000},{0.362500,0.562500,0.100000,0.100000},{0.387500,0.562500,0.050000,0.050000},{0.387500,0.562500,0.100000,0.100000},{0.412500,0.562500,0.050000,0.050000},{0.412500,0.562500,0.100000,0.100000},{0.437500,0.562500,0.050000,0.050000},{0.437500,0.562500,0.100000,0.100000},{0.462500,0.562500,0.050000,0.050000},{0.462500,0.562500,0.100000,0.100000},{0.487500,0.562500,0.050000,0.050000},{0.487500,0.562500,0.100000,0.100000},{0.512500,0.562500,0.050000,0.050000},{0.512500,0.562500,0.100000,0.100000},{0.537500,0.562500,0.050000,0.050000},{0.537500,0.562500,0.100000,0.100000},{0.562500,0.562500,0.050000,0.050000},{0.562500,0.562500,0.100000,0.100000},{0.587500,0.562500,0.050000,0.050000},{0.587500,0.562500,0.100000,0.100000},{0.612500,0.562500,0.050000,0.050000},{0.612500,0.562500,0.100000,0.100000},{0.637500,0.562500,0.050000,0.050000},{0.637500,0.562500,0.100000,0.100000},{0.662500,0.562500,0.050000,0.050000},{0.662500,0.562500,0.100000,0.100000},{0.687500,0.562500,0.050000,0.050000},{0.687500,0.562500,0.100000,0.100000},{0.712500,0.562500,0.050000,0.050000},{0.712500,0.562500,0.100000,0.100000},{0.737500,0.562500,0.050000,0.050000},{0.737500,0.562500,0.100000,0.100000},{0.762500,0.562500,0.050000,0.050000},{0.762500,0.562500,0.100000,0.100000},{0.787500,0.562500,0.050000,0.050000},{0.787500,0.562500,0.100000,0.100000},{0.812500,0.562500,0.050000,0.050000},{0.812500,0.562500,0.100000,0.100000},{0.837500,0.562500,0.050000,0.050000},{0.837500,0.562500,0.100000,0.100000},{0.862500,0.562500,0.050000,0.050000},{0.862500,0.562500,0.100000,0.100000},{0.887500,0.562500,0.050000,0.050000},{0.887500,0.562500,0.100000,0.100000},{0.912500,0.562500,0.050000,0.050000},{0.912500,0.562500,0.100000,0.100000},{0.937500,0.562500,0.050000,0.050000},{0.937500,0.562500,0.100000,0.100000},{0.962500,0.562500,0.050000,0.050000},{0.962500,0.562500,0.100000,0.100000},{0.987500,0.562500,0.050000,0.050000},{0.987500,0.562500,0.100000,0.100000},{0.012500,0.587500,0.050000,0.050000},{0.012500,0.587500,0.100000,0.100000},{0.037500,0.587500,0.050000,0.050000},{0.037500,0.587500,0.100000,0.100000},{0.062500,0.587500,0.050000,0.050000},{0.062500,0.587500,0.100000,0.100000},{0.087500,0.587500,0.050000,0.050000},{0.087500,0.587500,0.100000,0.100000},{0.112500,0.587500,0.050000,0.050000},{0.112500,0.587500,0.100000,0.100000},{0.137500,0.587500,0.050000,0.050000},{0.137500,0.587500,0.100000,0.100000},{0.162500,0.587500,0.050000,0.050000},{0.162500,0.587500,0.100000,0.100000},{0.187500,0.587500,0.050000,0.050000},{0.187500,0.587500,0.100000,0.100000},{0.212500,0.587500,0.050000,0.050000},{0.212500,0.587500,0.100000,0.100000},{0.237500,0.587500,0.050000,0.050000},{0.237500,0.587500,0.100000,0.100000},{0.262500,0.587500,0.050000,0.050000},{0.262500,0.587500,0.100000,0.100000},{0.287500,0.587500,0.050000,0.050000},{0.287500,0.587500,0.100000,0.100000},{0.312500,0.587500,0.050000,0.050000},{0.312500,0.587500,0.100000,0.100000},{0.337500,0.587500,0.050000,0.050000},{0.337500,0.587500,0.100000,0.100000},{0.362500,0.587500,0.050000,0.050000},{0.362500,0.587500,0.100000,0.100000},{0.387500,0.587500,0.050000,0.050000},{0.387500,0.587500,0.100000,0.100000},{0.412500,0.587500,0.050000,0.050000},{0.412500,0.587500,0.100000,0.100000},{0.437500,0.587500,0.050000,0.050000},{0.437500,0.587500,0.100000,0.100000},{0.462500,0.587500,0.050000,0.050000},{0.462500,0.587500,0.100000,0.100000},{0.487500,0.587500,0.050000,0.050000},{0.487500,0.587500,0.100000,0.100000},{0.512500,0.587500,0.050000,0.050000},{0.512500,0.587500,0.100000,0.100000},{0.537500,0.587500,0.050000,0.050000},{0.537500,0.587500,0.100000,0.100000},{0.562500,0.587500,0.050000,0.050000},{0.562500,0.587500,0.100000,0.100000},{0.587500,0.587500,0.050000,0.050000},{0.587500,0.587500,0.100000,0.100000},{0.612500,0.587500,0.050000,0.050000},{0.612500,0.587500,0.100000,0.100000},{0.637500,0.587500,0.050000,0.050000},{0.637500,0.587500,0.100000,0.100000},{0.662500,0.587500,0.050000,0.050000},{0.662500,0.587500,0.100000,0.100000},{0.687500,0.587500,0.050000,0.050000},{0.687500,0.587500,0.100000,0.100000},{0.712500,0.587500,0.050000,0.050000},{0.712500,0.587500,0.100000,0.100000},{0.737500,0.587500,0.050000,0.050000},{0.737500,0.587500,0.100000,0.100000},{0.762500,0.587500,0.050000,0.050000},{0.762500,0.587500,0.100000,0.100000},{0.787500,0.587500,0.050000,0.050000},{0.787500,0.587500,0.100000,0.100000},{0.812500,0.587500,0.050000,0.050000},{0.812500,0.587500,0.100000,0.100000},{0.837500,0.587500,0.050000,0.050000},{0.837500,0.587500,0.100000,0.100000},{0.862500,0.587500,0.050000,0.050000},{0.862500,0.587500,0.100000,0.100000},{0.887500,0.587500,0.050000,0.050000},{0.887500,0.587500,0.100000,0.100000},{0.912500,0.587500,0.050000,0.050000},{0.912500,0.587500,0.100000,0.100000},{0.937500,0.587500,0.050000,0.050000},{0.937500,0.587500,0.100000,0.100000},{0.962500,0.587500,0.050000,0.050000},{0.962500,0.587500,0.100000,0.100000},{0.987500,0.587500,0.050000,0.050000},{0.987500,0.587500,0.100000,0.100000},{0.012500,0.612500,0.050000,0.050000},{0.012500,0.612500,0.100000,0.100000},{0.037500,0.612500,0.050000,0.050000},{0.037500,0.612500,0.100000,0.100000},{0.062500,0.612500,0.050000,0.050000},{0.062500,0.612500,0.100000,0.100000},{0.087500,0.612500,0.050000,0.050000},{0.087500,0.612500,0.100000,0.100000},{0.112500,0.612500,0.050000,0.050000},{0.112500,0.612500,0.100000,0.100000},{0.137500,0.612500,0.050000,0.050000},{0.137500,0.612500,0.100000,0.100000},{0.162500,0.612500,0.050000,0.050000},{0.162500,0.612500,0.100000,0.100000},{0.187500,0.612500,0.050000,0.050000},{0.187500,0.612500,0.100000,0.100000},{0.212500,0.612500,0.050000,0.050000},{0.212500,0.612500,0.100000,0.100000},{0.237500,0.612500,0.050000,0.050000},{0.237500,0.612500,0.100000,0.100000},{0.262500,0.612500,0.050000,0.050000},{0.262500,0.612500,0.100000,0.100000},{0.287500,0.612500,0.050000,0.050000},{0.287500,0.612500,0.100000,0.100000},{0.312500,0.612500,0.050000,0.050000},{0.312500,0.612500,0.100000,0.100000},{0.337500,0.612500,0.050000,0.050000},{0.337500,0.612500,0.100000,0.100000},{0.362500,0.612500,0.050000,0.050000},{0.362500,0.612500,0.100000,0.100000},{0.387500,0.612500,0.050000,0.050000},{0.387500,0.612500,0.100000,0.100000},{0.412500,0.612500,0.050000,0.050000},{0.412500,0.612500,0.100000,0.100000},{0.437500,0.612500,0.050000,0.050000},{0.437500,0.612500,0.100000,0.100000},{0.462500,0.612500,0.050000,0.050000},{0.462500,0.612500,0.100000,0.100000},{0.487500,0.612500,0.050000,0.050000},{0.487500,0.612500,0.100000,0.100000},{0.512500,0.612500,0.050000,0.050000},{0.512500,0.612500,0.100000,0.100000},{0.537500,0.612500,0.050000,0.050000},{0.537500,0.612500,0.100000,0.100000},{0.562500,0.612500,0.050000,0.050000},{0.562500,0.612500,0.100000,0.100000},{0.587500,0.612500,0.050000,0.050000},{0.587500,0.612500,0.100000,0.100000},{0.612500,0.612500,0.050000,0.050000},{0.612500,0.612500,0.100000,0.100000},{0.637500,0.612500,0.050000,0.050000},{0.637500,0.612500,0.100000,0.100000},{0.662500,0.612500,0.050000,0.050000},{0.662500,0.612500,0.100000,0.100000},{0.687500,0.612500,0.050000,0.050000},{0.687500,0.612500,0.100000,0.100000},{0.712500,0.612500,0.050000,0.050000},{0.712500,0.612500,0.100000,0.100000},{0.737500,0.612500,0.050000,0.050000},{0.737500,0.612500,0.100000,0.100000},{0.762500,0.612500,0.050000,0.050000},{0.762500,0.612500,0.100000,0.100000},{0.787500,0.612500,0.050000,0.050000},{0.787500,0.612500,0.100000,0.100000},{0.812500,0.612500,0.050000,0.050000},{0.812500,0.612500,0.100000,0.100000},{0.837500,0.612500,0.050000,0.050000},{0.837500,0.612500,0.100000,0.100000},{0.862500,0.612500,0.050000,0.050000},{0.862500,0.612500,0.100000,0.100000},{0.887500,0.612500,0.050000,0.050000},{0.887500,0.612500,0.100000,0.100000},{0.912500,0.612500,0.050000,0.050000},{0.912500,0.612500,0.100000,0.100000},{0.937500,0.612500,0.050000,0.050000},{0.937500,0.612500,0.100000,0.100000},{0.962500,0.612500,0.050000,0.050000},{0.962500,0.612500,0.100000,0.100000},{0.987500,0.612500,0.050000,0.050000},{0.987500,0.612500,0.100000,0.100000},{0.012500,0.637500,0.050000,0.050000},{0.012500,0.637500,0.100000,0.100000},{0.037500,0.637500,0.050000,0.050000},{0.037500,0.637500,0.100000,0.100000},{0.062500,0.637500,0.050000,0.050000},{0.062500,0.637500,0.100000,0.100000},{0.087500,0.637500,0.050000,0.050000},{0.087500,0.637500,0.100000,0.100000},{0.112500,0.637500,0.050000,0.050000},{0.112500,0.637500,0.100000,0.100000},{0.137500,0.637500,0.050000,0.050000},{0.137500,0.637500,0.100000,0.100000},{0.162500,0.637500,0.050000,0.050000},{0.162500,0.637500,0.100000,0.100000},{0.187500,0.637500,0.050000,0.050000},{0.187500,0.637500,0.100000,0.100000},{0.212500,0.637500,0.050000,0.050000},{0.212500,0.637500,0.100000,0.100000},{0.237500,0.637500,0.050000,0.050000},{0.237500,0.637500,0.100000,0.100000},{0.262500,0.637500,0.050000,0.050000},{0.262500,0.637500,0.100000,0.100000},{0.287500,0.637500,0.050000,0.050000},{0.287500,0.637500,0.100000,0.100000},{0.312500,0.637500,0.050000,0.050000},{0.312500,0.637500,0.100000,0.100000},{0.337500,0.637500,0.050000,0.050000},{0.337500,0.637500,0.100000,0.100000},{0.362500,0.637500,0.050000,0.050000},{0.362500,0.637500,0.100000,0.100000},{0.387500,0.637500,0.050000,0.050000},{0.387500,0.637500,0.100000,0.100000},{0.412500,0.637500,0.050000,0.050000},{0.412500,0.637500,0.100000,0.100000},{0.437500,0.637500,0.050000,0.050000},{0.437500,0.637500,0.100000,0.100000},{0.462500,0.637500,0.050000,0.050000},{0.462500,0.637500,0.100000,0.100000},{0.487500,0.637500,0.050000,0.050000},{0.487500,0.637500,0.100000,0.100000},{0.512500,0.637500,0.050000,0.050000},{0.512500,0.637500,0.100000,0.100000},{0.537500,0.637500,0.050000,0.050000},{0.537500,0.637500,0.100000,0.100000},{0.562500,0.637500,0.050000,0.050000},{0.562500,0.637500,0.100000,0.100000},{0.587500,0.637500,0.050000,0.050000},{0.587500,0.637500,0.100000,0.100000},{0.612500,0.637500,0.050000,0.050000},{0.612500,0.637500,0.100000,0.100000},{0.637500,0.637500,0.050000,0.050000},{0.637500,0.637500,0.100000,0.100000},{0.662500,0.637500,0.050000,0.050000},{0.662500,0.637500,0.100000,0.100000},{0.687500,0.637500,0.050000,0.050000},{0.687500,0.637500,0.100000,0.100000},{0.712500,0.637500,0.050000,0.050000},{0.712500,0.637500,0.100000,0.100000},{0.737500,0.637500,0.050000,0.050000},{0.737500,0.637500,0.100000,0.100000},{0.762500,0.637500,0.050000,0.050000},{0.762500,0.637500,0.100000,0.100000},{0.787500,0.637500,0.050000,0.050000},{0.787500,0.637500,0.100000,0.100000},{0.812500,0.637500,0.050000,0.050000},{0.812500,0.637500,0.100000,0.100000},{0.837500,0.637500,0.050000,0.050000},{0.837500,0.637500,0.100000,0.100000},{0.862500,0.637500,0.050000,0.050000},{0.862500,0.637500,0.100000,0.100000},{0.887500,0.637500,0.050000,0.050000},{0.887500,0.637500,0.100000,0.100000},{0.912500,0.637500,0.050000,0.050000},{0.912500,0.637500,0.100000,0.100000},{0.937500,0.637500,0.050000,0.050000},{0.937500,0.637500,0.100000,0.100000},{0.962500,0.637500,0.050000,0.050000},{0.962500,0.637500,0.100000,0.100000},{0.987500,0.637500,0.050000,0.050000},{0.987500,0.637500,0.100000,0.100000},{0.012500,0.662500,0.050000,0.050000},{0.012500,0.662500,0.100000,0.100000},{0.037500,0.662500,0.050000,0.050000},{0.037500,0.662500,0.100000,0.100000},{0.062500,0.662500,0.050000,0.050000},{0.062500,0.662500,0.100000,0.100000},{0.087500,0.662500,0.050000,0.050000},{0.087500,0.662500,0.100000,0.100000},{0.112500,0.662500,0.050000,0.050000},{0.112500,0.662500,0.100000,0.100000},{0.137500,0.662500,0.050000,0.050000},{0.137500,0.662500,0.100000,0.100000},{0.162500,0.662500,0.050000,0.050000},{0.162500,0.662500,0.100000,0.100000},{0.187500,0.662500,0.050000,0.050000},{0.187500,0.662500,0.100000,0.100000},{0.212500,0.662500,0.050000,0.050000},{0.212500,0.662500,0.100000,0.100000},{0.237500,0.662500,0.050000,0.050000},{0.237500,0.662500,0.100000,0.100000},{0.262500,0.662500,0.050000,0.050000},{0.262500,0.662500,0.100000,0.100000},{0.287500,0.662500,0.050000,0.050000},{0.287500,0.662500,0.100000,0.100000},{0.312500,0.662500,0.050000,0.050000},{0.312500,0.662500,0.100000,0.100000},{0.337500,0.662500,0.050000,0.050000},{0.337500,0.662500,0.100000,0.100000},{0.362500,0.662500,0.050000,0.050000},{0.362500,0.662500,0.100000,0.100000},{0.387500,0.662500,0.050000,0.050000},{0.387500,0.662500,0.100000,0.100000},{0.412500,0.662500,0.050000,0.050000},{0.412500,0.662500,0.100000,0.100000},{0.437500,0.662500,0.050000,0.050000},{0.437500,0.662500,0.100000,0.100000},{0.462500,0.662500,0.050000,0.050000},{0.462500,0.662500,0.100000,0.100000},{0.487500,0.662500,0.050000,0.050000},{0.487500,0.662500,0.100000,0.100000},{0.512500,0.662500,0.050000,0.050000},{0.512500,0.662500,0.100000,0.100000},{0.537500,0.662500,0.050000,0.050000},{0.537500,0.662500,0.100000,0.100000},{0.562500,0.662500,0.050000,0.050000},{0.562500,0.662500,0.100000,0.100000},{0.587500,0.662500,0.050000,0.050000},{0.587500,0.662500,0.100000,0.100000},{0.612500,0.662500,0.050000,0.050000},{0.612500,0.662500,0.100000,0.100000},{0.637500,0.662500,0.050000,0.050000},{0.637500,0.662500,0.100000,0.100000},{0.662500,0.662500,0.050000,0.050000},{0.662500,0.662500,0.100000,0.100000},{0.687500,0.662500,0.050000,0.050000},{0.687500,0.662500,0.100000,0.100000},{0.712500,0.662500,0.050000,0.050000},{0.712500,0.662500,0.100000,0.100000},{0.737500,0.662500,0.050000,0.050000},{0.737500,0.662500,0.100000,0.100000},{0.762500,0.662500,0.050000,0.050000},{0.762500,0.662500,0.100000,0.100000},{0.787500,0.662500,0.050000,0.050000},{0.787500,0.662500,0.100000,0.100000},{0.812500,0.662500,0.050000,0.050000},{0.812500,0.662500,0.100000,0.100000},{0.837500,0.662500,0.050000,0.050000},{0.837500,0.662500,0.100000,0.100000},{0.862500,0.662500,0.050000,0.050000},{0.862500,0.662500,0.100000,0.100000},{0.887500,0.662500,0.050000,0.050000},{0.887500,0.662500,0.100000,0.100000},{0.912500,0.662500,0.050000,0.050000},{0.912500,0.662500,0.100000,0.100000},{0.937500,0.662500,0.050000,0.050000},{0.937500,0.662500,0.100000,0.100000},{0.962500,0.662500,0.050000,0.050000},{0.962500,0.662500,0.100000,0.100000},{0.987500,0.662500,0.050000,0.050000},{0.987500,0.662500,0.100000,0.100000},{0.012500,0.687500,0.050000,0.050000},{0.012500,0.687500,0.100000,0.100000},{0.037500,0.687500,0.050000,0.050000},{0.037500,0.687500,0.100000,0.100000},{0.062500,0.687500,0.050000,0.050000},{0.062500,0.687500,0.100000,0.100000},{0.087500,0.687500,0.050000,0.050000},{0.087500,0.687500,0.100000,0.100000},{0.112500,0.687500,0.050000,0.050000},{0.112500,0.687500,0.100000,0.100000},{0.137500,0.687500,0.050000,0.050000},{0.137500,0.687500,0.100000,0.100000},{0.162500,0.687500,0.050000,0.050000},{0.162500,0.687500,0.100000,0.100000},{0.187500,0.687500,0.050000,0.050000},{0.187500,0.687500,0.100000,0.100000},{0.212500,0.687500,0.050000,0.050000},{0.212500,0.687500,0.100000,0.100000},{0.237500,0.687500,0.050000,0.050000},{0.237500,0.687500,0.100000,0.100000},{0.262500,0.687500,0.050000,0.050000},{0.262500,0.687500,0.100000,0.100000},{0.287500,0.687500,0.050000,0.050000},{0.287500,0.687500,0.100000,0.100000},{0.312500,0.687500,0.050000,0.050000},{0.312500,0.687500,0.100000,0.100000},{0.337500,0.687500,0.050000,0.050000},{0.337500,0.687500,0.100000,0.100000},{0.362500,0.687500,0.050000,0.050000},{0.362500,0.687500,0.100000,0.100000},{0.387500,0.687500,0.050000,0.050000},{0.387500,0.687500,0.100000,0.100000},{0.412500,0.687500,0.050000,0.050000},{0.412500,0.687500,0.100000,0.100000},{0.437500,0.687500,0.050000,0.050000},{0.437500,0.687500,0.100000,0.100000},{0.462500,0.687500,0.050000,0.050000},{0.462500,0.687500,0.100000,0.100000},{0.487500,0.687500,0.050000,0.050000},{0.487500,0.687500,0.100000,0.100000},{0.512500,0.687500,0.050000,0.050000},{0.512500,0.687500,0.100000,0.100000},{0.537500,0.687500,0.050000,0.050000},{0.537500,0.687500,0.100000,0.100000},{0.562500,0.687500,0.050000,0.050000},{0.562500,0.687500,0.100000,0.100000},{0.587500,0.687500,0.050000,0.050000},{0.587500,0.687500,0.100000,0.100000},{0.612500,0.687500,0.050000,0.050000},{0.612500,0.687500,0.100000,0.100000},{0.637500,0.687500,0.050000,0.050000},{0.637500,0.687500,0.100000,0.100000},{0.662500,0.687500,0.050000,0.050000},{0.662500,0.687500,0.100000,0.100000},{0.687500,0.687500,0.050000,0.050000},{0.687500,0.687500,0.100000,0.100000},{0.712500,0.687500,0.050000,0.050000},{0.712500,0.687500,0.100000,0.100000},{0.737500,0.687500,0.050000,0.050000},{0.737500,0.687500,0.100000,0.100000},{0.762500,0.687500,0.050000,0.050000},{0.762500,0.687500,0.100000,0.100000},{0.787500,0.687500,0.050000,0.050000},{0.787500,0.687500,0.100000,0.100000},{0.812500,0.687500,0.050000,0.050000},{0.812500,0.687500,0.100000,0.100000},{0.837500,0.687500,0.050000,0.050000},{0.837500,0.687500,0.100000,0.100000},{0.862500,0.687500,0.050000,0.050000},{0.862500,0.687500,0.100000,0.100000},{0.887500,0.687500,0.050000,0.050000},{0.887500,0.687500,0.100000,0.100000},{0.912500,0.687500,0.050000,0.050000},{0.912500,0.687500,0.100000,0.100000},{0.937500,0.687500,0.050000,0.050000},{0.937500,0.687500,0.100000,0.100000},{0.962500,0.687500,0.050000,0.050000},{0.962500,0.687500,0.100000,0.100000},{0.987500,0.687500,0.050000,0.050000},{0.987500,0.687500,0.100000,0.100000},{0.012500,0.712500,0.050000,0.050000},{0.012500,0.712500,0.100000,0.100000},{0.037500,0.712500,0.050000,0.050000},{0.037500,0.712500,0.100000,0.100000},{0.062500,0.712500,0.050000,0.050000},{0.062500,0.712500,0.100000,0.100000},{0.087500,0.712500,0.050000,0.050000},{0.087500,0.712500,0.100000,0.100000},{0.112500,0.712500,0.050000,0.050000},{0.112500,0.712500,0.100000,0.100000},{0.137500,0.712500,0.050000,0.050000},{0.137500,0.712500,0.100000,0.100000},{0.162500,0.712500,0.050000,0.050000},{0.162500,0.712500,0.100000,0.100000},{0.187500,0.712500,0.050000,0.050000},{0.187500,0.712500,0.100000,0.100000},{0.212500,0.712500,0.050000,0.050000},{0.212500,0.712500,0.100000,0.100000},{0.237500,0.712500,0.050000,0.050000},{0.237500,0.712500,0.100000,0.100000},{0.262500,0.712500,0.050000,0.050000},{0.262500,0.712500,0.100000,0.100000},{0.287500,0.712500,0.050000,0.050000},{0.287500,0.712500,0.100000,0.100000},{0.312500,0.712500,0.050000,0.050000},{0.312500,0.712500,0.100000,0.100000},{0.337500,0.712500,0.050000,0.050000},{0.337500,0.712500,0.100000,0.100000},{0.362500,0.712500,0.050000,0.050000},{0.362500,0.712500,0.100000,0.100000},{0.387500,0.712500,0.050000,0.050000},{0.387500,0.712500,0.100000,0.100000},{0.412500,0.712500,0.050000,0.050000},{0.412500,0.712500,0.100000,0.100000},{0.437500,0.712500,0.050000,0.050000},{0.437500,0.712500,0.100000,0.100000},{0.462500,0.712500,0.050000,0.050000},{0.462500,0.712500,0.100000,0.100000},{0.487500,0.712500,0.050000,0.050000},{0.487500,0.712500,0.100000,0.100000},{0.512500,0.712500,0.050000,0.050000},{0.512500,0.712500,0.100000,0.100000},{0.537500,0.712500,0.050000,0.050000},{0.537500,0.712500,0.100000,0.100000},{0.562500,0.712500,0.050000,0.050000},{0.562500,0.712500,0.100000,0.100000},{0.587500,0.712500,0.050000,0.050000},{0.587500,0.712500,0.100000,0.100000},{0.612500,0.712500,0.050000,0.050000},{0.612500,0.712500,0.100000,0.100000},{0.637500,0.712500,0.050000,0.050000},{0.637500,0.712500,0.100000,0.100000},{0.662500,0.712500,0.050000,0.050000},{0.662500,0.712500,0.100000,0.100000},{0.687500,0.712500,0.050000,0.050000},{0.687500,0.712500,0.100000,0.100000},{0.712500,0.712500,0.050000,0.050000},{0.712500,0.712500,0.100000,0.100000},{0.737500,0.712500,0.050000,0.050000},{0.737500,0.712500,0.100000,0.100000},{0.762500,0.712500,0.050000,0.050000},{0.762500,0.712500,0.100000,0.100000},{0.787500,0.712500,0.050000,0.050000},{0.787500,0.712500,0.100000,0.100000},{0.812500,0.712500,0.050000,0.050000},{0.812500,0.712500,0.100000,0.100000},{0.837500,0.712500,0.050000,0.050000},{0.837500,0.712500,0.100000,0.100000},{0.862500,0.712500,0.050000,0.050000},{0.862500,0.712500,0.100000,0.100000},{0.887500,0.712500,0.050000,0.050000},{0.887500,0.712500,0.100000,0.100000},{0.912500,0.712500,0.050000,0.050000},{0.912500,0.712500,0.100000,0.100000},{0.937500,0.712500,0.050000,0.050000},{0.937500,0.712500,0.100000,0.100000},{0.962500,0.712500,0.050000,0.050000},{0.962500,0.712500,0.100000,0.100000},{0.987500,0.712500,0.050000,0.050000},{0.987500,0.712500,0.100000,0.100000},{0.012500,0.737500,0.050000,0.050000},{0.012500,0.737500,0.100000,0.100000},{0.037500,0.737500,0.050000,0.050000},{0.037500,0.737500,0.100000,0.100000},{0.062500,0.737500,0.050000,0.050000},{0.062500,0.737500,0.100000,0.100000},{0.087500,0.737500,0.050000,0.050000},{0.087500,0.737500,0.100000,0.100000},{0.112500,0.737500,0.050000,0.050000},{0.112500,0.737500,0.100000,0.100000},{0.137500,0.737500,0.050000,0.050000},{0.137500,0.737500,0.100000,0.100000},{0.162500,0.737500,0.050000,0.050000},{0.162500,0.737500,0.100000,0.100000},{0.187500,0.737500,0.050000,0.050000},{0.187500,0.737500,0.100000,0.100000},{0.212500,0.737500,0.050000,0.050000},{0.212500,0.737500,0.100000,0.100000},{0.237500,0.737500,0.050000,0.050000},{0.237500,0.737500,0.100000,0.100000},{0.262500,0.737500,0.050000,0.050000},{0.262500,0.737500,0.100000,0.100000},{0.287500,0.737500,0.050000,0.050000},{0.287500,0.737500,0.100000,0.100000},{0.312500,0.737500,0.050000,0.050000},{0.312500,0.737500,0.100000,0.100000},{0.337500,0.737500,0.050000,0.050000},{0.337500,0.737500,0.100000,0.100000},{0.362500,0.737500,0.050000,0.050000},{0.362500,0.737500,0.100000,0.100000},{0.387500,0.737500,0.050000,0.050000},{0.387500,0.737500,0.100000,0.100000},{0.412500,0.737500,0.050000,0.050000},{0.412500,0.737500,0.100000,0.100000},{0.437500,0.737500,0.050000,0.050000},{0.437500,0.737500,0.100000,0.100000},{0.462500,0.737500,0.050000,0.050000},{0.462500,0.737500,0.100000,0.100000},{0.487500,0.737500,0.050000,0.050000},{0.487500,0.737500,0.100000,0.100000},{0.512500,0.737500,0.050000,0.050000},{0.512500,0.737500,0.100000,0.100000},{0.537500,0.737500,0.050000,0.050000},{0.537500,0.737500,0.100000,0.100000},{0.562500,0.737500,0.050000,0.050000},{0.562500,0.737500,0.100000,0.100000},{0.587500,0.737500,0.050000,0.050000},{0.587500,0.737500,0.100000,0.100000},{0.612500,0.737500,0.050000,0.050000},{0.612500,0.737500,0.100000,0.100000},{0.637500,0.737500,0.050000,0.050000},{0.637500,0.737500,0.100000,0.100000},{0.662500,0.737500,0.050000,0.050000},{0.662500,0.737500,0.100000,0.100000},{0.687500,0.737500,0.050000,0.050000},{0.687500,0.737500,0.100000,0.100000},{0.712500,0.737500,0.050000,0.050000},{0.712500,0.737500,0.100000,0.100000},{0.737500,0.737500,0.050000,0.050000},{0.737500,0.737500,0.100000,0.100000},{0.762500,0.737500,0.050000,0.050000},{0.762500,0.737500,0.100000,0.100000},{0.787500,0.737500,0.050000,0.050000},{0.787500,0.737500,0.100000,0.100000},{0.812500,0.737500,0.050000,0.050000},{0.812500,0.737500,0.100000,0.100000},{0.837500,0.737500,0.050000,0.050000},{0.837500,0.737500,0.100000,0.100000},{0.862500,0.737500,0.050000,0.050000},{0.862500,0.737500,0.100000,0.100000},{0.887500,0.737500,0.050000,0.050000},{0.887500,0.737500,0.100000,0.100000},{0.912500,0.737500,0.050000,0.050000},{0.912500,0.737500,0.100000,0.100000},{0.937500,0.737500,0.050000,0.050000},{0.937500,0.737500,0.100000,0.100000},{0.962500,0.737500,0.050000,0.050000},{0.962500,0.737500,0.100000,0.100000},{0.987500,0.737500,0.050000,0.050000},{0.987500,0.737500,0.100000,0.100000},{0.012500,0.762500,0.050000,0.050000},{0.012500,0.762500,0.100000,0.100000},{0.037500,0.762500,0.050000,0.050000},{0.037500,0.762500,0.100000,0.100000},{0.062500,0.762500,0.050000,0.050000},{0.062500,0.762500,0.100000,0.100000},{0.087500,0.762500,0.050000,0.050000},{0.087500,0.762500,0.100000,0.100000},{0.112500,0.762500,0.050000,0.050000},{0.112500,0.762500,0.100000,0.100000},{0.137500,0.762500,0.050000,0.050000},{0.137500,0.762500,0.100000,0.100000},{0.162500,0.762500,0.050000,0.050000},{0.162500,0.762500,0.100000,0.100000},{0.187500,0.762500,0.050000,0.050000},{0.187500,0.762500,0.100000,0.100000},{0.212500,0.762500,0.050000,0.050000},{0.212500,0.762500,0.100000,0.100000},{0.237500,0.762500,0.050000,0.050000},{0.237500,0.762500,0.100000,0.100000},{0.262500,0.762500,0.050000,0.050000},{0.262500,0.762500,0.100000,0.100000},{0.287500,0.762500,0.050000,0.050000},{0.287500,0.762500,0.100000,0.100000},{0.312500,0.762500,0.050000,0.050000},{0.312500,0.762500,0.100000,0.100000},{0.337500,0.762500,0.050000,0.050000},{0.337500,0.762500,0.100000,0.100000},{0.362500,0.762500,0.050000,0.050000},{0.362500,0.762500,0.100000,0.100000},{0.387500,0.762500,0.050000,0.050000},{0.387500,0.762500,0.100000,0.100000},{0.412500,0.762500,0.050000,0.050000},{0.412500,0.762500,0.100000,0.100000},{0.437500,0.762500,0.050000,0.050000},{0.437500,0.762500,0.100000,0.100000},{0.462500,0.762500,0.050000,0.050000},{0.462500,0.762500,0.100000,0.100000},{0.487500,0.762500,0.050000,0.050000},{0.487500,0.762500,0.100000,0.100000},{0.512500,0.762500,0.050000,0.050000},{0.512500,0.762500,0.100000,0.100000},{0.537500,0.762500,0.050000,0.050000},{0.537500,0.762500,0.100000,0.100000},{0.562500,0.762500,0.050000,0.050000},{0.562500,0.762500,0.100000,0.100000},{0.587500,0.762500,0.050000,0.050000},{0.587500,0.762500,0.100000,0.100000},{0.612500,0.762500,0.050000,0.050000},{0.612500,0.762500,0.100000,0.100000},{0.637500,0.762500,0.050000,0.050000},{0.637500,0.762500,0.100000,0.100000},{0.662500,0.762500,0.050000,0.050000},{0.662500,0.762500,0.100000,0.100000},{0.687500,0.762500,0.050000,0.050000},{0.687500,0.762500,0.100000,0.100000},{0.712500,0.762500,0.050000,0.050000},{0.712500,0.762500,0.100000,0.100000},{0.737500,0.762500,0.050000,0.050000},{0.737500,0.762500,0.100000,0.100000},{0.762500,0.762500,0.050000,0.050000},{0.762500,0.762500,0.100000,0.100000},{0.787500,0.762500,0.050000,0.050000},{0.787500,0.762500,0.100000,0.100000},{0.812500,0.762500,0.050000,0.050000},{0.812500,0.762500,0.100000,0.100000},{0.837500,0.762500,0.050000,0.050000},{0.837500,0.762500,0.100000,0.100000},{0.862500,0.762500,0.050000,0.050000},{0.862500,0.762500,0.100000,0.100000},{0.887500,0.762500,0.050000,0.050000},{0.887500,0.762500,0.100000,0.100000},{0.912500,0.762500,0.050000,0.050000},{0.912500,0.762500,0.100000,0.100000},{0.937500,0.762500,0.050000,0.050000},{0.937500,0.762500,0.100000,0.100000},{0.962500,0.762500,0.050000,0.050000},{0.962500,0.762500,0.100000,0.100000},{0.987500,0.762500,0.050000,0.050000},{0.987500,0.762500,0.100000,0.100000},{0.012500,0.787500,0.050000,0.050000},{0.012500,0.787500,0.100000,0.100000},{0.037500,0.787500,0.050000,0.050000},{0.037500,0.787500,0.100000,0.100000},{0.062500,0.787500,0.050000,0.050000},{0.062500,0.787500,0.100000,0.100000},{0.087500,0.787500,0.050000,0.050000},{0.087500,0.787500,0.100000,0.100000},{0.112500,0.787500,0.050000,0.050000},{0.112500,0.787500,0.100000,0.100000},{0.137500,0.787500,0.050000,0.050000},{0.137500,0.787500,0.100000,0.100000},{0.162500,0.787500,0.050000,0.050000},{0.162500,0.787500,0.100000,0.100000},{0.187500,0.787500,0.050000,0.050000},{0.187500,0.787500,0.100000,0.100000},{0.212500,0.787500,0.050000,0.050000},{0.212500,0.787500,0.100000,0.100000},{0.237500,0.787500,0.050000,0.050000},{0.237500,0.787500,0.100000,0.100000},{0.262500,0.787500,0.050000,0.050000},{0.262500,0.787500,0.100000,0.100000},{0.287500,0.787500,0.050000,0.050000},{0.287500,0.787500,0.100000,0.100000},{0.312500,0.787500,0.050000,0.050000},{0.312500,0.787500,0.100000,0.100000},{0.337500,0.787500,0.050000,0.050000},{0.337500,0.787500,0.100000,0.100000},{0.362500,0.787500,0.050000,0.050000},{0.362500,0.787500,0.100000,0.100000},{0.387500,0.787500,0.050000,0.050000},{0.387500,0.787500,0.100000,0.100000},{0.412500,0.787500,0.050000,0.050000},{0.412500,0.787500,0.100000,0.100000},{0.437500,0.787500,0.050000,0.050000},{0.437500,0.787500,0.100000,0.100000},{0.462500,0.787500,0.050000,0.050000},{0.462500,0.787500,0.100000,0.100000},{0.487500,0.787500,0.050000,0.050000},{0.487500,0.787500,0.100000,0.100000},{0.512500,0.787500,0.050000,0.050000},{0.512500,0.787500,0.100000,0.100000},{0.537500,0.787500,0.050000,0.050000},{0.537500,0.787500,0.100000,0.100000},{0.562500,0.787500,0.050000,0.050000},{0.562500,0.787500,0.100000,0.100000},{0.587500,0.787500,0.050000,0.050000},{0.587500,0.787500,0.100000,0.100000},{0.612500,0.787500,0.050000,0.050000},{0.612500,0.787500,0.100000,0.100000},{0.637500,0.787500,0.050000,0.050000},{0.637500,0.787500,0.100000,0.100000},{0.662500,0.787500,0.050000,0.050000},{0.662500,0.787500,0.100000,0.100000},{0.687500,0.787500,0.050000,0.050000},{0.687500,0.787500,0.100000,0.100000},{0.712500,0.787500,0.050000,0.050000},{0.712500,0.787500,0.100000,0.100000},{0.737500,0.787500,0.050000,0.050000},{0.737500,0.787500,0.100000,0.100000},{0.762500,0.787500,0.050000,0.050000},{0.762500,0.787500,0.100000,0.100000},{0.787500,0.787500,0.050000,0.050000},{0.787500,0.787500,0.100000,0.100000},{0.812500,0.787500,0.050000,0.050000},{0.812500,0.787500,0.100000,0.100000},{0.837500,0.787500,0.050000,0.050000},{0.837500,0.787500,0.100000,0.100000},{0.862500,0.787500,0.050000,0.050000},{0.862500,0.787500,0.100000,0.100000},{0.887500,0.787500,0.050000,0.050000},{0.887500,0.787500,0.100000,0.100000},{0.912500,0.787500,0.050000,0.050000},{0.912500,0.787500,0.100000,0.100000},{0.937500,0.787500,0.050000,0.050000},{0.937500,0.787500,0.100000,0.100000},{0.962500,0.787500,0.050000,0.050000},{0.962500,0.787500,0.100000,0.100000},{0.987500,0.787500,0.050000,0.050000},{0.987500,0.787500,0.100000,0.100000},{0.012500,0.812500,0.050000,0.050000},{0.012500,0.812500,0.100000,0.100000},{0.037500,0.812500,0.050000,0.050000},{0.037500,0.812500,0.100000,0.100000},{0.062500,0.812500,0.050000,0.050000},{0.062500,0.812500,0.100000,0.100000},{0.087500,0.812500,0.050000,0.050000},{0.087500,0.812500,0.100000,0.100000},{0.112500,0.812500,0.050000,0.050000},{0.112500,0.812500,0.100000,0.100000},{0.137500,0.812500,0.050000,0.050000},{0.137500,0.812500,0.100000,0.100000},{0.162500,0.812500,0.050000,0.050000},{0.162500,0.812500,0.100000,0.100000},{0.187500,0.812500,0.050000,0.050000},{0.187500,0.812500,0.100000,0.100000},{0.212500,0.812500,0.050000,0.050000},{0.212500,0.812500,0.100000,0.100000},{0.237500,0.812500,0.050000,0.050000},{0.237500,0.812500,0.100000,0.100000},{0.262500,0.812500,0.050000,0.050000},{0.262500,0.812500,0.100000,0.100000},{0.287500,0.812500,0.050000,0.050000},{0.287500,0.812500,0.100000,0.100000},{0.312500,0.812500,0.050000,0.050000},{0.312500,0.812500,0.100000,0.100000},{0.337500,0.812500,0.050000,0.050000},{0.337500,0.812500,0.100000,0.100000},{0.362500,0.812500,0.050000,0.050000},{0.362500,0.812500,0.100000,0.100000},{0.387500,0.812500,0.050000,0.050000},{0.387500,0.812500,0.100000,0.100000},{0.412500,0.812500,0.050000,0.050000},{0.412500,0.812500,0.100000,0.100000},{0.437500,0.812500,0.050000,0.050000},{0.437500,0.812500,0.100000,0.100000},{0.462500,0.812500,0.050000,0.050000},{0.462500,0.812500,0.100000,0.100000},{0.487500,0.812500,0.050000,0.050000},{0.487500,0.812500,0.100000,0.100000},{0.512500,0.812500,0.050000,0.050000},{0.512500,0.812500,0.100000,0.100000},{0.537500,0.812500,0.050000,0.050000},{0.537500,0.812500,0.100000,0.100000},{0.562500,0.812500,0.050000,0.050000},{0.562500,0.812500,0.100000,0.100000},{0.587500,0.812500,0.050000,0.050000},{0.587500,0.812500,0.100000,0.100000},{0.612500,0.812500,0.050000,0.050000},{0.612500,0.812500,0.100000,0.100000},{0.637500,0.812500,0.050000,0.050000},{0.637500,0.812500,0.100000,0.100000},{0.662500,0.812500,0.050000,0.050000},{0.662500,0.812500,0.100000,0.100000},{0.687500,0.812500,0.050000,0.050000},{0.687500,0.812500,0.100000,0.100000},{0.712500,0.812500,0.050000,0.050000},{0.712500,0.812500,0.100000,0.100000},{0.737500,0.812500,0.050000,0.050000},{0.737500,0.812500,0.100000,0.100000},{0.762500,0.812500,0.050000,0.050000},{0.762500,0.812500,0.100000,0.100000},{0.787500,0.812500,0.050000,0.050000},{0.787500,0.812500,0.100000,0.100000},{0.812500,0.812500,0.050000,0.050000},{0.812500,0.812500,0.100000,0.100000},{0.837500,0.812500,0.050000,0.050000},{0.837500,0.812500,0.100000,0.100000},{0.862500,0.812500,0.050000,0.050000},{0.862500,0.812500,0.100000,0.100000},{0.887500,0.812500,0.050000,0.050000},{0.887500,0.812500,0.100000,0.100000},{0.912500,0.812500,0.050000,0.050000},{0.912500,0.812500,0.100000,0.100000},{0.937500,0.812500,0.050000,0.050000},{0.937500,0.812500,0.100000,0.100000},{0.962500,0.812500,0.050000,0.050000},{0.962500,0.812500,0.100000,0.100000},{0.987500,0.812500,0.050000,0.050000},{0.987500,0.812500,0.100000,0.100000},{0.012500,0.837500,0.050000,0.050000},{0.012500,0.837500,0.100000,0.100000},{0.037500,0.837500,0.050000,0.050000},{0.037500,0.837500,0.100000,0.100000},{0.062500,0.837500,0.050000,0.050000},{0.062500,0.837500,0.100000,0.100000},{0.087500,0.837500,0.050000,0.050000},{0.087500,0.837500,0.100000,0.100000},{0.112500,0.837500,0.050000,0.050000},{0.112500,0.837500,0.100000,0.100000},{0.137500,0.837500,0.050000,0.050000},{0.137500,0.837500,0.100000,0.100000},{0.162500,0.837500,0.050000,0.050000},{0.162500,0.837500,0.100000,0.100000},{0.187500,0.837500,0.050000,0.050000},{0.187500,0.837500,0.100000,0.100000},{0.212500,0.837500,0.050000,0.050000},{0.212500,0.837500,0.100000,0.100000},{0.237500,0.837500,0.050000,0.050000},{0.237500,0.837500,0.100000,0.100000},{0.262500,0.837500,0.050000,0.050000},{0.262500,0.837500,0.100000,0.100000},{0.287500,0.837500,0.050000,0.050000},{0.287500,0.837500,0.100000,0.100000},{0.312500,0.837500,0.050000,0.050000},{0.312500,0.837500,0.100000,0.100000},{0.337500,0.837500,0.050000,0.050000},{0.337500,0.837500,0.100000,0.100000},{0.362500,0.837500,0.050000,0.050000},{0.362500,0.837500,0.100000,0.100000},{0.387500,0.837500,0.050000,0.050000},{0.387500,0.837500,0.100000,0.100000},{0.412500,0.837500,0.050000,0.050000},{0.412500,0.837500,0.100000,0.100000},{0.437500,0.837500,0.050000,0.050000},{0.437500,0.837500,0.100000,0.100000},{0.462500,0.837500,0.050000,0.050000},{0.462500,0.837500,0.100000,0.100000},{0.487500,0.837500,0.050000,0.050000},{0.487500,0.837500,0.100000,0.100000},{0.512500,0.837500,0.050000,0.050000},{0.512500,0.837500,0.100000,0.100000},{0.537500,0.837500,0.050000,0.050000},{0.537500,0.837500,0.100000,0.100000},{0.562500,0.837500,0.050000,0.050000},{0.562500,0.837500,0.100000,0.100000},{0.587500,0.837500,0.050000,0.050000},{0.587500,0.837500,0.100000,0.100000},{0.612500,0.837500,0.050000,0.050000},{0.612500,0.837500,0.100000,0.100000},{0.637500,0.837500,0.050000,0.050000},{0.637500,0.837500,0.100000,0.100000},{0.662500,0.837500,0.050000,0.050000},{0.662500,0.837500,0.100000,0.100000},{0.687500,0.837500,0.050000,0.050000},{0.687500,0.837500,0.100000,0.100000},{0.712500,0.837500,0.050000,0.050000},{0.712500,0.837500,0.100000,0.100000},{0.737500,0.837500,0.050000,0.050000},{0.737500,0.837500,0.100000,0.100000},{0.762500,0.837500,0.050000,0.050000},{0.762500,0.837500,0.100000,0.100000},{0.787500,0.837500,0.050000,0.050000},{0.787500,0.837500,0.100000,0.100000},{0.812500,0.837500,0.050000,0.050000},{0.812500,0.837500,0.100000,0.100000},{0.837500,0.837500,0.050000,0.050000},{0.837500,0.837500,0.100000,0.100000},{0.862500,0.837500,0.050000,0.050000},{0.862500,0.837500,0.100000,0.100000},{0.887500,0.837500,0.050000,0.050000},{0.887500,0.837500,0.100000,0.100000},{0.912500,0.837500,0.050000,0.050000},{0.912500,0.837500,0.100000,0.100000},{0.937500,0.837500,0.050000,0.050000},{0.937500,0.837500,0.100000,0.100000},{0.962500,0.837500,0.050000,0.050000},{0.962500,0.837500,0.100000,0.100000},{0.987500,0.837500,0.050000,0.050000},{0.987500,0.837500,0.100000,0.100000},{0.012500,0.862500,0.050000,0.050000},{0.012500,0.862500,0.100000,0.100000},{0.037500,0.862500,0.050000,0.050000},{0.037500,0.862500,0.100000,0.100000},{0.062500,0.862500,0.050000,0.050000},{0.062500,0.862500,0.100000,0.100000},{0.087500,0.862500,0.050000,0.050000},{0.087500,0.862500,0.100000,0.100000},{0.112500,0.862500,0.050000,0.050000},{0.112500,0.862500,0.100000,0.100000},{0.137500,0.862500,0.050000,0.050000},{0.137500,0.862500,0.100000,0.100000},{0.162500,0.862500,0.050000,0.050000},{0.162500,0.862500,0.100000,0.100000},{0.187500,0.862500,0.050000,0.050000},{0.187500,0.862500,0.100000,0.100000},{0.212500,0.862500,0.050000,0.050000},{0.212500,0.862500,0.100000,0.100000},{0.237500,0.862500,0.050000,0.050000},{0.237500,0.862500,0.100000,0.100000},{0.262500,0.862500,0.050000,0.050000},{0.262500,0.862500,0.100000,0.100000},{0.287500,0.862500,0.050000,0.050000},{0.287500,0.862500,0.100000,0.100000},{0.312500,0.862500,0.050000,0.050000},{0.312500,0.862500,0.100000,0.100000},{0.337500,0.862500,0.050000,0.050000},{0.337500,0.862500,0.100000,0.100000},{0.362500,0.862500,0.050000,0.050000},{0.362500,0.862500,0.100000,0.100000},{0.387500,0.862500,0.050000,0.050000},{0.387500,0.862500,0.100000,0.100000},{0.412500,0.862500,0.050000,0.050000},{0.412500,0.862500,0.100000,0.100000},{0.437500,0.862500,0.050000,0.050000},{0.437500,0.862500,0.100000,0.100000},{0.462500,0.862500,0.050000,0.050000},{0.462500,0.862500,0.100000,0.100000},{0.487500,0.862500,0.050000,0.050000},{0.487500,0.862500,0.100000,0.100000},{0.512500,0.862500,0.050000,0.050000},{0.512500,0.862500,0.100000,0.100000},{0.537500,0.862500,0.050000,0.050000},{0.537500,0.862500,0.100000,0.100000},{0.562500,0.862500,0.050000,0.050000},{0.562500,0.862500,0.100000,0.100000},{0.587500,0.862500,0.050000,0.050000},{0.587500,0.862500,0.100000,0.100000},{0.612500,0.862500,0.050000,0.050000},{0.612500,0.862500,0.100000,0.100000},{0.637500,0.862500,0.050000,0.050000},{0.637500,0.862500,0.100000,0.100000},{0.662500,0.862500,0.050000,0.050000},{0.662500,0.862500,0.100000,0.100000},{0.687500,0.862500,0.050000,0.050000},{0.687500,0.862500,0.100000,0.100000},{0.712500,0.862500,0.050000,0.050000},{0.712500,0.862500,0.100000,0.100000},{0.737500,0.862500,0.050000,0.050000},{0.737500,0.862500,0.100000,0.100000},{0.762500,0.862500,0.050000,0.050000},{0.762500,0.862500,0.100000,0.100000},{0.787500,0.862500,0.050000,0.050000},{0.787500,0.862500,0.100000,0.100000},{0.812500,0.862500,0.050000,0.050000},{0.812500,0.862500,0.100000,0.100000},{0.837500,0.862500,0.050000,0.050000},{0.837500,0.862500,0.100000,0.100000},{0.862500,0.862500,0.050000,0.050000},{0.862500,0.862500,0.100000,0.100000},{0.887500,0.862500,0.050000,0.050000},{0.887500,0.862500,0.100000,0.100000},{0.912500,0.862500,0.050000,0.050000},{0.912500,0.862500,0.100000,0.100000},{0.937500,0.862500,0.050000,0.050000},{0.937500,0.862500,0.100000,0.100000},{0.962500,0.862500,0.050000,0.050000},{0.962500,0.862500,0.100000,0.100000},{0.987500,0.862500,0.050000,0.050000},{0.987500,0.862500,0.100000,0.100000},{0.012500,0.887500,0.050000,0.050000},{0.012500,0.887500,0.100000,0.100000},{0.037500,0.887500,0.050000,0.050000},{0.037500,0.887500,0.100000,0.100000},{0.062500,0.887500,0.050000,0.050000},{0.062500,0.887500,0.100000,0.100000},{0.087500,0.887500,0.050000,0.050000},{0.087500,0.887500,0.100000,0.100000},{0.112500,0.887500,0.050000,0.050000},{0.112500,0.887500,0.100000,0.100000},{0.137500,0.887500,0.050000,0.050000},{0.137500,0.887500,0.100000,0.100000},{0.162500,0.887500,0.050000,0.050000},{0.162500,0.887500,0.100000,0.100000},{0.187500,0.887500,0.050000,0.050000},{0.187500,0.887500,0.100000,0.100000},{0.212500,0.887500,0.050000,0.050000},{0.212500,0.887500,0.100000,0.100000},{0.237500,0.887500,0.050000,0.050000},{0.237500,0.887500,0.100000,0.100000},{0.262500,0.887500,0.050000,0.050000},{0.262500,0.887500,0.100000,0.100000},{0.287500,0.887500,0.050000,0.050000},{0.287500,0.887500,0.100000,0.100000},{0.312500,0.887500,0.050000,0.050000},{0.312500,0.887500,0.100000,0.100000},{0.337500,0.887500,0.050000,0.050000},{0.337500,0.887500,0.100000,0.100000},{0.362500,0.887500,0.050000,0.050000},{0.362500,0.887500,0.100000,0.100000},{0.387500,0.887500,0.050000,0.050000},{0.387500,0.887500,0.100000,0.100000},{0.412500,0.887500,0.050000,0.050000},{0.412500,0.887500,0.100000,0.100000},{0.437500,0.887500,0.050000,0.050000},{0.437500,0.887500,0.100000,0.100000},{0.462500,0.887500,0.050000,0.050000},{0.462500,0.887500,0.100000,0.100000},{0.487500,0.887500,0.050000,0.050000},{0.487500,0.887500,0.100000,0.100000},{0.512500,0.887500,0.050000,0.050000},{0.512500,0.887500,0.100000,0.100000},{0.537500,0.887500,0.050000,0.050000},{0.537500,0.887500,0.100000,0.100000},{0.562500,0.887500,0.050000,0.050000},{0.562500,0.887500,0.100000,0.100000},{0.587500,0.887500,0.050000,0.050000},{0.587500,0.887500,0.100000,0.100000},{0.612500,0.887500,0.050000,0.050000},{0.612500,0.887500,0.100000,0.100000},{0.637500,0.887500,0.050000,0.050000},{0.637500,0.887500,0.100000,0.100000},{0.662500,0.887500,0.050000,0.050000},{0.662500,0.887500,0.100000,0.100000},{0.687500,0.887500,0.050000,0.050000},{0.687500,0.887500,0.100000,0.100000},{0.712500,0.887500,0.050000,0.050000},{0.712500,0.887500,0.100000,0.100000},{0.737500,0.887500,0.050000,0.050000},{0.737500,0.887500,0.100000,0.100000},{0.762500,0.887500,0.050000,0.050000},{0.762500,0.887500,0.100000,0.100000},{0.787500,0.887500,0.050000,0.050000},{0.787500,0.887500,0.100000,0.100000},{0.812500,0.887500,0.050000,0.050000},{0.812500,0.887500,0.100000,0.100000},{0.837500,0.887500,0.050000,0.050000},{0.837500,0.887500,0.100000,0.100000},{0.862500,0.887500,0.050000,0.050000},{0.862500,0.887500,0.100000,0.100000},{0.887500,0.887500,0.050000,0.050000},{0.887500,0.887500,0.100000,0.100000},{0.912500,0.887500,0.050000,0.050000},{0.912500,0.887500,0.100000,0.100000},{0.937500,0.887500,0.050000,0.050000},{0.937500,0.887500,0.100000,0.100000},{0.962500,0.887500,0.050000,0.050000},{0.962500,0.887500,0.100000,0.100000},{0.987500,0.887500,0.050000,0.050000},{0.987500,0.887500,0.100000,0.100000},{0.012500,0.912500,0.050000,0.050000},{0.012500,0.912500,0.100000,0.100000},{0.037500,0.912500,0.050000,0.050000},{0.037500,0.912500,0.100000,0.100000},{0.062500,0.912500,0.050000,0.050000},{0.062500,0.912500,0.100000,0.100000},{0.087500,0.912500,0.050000,0.050000},{0.087500,0.912500,0.100000,0.100000},{0.112500,0.912500,0.050000,0.050000},{0.112500,0.912500,0.100000,0.100000},{0.137500,0.912500,0.050000,0.050000},{0.137500,0.912500,0.100000,0.100000},{0.162500,0.912500,0.050000,0.050000},{0.162500,0.912500,0.100000,0.100000},{0.187500,0.912500,0.050000,0.050000},{0.187500,0.912500,0.100000,0.100000},{0.212500,0.912500,0.050000,0.050000},{0.212500,0.912500,0.100000,0.100000},{0.237500,0.912500,0.050000,0.050000},{0.237500,0.912500,0.100000,0.100000},{0.262500,0.912500,0.050000,0.050000},{0.262500,0.912500,0.100000,0.100000},{0.287500,0.912500,0.050000,0.050000},{0.287500,0.912500,0.100000,0.100000},{0.312500,0.912500,0.050000,0.050000},{0.312500,0.912500,0.100000,0.100000},{0.337500,0.912500,0.050000,0.050000},{0.337500,0.912500,0.100000,0.100000},{0.362500,0.912500,0.050000,0.050000},{0.362500,0.912500,0.100000,0.100000},{0.387500,0.912500,0.050000,0.050000},{0.387500,0.912500,0.100000,0.100000},{0.412500,0.912500,0.050000,0.050000},{0.412500,0.912500,0.100000,0.100000},{0.437500,0.912500,0.050000,0.050000},{0.437500,0.912500,0.100000,0.100000},{0.462500,0.912500,0.050000,0.050000},{0.462500,0.912500,0.100000,0.100000},{0.487500,0.912500,0.050000,0.050000},{0.487500,0.912500,0.100000,0.100000},{0.512500,0.912500,0.050000,0.050000},{0.512500,0.912500,0.100000,0.100000},{0.537500,0.912500,0.050000,0.050000},{0.537500,0.912500,0.100000,0.100000},{0.562500,0.912500,0.050000,0.050000},{0.562500,0.912500,0.100000,0.100000},{0.587500,0.912500,0.050000,0.050000},{0.587500,0.912500,0.100000,0.100000},{0.612500,0.912500,0.050000,0.050000},{0.612500,0.912500,0.100000,0.100000},{0.637500,0.912500,0.050000,0.050000},{0.637500,0.912500,0.100000,0.100000},{0.662500,0.912500,0.050000,0.050000},{0.662500,0.912500,0.100000,0.100000},{0.687500,0.912500,0.050000,0.050000},{0.687500,0.912500,0.100000,0.100000},{0.712500,0.912500,0.050000,0.050000},{0.712500,0.912500,0.100000,0.100000},{0.737500,0.912500,0.050000,0.050000},{0.737500,0.912500,0.100000,0.100000},{0.762500,0.912500,0.050000,0.050000},{0.762500,0.912500,0.100000,0.100000},{0.787500,0.912500,0.050000,0.050000},{0.787500,0.912500,0.100000,0.100000},{0.812500,0.912500,0.050000,0.050000},{0.812500,0.912500,0.100000,0.100000},{0.837500,0.912500,0.050000,0.050000},{0.837500,0.912500,0.100000,0.100000},{0.862500,0.912500,0.050000,0.050000},{0.862500,0.912500,0.100000,0.100000},{0.887500,0.912500,0.050000,0.050000},{0.887500,0.912500,0.100000,0.100000},{0.912500,0.912500,0.050000,0.050000},{0.912500,0.912500,0.100000,0.100000},{0.937500,0.912500,0.050000,0.050000},{0.937500,0.912500,0.100000,0.100000},{0.962500,0.912500,0.050000,0.050000},{0.962500,0.912500,0.100000,0.100000},{0.987500,0.912500,0.050000,0.050000},{0.987500,0.912500,0.100000,0.100000},{0.012500,0.937500,0.050000,0.050000},{0.012500,0.937500,0.100000,0.100000},{0.037500,0.937500,0.050000,0.050000},{0.037500,0.937500,0.100000,0.100000},{0.062500,0.937500,0.050000,0.050000},{0.062500,0.937500,0.100000,0.100000},{0.087500,0.937500,0.050000,0.050000},{0.087500,0.937500,0.100000,0.100000},{0.112500,0.937500,0.050000,0.050000},{0.112500,0.937500,0.100000,0.100000},{0.137500,0.937500,0.050000,0.050000},{0.137500,0.937500,0.100000,0.100000},{0.162500,0.937500,0.050000,0.050000},{0.162500,0.937500,0.100000,0.100000},{0.187500,0.937500,0.050000,0.050000},{0.187500,0.937500,0.100000,0.100000},{0.212500,0.937500,0.050000,0.050000},{0.212500,0.937500,0.100000,0.100000},{0.237500,0.937500,0.050000,0.050000},{0.237500,0.937500,0.100000,0.100000},{0.262500,0.937500,0.050000,0.050000},{0.262500,0.937500,0.100000,0.100000},{0.287500,0.937500,0.050000,0.050000},{0.287500,0.937500,0.100000,0.100000},{0.312500,0.937500,0.050000,0.050000},{0.312500,0.937500,0.100000,0.100000},{0.337500,0.937500,0.050000,0.050000},{0.337500,0.937500,0.100000,0.100000},{0.362500,0.937500,0.050000,0.050000},{0.362500,0.937500,0.100000,0.100000},{0.387500,0.937500,0.050000,0.050000},{0.387500,0.937500,0.100000,0.100000},{0.412500,0.937500,0.050000,0.050000},{0.412500,0.937500,0.100000,0.100000},{0.437500,0.937500,0.050000,0.050000},{0.437500,0.937500,0.100000,0.100000},{0.462500,0.937500,0.050000,0.050000},{0.462500,0.937500,0.100000,0.100000},{0.487500,0.937500,0.050000,0.050000},{0.487500,0.937500,0.100000,0.100000},{0.512500,0.937500,0.050000,0.050000},{0.512500,0.937500,0.100000,0.100000},{0.537500,0.937500,0.050000,0.050000},{0.537500,0.937500,0.100000,0.100000},{0.562500,0.937500,0.050000,0.050000},{0.562500,0.937500,0.100000,0.100000},{0.587500,0.937500,0.050000,0.050000},{0.587500,0.937500,0.100000,0.100000},{0.612500,0.937500,0.050000,0.050000},{0.612500,0.937500,0.100000,0.100000},{0.637500,0.937500,0.050000,0.050000},{0.637500,0.937500,0.100000,0.100000},{0.662500,0.937500,0.050000,0.050000},{0.662500,0.937500,0.100000,0.100000},{0.687500,0.937500,0.050000,0.050000},{0.687500,0.937500,0.100000,0.100000},{0.712500,0.937500,0.050000,0.050000},{0.712500,0.937500,0.100000,0.100000},{0.737500,0.937500,0.050000,0.050000},{0.737500,0.937500,0.100000,0.100000},{0.762500,0.937500,0.050000,0.050000},{0.762500,0.937500,0.100000,0.100000},{0.787500,0.937500,0.050000,0.050000},{0.787500,0.937500,0.100000,0.100000},{0.812500,0.937500,0.050000,0.050000},{0.812500,0.937500,0.100000,0.100000},{0.837500,0.937500,0.050000,0.050000},{0.837500,0.937500,0.100000,0.100000},{0.862500,0.937500,0.050000,0.050000},{0.862500,0.937500,0.100000,0.100000},{0.887500,0.937500,0.050000,0.050000},{0.887500,0.937500,0.100000,0.100000},{0.912500,0.937500,0.050000,0.050000},{0.912500,0.937500,0.100000,0.100000},{0.937500,0.937500,0.050000,0.050000},{0.937500,0.937500,0.100000,0.100000},{0.962500,0.937500,0.050000,0.050000},{0.962500,0.937500,0.100000,0.100000},{0.987500,0.937500,0.050000,0.050000},{0.987500,0.937500,0.100000,0.100000},{0.012500,0.962500,0.050000,0.050000},{0.012500,0.962500,0.100000,0.100000},{0.037500,0.962500,0.050000,0.050000},{0.037500,0.962500,0.100000,0.100000},{0.062500,0.962500,0.050000,0.050000},{0.062500,0.962500,0.100000,0.100000},{0.087500,0.962500,0.050000,0.050000},{0.087500,0.962500,0.100000,0.100000},{0.112500,0.962500,0.050000,0.050000},{0.112500,0.962500,0.100000,0.100000},{0.137500,0.962500,0.050000,0.050000},{0.137500,0.962500,0.100000,0.100000},{0.162500,0.962500,0.050000,0.050000},{0.162500,0.962500,0.100000,0.100000},{0.187500,0.962500,0.050000,0.050000},{0.187500,0.962500,0.100000,0.100000},{0.212500,0.962500,0.050000,0.050000},{0.212500,0.962500,0.100000,0.100000},{0.237500,0.962500,0.050000,0.050000},{0.237500,0.962500,0.100000,0.100000},{0.262500,0.962500,0.050000,0.050000},{0.262500,0.962500,0.100000,0.100000},{0.287500,0.962500,0.050000,0.050000},{0.287500,0.962500,0.100000,0.100000},{0.312500,0.962500,0.050000,0.050000},{0.312500,0.962500,0.100000,0.100000},{0.337500,0.962500,0.050000,0.050000},{0.337500,0.962500,0.100000,0.100000},{0.362500,0.962500,0.050000,0.050000},{0.362500,0.962500,0.100000,0.100000},{0.387500,0.962500,0.050000,0.050000},{0.387500,0.962500,0.100000,0.100000},{0.412500,0.962500,0.050000,0.050000},{0.412500,0.962500,0.100000,0.100000},{0.437500,0.962500,0.050000,0.050000},{0.437500,0.962500,0.100000,0.100000},{0.462500,0.962500,0.050000,0.050000},{0.462500,0.962500,0.100000,0.100000},{0.487500,0.962500,0.050000,0.050000},{0.487500,0.962500,0.100000,0.100000},{0.512500,0.962500,0.050000,0.050000},{0.512500,0.962500,0.100000,0.100000},{0.537500,0.962500,0.050000,0.050000},{0.537500,0.962500,0.100000,0.100000},{0.562500,0.962500,0.050000,0.050000},{0.562500,0.962500,0.100000,0.100000},{0.587500,0.962500,0.050000,0.050000},{0.587500,0.962500,0.100000,0.100000},{0.612500,0.962500,0.050000,0.050000},{0.612500,0.962500,0.100000,0.100000},{0.637500,0.962500,0.050000,0.050000},{0.637500,0.962500,0.100000,0.100000},{0.662500,0.962500,0.050000,0.050000},{0.662500,0.962500,0.100000,0.100000},{0.687500,0.962500,0.050000,0.050000},{0.687500,0.962500,0.100000,0.100000},{0.712500,0.962500,0.050000,0.050000},{0.712500,0.962500,0.100000,0.100000},{0.737500,0.962500,0.050000,0.050000},{0.737500,0.962500,0.100000,0.100000},{0.762500,0.962500,0.050000,0.050000},{0.762500,0.962500,0.100000,0.100000},{0.787500,0.962500,0.050000,0.050000},{0.787500,0.962500,0.100000,0.100000},{0.812500,0.962500,0.050000,0.050000},{0.812500,0.962500,0.100000,0.100000},{0.837500,0.962500,0.050000,0.050000},{0.837500,0.962500,0.100000,0.100000},{0.862500,0.962500,0.050000,0.050000},{0.862500,0.962500,0.100000,0.100000},{0.887500,0.962500,0.050000,0.050000},{0.887500,0.962500,0.100000,0.100000},{0.912500,0.962500,0.050000,0.050000},{0.912500,0.962500,0.100000,0.100000},{0.937500,0.962500,0.050000,0.050000},{0.937500,0.962500,0.100000,0.100000},{0.962500,0.962500,0.050000,0.050000},{0.962500,0.962500,0.100000,0.100000},{0.987500,0.962500,0.050000,0.050000},{0.987500,0.962500,0.100000,0.100000},{0.012500,0.987500,0.050000,0.050000},{0.012500,0.987500,0.100000,0.100000},{0.037500,0.987500,0.050000,0.050000},{0.037500,0.987500,0.100000,0.100000},{0.062500,0.987500,0.050000,0.050000},{0.062500,0.987500,0.100000,0.100000},{0.087500,0.987500,0.050000,0.050000},{0.087500,0.987500,0.100000,0.100000},{0.112500,0.987500,0.050000,0.050000},{0.112500,0.987500,0.100000,0.100000},{0.137500,0.987500,0.050000,0.050000},{0.137500,0.987500,0.100000,0.100000},{0.162500,0.987500,0.050000,0.050000},{0.162500,0.987500,0.100000,0.100000},{0.187500,0.987500,0.050000,0.050000},{0.187500,0.987500,0.100000,0.100000},{0.212500,0.987500,0.050000,0.050000},{0.212500,0.987500,0.100000,0.100000},{0.237500,0.987500,0.050000,0.050000},{0.237500,0.987500,0.100000,0.100000},{0.262500,0.987500,0.050000,0.050000},{0.262500,0.987500,0.100000,0.100000},{0.287500,0.987500,0.050000,0.050000},{0.287500,0.987500,0.100000,0.100000},{0.312500,0.987500,0.050000,0.050000},{0.312500,0.987500,0.100000,0.100000},{0.337500,0.987500,0.050000,0.050000},{0.337500,0.987500,0.100000,0.100000},{0.362500,0.987500,0.050000,0.050000},{0.362500,0.987500,0.100000,0.100000},{0.387500,0.987500,0.050000,0.050000},{0.387500,0.987500,0.100000,0.100000},{0.412500,0.987500,0.050000,0.050000},{0.412500,0.987500,0.100000,0.100000},{0.437500,0.987500,0.050000,0.050000},{0.437500,0.987500,0.100000,0.100000},{0.462500,0.987500,0.050000,0.050000},{0.462500,0.987500,0.100000,0.100000},{0.487500,0.987500,0.050000,0.050000},{0.487500,0.987500,0.100000,0.100000},{0.512500,0.987500,0.050000,0.050000},{0.512500,0.987500,0.100000,0.100000},{0.537500,0.987500,0.050000,0.050000},{0.537500,0.987500,0.100000,0.100000},{0.562500,0.987500,0.050000,0.050000},{0.562500,0.987500,0.100000,0.100000},{0.587500,0.987500,0.050000,0.050000},{0.587500,0.987500,0.100000,0.100000},{0.612500,0.987500,0.050000,0.050000},{0.612500,0.987500,0.100000,0.100000},{0.637500,0.987500,0.050000,0.050000},{0.637500,0.987500,0.100000,0.100000},{0.662500,0.987500,0.050000,0.050000},{0.662500,0.987500,0.100000,0.100000},{0.687500,0.987500,0.050000,0.050000},{0.687500,0.987500,0.100000,0.100000},{0.712500,0.987500,0.050000,0.050000},{0.712500,0.987500,0.100000,0.100000},{0.737500,0.987500,0.050000,0.050000},{0.737500,0.987500,0.100000,0.100000},{0.762500,0.987500,0.050000,0.050000},{0.762500,0.987500,0.100000,0.100000},{0.787500,0.987500,0.050000,0.050000},{0.787500,0.987500,0.100000,0.100000},{0.812500,0.987500,0.050000,0.050000},{0.812500,0.987500,0.100000,0.100000},{0.837500,0.987500,0.050000,0.050000},{0.837500,0.987500,0.100000,0.100000},{0.862500,0.987500,0.050000,0.050000},{0.862500,0.987500,0.100000,0.100000},{0.887500,0.987500,0.050000,0.050000},{0.887500,0.987500,0.100000,0.100000},{0.912500,0.987500,0.050000,0.050000},{0.912500,0.987500,0.100000,0.100000},{0.937500,0.987500,0.050000,0.050000},{0.937500,0.987500,0.100000,0.100000},{0.962500,0.987500,0.050000,0.050000},{0.962500,0.987500,0.100000,0.100000},{0.987500,0.987500,0.050000,0.050000},{0.987500,0.987500,0.100000,0.100000},{0.025000,0.025000,0.200000,0.200000},{0.025000,0.025000,0.400000,0.400000},{0.075000,0.025000,0.200000,0.200000},{0.075000,0.025000,0.400000,0.400000},{0.125000,0.025000,0.200000,0.200000},{0.125000,0.025000,0.400000,0.400000},{0.175000,0.025000,0.200000,0.200000},{0.175000,0.025000,0.400000,0.400000},{0.225000,0.025000,0.200000,0.200000},{0.225000,0.025000,0.400000,0.400000},{0.275000,0.025000,0.200000,0.200000},{0.275000,0.025000,0.400000,0.400000},{0.325000,0.025000,0.200000,0.200000},{0.325000,0.025000,0.400000,0.400000},{0.375000,0.025000,0.200000,0.200000},{0.375000,0.025000,0.400000,0.400000},{0.425000,0.025000,0.200000,0.200000},{0.425000,0.025000,0.400000,0.400000},{0.475000,0.025000,0.200000,0.200000},{0.475000,0.025000,0.400000,0.400000},{0.525000,0.025000,0.200000,0.200000},{0.525000,0.025000,0.400000,0.400000},{0.575000,0.025000,0.200000,0.200000},{0.575000,0.025000,0.400000,0.400000},{0.625000,0.025000,0.200000,0.200000},{0.625000,0.025000,0.400000,0.400000},{0.675000,0.025000,0.200000,0.200000},{0.675000,0.025000,0.400000,0.400000},{0.725000,0.025000,0.200000,0.200000},{0.725000,0.025000,0.400000,0.400000},{0.775000,0.025000,0.200000,0.200000},{0.775000,0.025000,0.400000,0.400000},{0.825000,0.025000,0.200000,0.200000},{0.825000,0.025000,0.400000,0.400000},{0.875000,0.025000,0.200000,0.200000},{0.875000,0.025000,0.400000,0.400000},{0.925000,0.025000,0.200000,0.200000},{0.925000,0.025000,0.400000,0.400000},{0.975000,0.025000,0.200000,0.200000},{0.975000,0.025000,0.400000,0.400000},{0.025000,0.075000,0.200000,0.200000},{0.025000,0.075000,0.400000,0.400000},{0.075000,0.075000,0.200000,0.200000},{0.075000,0.075000,0.400000,0.400000},{0.125000,0.075000,0.200000,0.200000},{0.125000,0.075000,0.400000,0.400000},{0.175000,0.075000,0.200000,0.200000},{0.175000,0.075000,0.400000,0.400000},{0.225000,0.075000,0.200000,0.200000},{0.225000,0.075000,0.400000,0.400000},{0.275000,0.075000,0.200000,0.200000},{0.275000,0.075000,0.400000,0.400000},{0.325000,0.075000,0.200000,0.200000},{0.325000,0.075000,0.400000,0.400000},{0.375000,0.075000,0.200000,0.200000},{0.375000,0.075000,0.400000,0.400000},{0.425000,0.075000,0.200000,0.200000},{0.425000,0.075000,0.400000,0.400000},{0.475000,0.075000,0.200000,0.200000},{0.475000,0.075000,0.400000,0.400000},{0.525000,0.075000,0.200000,0.200000},{0.525000,0.075000,0.400000,0.400000},{0.575000,0.075000,0.200000,0.200000},{0.575000,0.075000,0.400000,0.400000},{0.625000,0.075000,0.200000,0.200000},{0.625000,0.075000,0.400000,0.400000},{0.675000,0.075000,0.200000,0.200000},{0.675000,0.075000,0.400000,0.400000},{0.725000,0.075000,0.200000,0.200000},{0.725000,0.075000,0.400000,0.400000},{0.775000,0.075000,0.200000,0.200000},{0.775000,0.075000,0.400000,0.400000},{0.825000,0.075000,0.200000,0.200000},{0.825000,0.075000,0.400000,0.400000},{0.875000,0.075000,0.200000,0.200000},{0.875000,0.075000,0.400000,0.400000},{0.925000,0.075000,0.200000,0.200000},{0.925000,0.075000,0.400000,0.400000},{0.975000,0.075000,0.200000,0.200000},{0.975000,0.075000,0.400000,0.400000},{0.025000,0.125000,0.200000,0.200000},{0.025000,0.125000,0.400000,0.400000},{0.075000,0.125000,0.200000,0.200000},{0.075000,0.125000,0.400000,0.400000},{0.125000,0.125000,0.200000,0.200000},{0.125000,0.125000,0.400000,0.400000},{0.175000,0.125000,0.200000,0.200000},{0.175000,0.125000,0.400000,0.400000},{0.225000,0.125000,0.200000,0.200000},{0.225000,0.125000,0.400000,0.400000},{0.275000,0.125000,0.200000,0.200000},{0.275000,0.125000,0.400000,0.400000},{0.325000,0.125000,0.200000,0.200000},{0.325000,0.125000,0.400000,0.400000},{0.375000,0.125000,0.200000,0.200000},{0.375000,0.125000,0.400000,0.400000},{0.425000,0.125000,0.200000,0.200000},{0.425000,0.125000,0.400000,0.400000},{0.475000,0.125000,0.200000,0.200000},{0.475000,0.125000,0.400000,0.400000},{0.525000,0.125000,0.200000,0.200000},{0.525000,0.125000,0.400000,0.400000},{0.575000,0.125000,0.200000,0.200000},{0.575000,0.125000,0.400000,0.400000},{0.625000,0.125000,0.200000,0.200000},{0.625000,0.125000,0.400000,0.400000},{0.675000,0.125000,0.200000,0.200000},{0.675000,0.125000,0.400000,0.400000},{0.725000,0.125000,0.200000,0.200000},{0.725000,0.125000,0.400000,0.400000},{0.775000,0.125000,0.200000,0.200000},{0.775000,0.125000,0.400000,0.400000},{0.825000,0.125000,0.200000,0.200000},{0.825000,0.125000,0.400000,0.400000},{0.875000,0.125000,0.200000,0.200000},{0.875000,0.125000,0.400000,0.400000},{0.925000,0.125000,0.200000,0.200000},{0.925000,0.125000,0.400000,0.400000},{0.975000,0.125000,0.200000,0.200000},{0.975000,0.125000,0.400000,0.400000},{0.025000,0.175000,0.200000,0.200000},{0.025000,0.175000,0.400000,0.400000},{0.075000,0.175000,0.200000,0.200000},{0.075000,0.175000,0.400000,0.400000},{0.125000,0.175000,0.200000,0.200000},{0.125000,0.175000,0.400000,0.400000},{0.175000,0.175000,0.200000,0.200000},{0.175000,0.175000,0.400000,0.400000},{0.225000,0.175000,0.200000,0.200000},{0.225000,0.175000,0.400000,0.400000},{0.275000,0.175000,0.200000,0.200000},{0.275000,0.175000,0.400000,0.400000},{0.325000,0.175000,0.200000,0.200000},{0.325000,0.175000,0.400000,0.400000},{0.375000,0.175000,0.200000,0.200000},{0.375000,0.175000,0.400000,0.400000},{0.425000,0.175000,0.200000,0.200000},{0.425000,0.175000,0.400000,0.400000},{0.475000,0.175000,0.200000,0.200000},{0.475000,0.175000,0.400000,0.400000},{0.525000,0.175000,0.200000,0.200000},{0.525000,0.175000,0.400000,0.400000},{0.575000,0.175000,0.200000,0.200000},{0.575000,0.175000,0.400000,0.400000},{0.625000,0.175000,0.200000,0.200000},{0.625000,0.175000,0.400000,0.400000},{0.675000,0.175000,0.200000,0.200000},{0.675000,0.175000,0.400000,0.400000},{0.725000,0.175000,0.200000,0.200000},{0.725000,0.175000,0.400000,0.400000},{0.775000,0.175000,0.200000,0.200000},{0.775000,0.175000,0.400000,0.400000},{0.825000,0.175000,0.200000,0.200000},{0.825000,0.175000,0.400000,0.400000},{0.875000,0.175000,0.200000,0.200000},{0.875000,0.175000,0.400000,0.400000},{0.925000,0.175000,0.200000,0.200000},{0.925000,0.175000,0.400000,0.400000},{0.975000,0.175000,0.200000,0.200000},{0.975000,0.175000,0.400000,0.400000},{0.025000,0.225000,0.200000,0.200000},{0.025000,0.225000,0.400000,0.400000},{0.075000,0.225000,0.200000,0.200000},{0.075000,0.225000,0.400000,0.400000},{0.125000,0.225000,0.200000,0.200000},{0.125000,0.225000,0.400000,0.400000},{0.175000,0.225000,0.200000,0.200000},{0.175000,0.225000,0.400000,0.400000},{0.225000,0.225000,0.200000,0.200000},{0.225000,0.225000,0.400000,0.400000},{0.275000,0.225000,0.200000,0.200000},{0.275000,0.225000,0.400000,0.400000},{0.325000,0.225000,0.200000,0.200000},{0.325000,0.225000,0.400000,0.400000},{0.375000,0.225000,0.200000,0.200000},{0.375000,0.225000,0.400000,0.400000},{0.425000,0.225000,0.200000,0.200000},{0.425000,0.225000,0.400000,0.400000},{0.475000,0.225000,0.200000,0.200000},{0.475000,0.225000,0.400000,0.400000},{0.525000,0.225000,0.200000,0.200000},{0.525000,0.225000,0.400000,0.400000},{0.575000,0.225000,0.200000,0.200000},{0.575000,0.225000,0.400000,0.400000},{0.625000,0.225000,0.200000,0.200000},{0.625000,0.225000,0.400000,0.400000},{0.675000,0.225000,0.200000,0.200000},{0.675000,0.225000,0.400000,0.400000},{0.725000,0.225000,0.200000,0.200000},{0.725000,0.225000,0.400000,0.400000},{0.775000,0.225000,0.200000,0.200000},{0.775000,0.225000,0.400000,0.400000},{0.825000,0.225000,0.200000,0.200000},{0.825000,0.225000,0.400000,0.400000},{0.875000,0.225000,0.200000,0.200000},{0.875000,0.225000,0.400000,0.400000},{0.925000,0.225000,0.200000,0.200000},{0.925000,0.225000,0.400000,0.400000},{0.975000,0.225000,0.200000,0.200000},{0.975000,0.225000,0.400000,0.400000},{0.025000,0.275000,0.200000,0.200000},{0.025000,0.275000,0.400000,0.400000},{0.075000,0.275000,0.200000,0.200000},{0.075000,0.275000,0.400000,0.400000},{0.125000,0.275000,0.200000,0.200000},{0.125000,0.275000,0.400000,0.400000},{0.175000,0.275000,0.200000,0.200000},{0.175000,0.275000,0.400000,0.400000},{0.225000,0.275000,0.200000,0.200000},{0.225000,0.275000,0.400000,0.400000},{0.275000,0.275000,0.200000,0.200000},{0.275000,0.275000,0.400000,0.400000},{0.325000,0.275000,0.200000,0.200000},{0.325000,0.275000,0.400000,0.400000},{0.375000,0.275000,0.200000,0.200000},{0.375000,0.275000,0.400000,0.400000},{0.425000,0.275000,0.200000,0.200000},{0.425000,0.275000,0.400000,0.400000},{0.475000,0.275000,0.200000,0.200000},{0.475000,0.275000,0.400000,0.400000},{0.525000,0.275000,0.200000,0.200000},{0.525000,0.275000,0.400000,0.400000},{0.575000,0.275000,0.200000,0.200000},{0.575000,0.275000,0.400000,0.400000},{0.625000,0.275000,0.200000,0.200000},{0.625000,0.275000,0.400000,0.400000},{0.675000,0.275000,0.200000,0.200000},{0.675000,0.275000,0.400000,0.400000},{0.725000,0.275000,0.200000,0.200000},{0.725000,0.275000,0.400000,0.400000},{0.775000,0.275000,0.200000,0.200000},{0.775000,0.275000,0.400000,0.400000},{0.825000,0.275000,0.200000,0.200000},{0.825000,0.275000,0.400000,0.400000},{0.875000,0.275000,0.200000,0.200000},{0.875000,0.275000,0.400000,0.400000},{0.925000,0.275000,0.200000,0.200000},{0.925000,0.275000,0.400000,0.400000},{0.975000,0.275000,0.200000,0.200000},{0.975000,0.275000,0.400000,0.400000},{0.025000,0.325000,0.200000,0.200000},{0.025000,0.325000,0.400000,0.400000},{0.075000,0.325000,0.200000,0.200000},{0.075000,0.325000,0.400000,0.400000},{0.125000,0.325000,0.200000,0.200000},{0.125000,0.325000,0.400000,0.400000},{0.175000,0.325000,0.200000,0.200000},{0.175000,0.325000,0.400000,0.400000},{0.225000,0.325000,0.200000,0.200000},{0.225000,0.325000,0.400000,0.400000},{0.275000,0.325000,0.200000,0.200000},{0.275000,0.325000,0.400000,0.400000},{0.325000,0.325000,0.200000,0.200000},{0.325000,0.325000,0.400000,0.400000},{0.375000,0.325000,0.200000,0.200000},{0.375000,0.325000,0.400000,0.400000},{0.425000,0.325000,0.200000,0.200000},{0.425000,0.325000,0.400000,0.400000},{0.475000,0.325000,0.200000,0.200000},{0.475000,0.325000,0.400000,0.400000},{0.525000,0.325000,0.200000,0.200000},{0.525000,0.325000,0.400000,0.400000},{0.575000,0.325000,0.200000,0.200000},{0.575000,0.325000,0.400000,0.400000},{0.625000,0.325000,0.200000,0.200000},{0.625000,0.325000,0.400000,0.400000},{0.675000,0.325000,0.200000,0.200000},{0.675000,0.325000,0.400000,0.400000},{0.725000,0.325000,0.200000,0.200000},{0.725000,0.325000,0.400000,0.400000},{0.775000,0.325000,0.200000,0.200000},{0.775000,0.325000,0.400000,0.400000},{0.825000,0.325000,0.200000,0.200000},{0.825000,0.325000,0.400000,0.400000},{0.875000,0.325000,0.200000,0.200000},{0.875000,0.325000,0.400000,0.400000},{0.925000,0.325000,0.200000,0.200000},{0.925000,0.325000,0.400000,0.400000},{0.975000,0.325000,0.200000,0.200000},{0.975000,0.325000,0.400000,0.400000},{0.025000,0.375000,0.200000,0.200000},{0.025000,0.375000,0.400000,0.400000},{0.075000,0.375000,0.200000,0.200000},{0.075000,0.375000,0.400000,0.400000},{0.125000,0.375000,0.200000,0.200000},{0.125000,0.375000,0.400000,0.400000},{0.175000,0.375000,0.200000,0.200000},{0.175000,0.375000,0.400000,0.400000},{0.225000,0.375000,0.200000,0.200000},{0.225000,0.375000,0.400000,0.400000},{0.275000,0.375000,0.200000,0.200000},{0.275000,0.375000,0.400000,0.400000},{0.325000,0.375000,0.200000,0.200000},{0.325000,0.375000,0.400000,0.400000},{0.375000,0.375000,0.200000,0.200000},{0.375000,0.375000,0.400000,0.400000},{0.425000,0.375000,0.200000,0.200000},{0.425000,0.375000,0.400000,0.400000},{0.475000,0.375000,0.200000,0.200000},{0.475000,0.375000,0.400000,0.400000},{0.525000,0.375000,0.200000,0.200000},{0.525000,0.375000,0.400000,0.400000},{0.575000,0.375000,0.200000,0.200000},{0.575000,0.375000,0.400000,0.400000},{0.625000,0.375000,0.200000,0.200000},{0.625000,0.375000,0.400000,0.400000},{0.675000,0.375000,0.200000,0.200000},{0.675000,0.375000,0.400000,0.400000},{0.725000,0.375000,0.200000,0.200000},{0.725000,0.375000,0.400000,0.400000},{0.775000,0.375000,0.200000,0.200000},{0.775000,0.375000,0.400000,0.400000},{0.825000,0.375000,0.200000,0.200000},{0.825000,0.375000,0.400000,0.400000},{0.875000,0.375000,0.200000,0.200000},{0.875000,0.375000,0.400000,0.400000},{0.925000,0.375000,0.200000,0.200000},{0.925000,0.375000,0.400000,0.400000},{0.975000,0.375000,0.200000,0.200000},{0.975000,0.375000,0.400000,0.400000},{0.025000,0.425000,0.200000,0.200000},{0.025000,0.425000,0.400000,0.400000},{0.075000,0.425000,0.200000,0.200000},{0.075000,0.425000,0.400000,0.400000},{0.125000,0.425000,0.200000,0.200000},{0.125000,0.425000,0.400000,0.400000},{0.175000,0.425000,0.200000,0.200000},{0.175000,0.425000,0.400000,0.400000},{0.225000,0.425000,0.200000,0.200000},{0.225000,0.425000,0.400000,0.400000},{0.275000,0.425000,0.200000,0.200000},{0.275000,0.425000,0.400000,0.400000},{0.325000,0.425000,0.200000,0.200000},{0.325000,0.425000,0.400000,0.400000},{0.375000,0.425000,0.200000,0.200000},{0.375000,0.425000,0.400000,0.400000},{0.425000,0.425000,0.200000,0.200000},{0.425000,0.425000,0.400000,0.400000},{0.475000,0.425000,0.200000,0.200000},{0.475000,0.425000,0.400000,0.400000},{0.525000,0.425000,0.200000,0.200000},{0.525000,0.425000,0.400000,0.400000},{0.575000,0.425000,0.200000,0.200000},{0.575000,0.425000,0.400000,0.400000},{0.625000,0.425000,0.200000,0.200000},{0.625000,0.425000,0.400000,0.400000},{0.675000,0.425000,0.200000,0.200000},{0.675000,0.425000,0.400000,0.400000},{0.725000,0.425000,0.200000,0.200000},{0.725000,0.425000,0.400000,0.400000},{0.775000,0.425000,0.200000,0.200000},{0.775000,0.425000,0.400000,0.400000},{0.825000,0.425000,0.200000,0.200000},{0.825000,0.425000,0.400000,0.400000},{0.875000,0.425000,0.200000,0.200000},{0.875000,0.425000,0.400000,0.400000},{0.925000,0.425000,0.200000,0.200000},{0.925000,0.425000,0.400000,0.400000},{0.975000,0.425000,0.200000,0.200000},{0.975000,0.425000,0.400000,0.400000},{0.025000,0.475000,0.200000,0.200000},{0.025000,0.475000,0.400000,0.400000},{0.075000,0.475000,0.200000,0.200000},{0.075000,0.475000,0.400000,0.400000},{0.125000,0.475000,0.200000,0.200000},{0.125000,0.475000,0.400000,0.400000},{0.175000,0.475000,0.200000,0.200000},{0.175000,0.475000,0.400000,0.400000},{0.225000,0.475000,0.200000,0.200000},{0.225000,0.475000,0.400000,0.400000},{0.275000,0.475000,0.200000,0.200000},{0.275000,0.475000,0.400000,0.400000},{0.325000,0.475000,0.200000,0.200000},{0.325000,0.475000,0.400000,0.400000},{0.375000,0.475000,0.200000,0.200000},{0.375000,0.475000,0.400000,0.400000},{0.425000,0.475000,0.200000,0.200000},{0.425000,0.475000,0.400000,0.400000},{0.475000,0.475000,0.200000,0.200000},{0.475000,0.475000,0.400000,0.400000},{0.525000,0.475000,0.200000,0.200000},{0.525000,0.475000,0.400000,0.400000},{0.575000,0.475000,0.200000,0.200000},{0.575000,0.475000,0.400000,0.400000},{0.625000,0.475000,0.200000,0.200000},{0.625000,0.475000,0.400000,0.400000},{0.675000,0.475000,0.200000,0.200000},{0.675000,0.475000,0.400000,0.400000},{0.725000,0.475000,0.200000,0.200000},{0.725000,0.475000,0.400000,0.400000},{0.775000,0.475000,0.200000,0.200000},{0.775000,0.475000,0.400000,0.400000},{0.825000,0.475000,0.200000,0.200000},{0.825000,0.475000,0.400000,0.400000},{0.875000,0.475000,0.200000,0.200000},{0.875000,0.475000,0.400000,0.400000},{0.925000,0.475000,0.200000,0.200000},{0.925000,0.475000,0.400000,0.400000},{0.975000,0.475000,0.200000,0.200000},{0.975000,0.475000,0.400000,0.400000},{0.025000,0.525000,0.200000,0.200000},{0.025000,0.525000,0.400000,0.400000},{0.075000,0.525000,0.200000,0.200000},{0.075000,0.525000,0.400000,0.400000},{0.125000,0.525000,0.200000,0.200000},{0.125000,0.525000,0.400000,0.400000},{0.175000,0.525000,0.200000,0.200000},{0.175000,0.525000,0.400000,0.400000},{0.225000,0.525000,0.200000,0.200000},{0.225000,0.525000,0.400000,0.400000},{0.275000,0.525000,0.200000,0.200000},{0.275000,0.525000,0.400000,0.400000},{0.325000,0.525000,0.200000,0.200000},{0.325000,0.525000,0.400000,0.400000},{0.375000,0.525000,0.200000,0.200000},{0.375000,0.525000,0.400000,0.400000},{0.425000,0.525000,0.200000,0.200000},{0.425000,0.525000,0.400000,0.400000},{0.475000,0.525000,0.200000,0.200000},{0.475000,0.525000,0.400000,0.400000},{0.525000,0.525000,0.200000,0.200000},{0.525000,0.525000,0.400000,0.400000},{0.575000,0.525000,0.200000,0.200000},{0.575000,0.525000,0.400000,0.400000},{0.625000,0.525000,0.200000,0.200000},{0.625000,0.525000,0.400000,0.400000},{0.675000,0.525000,0.200000,0.200000},{0.675000,0.525000,0.400000,0.400000},{0.725000,0.525000,0.200000,0.200000},{0.725000,0.525000,0.400000,0.400000},{0.775000,0.525000,0.200000,0.200000},{0.775000,0.525000,0.400000,0.400000},{0.825000,0.525000,0.200000,0.200000},{0.825000,0.525000,0.400000,0.400000},{0.875000,0.525000,0.200000,0.200000},{0.875000,0.525000,0.400000,0.400000},{0.925000,0.525000,0.200000,0.200000},{0.925000,0.525000,0.400000,0.400000},{0.975000,0.525000,0.200000,0.200000},{0.975000,0.525000,0.400000,0.400000},{0.025000,0.575000,0.200000,0.200000},{0.025000,0.575000,0.400000,0.400000},{0.075000,0.575000,0.200000,0.200000},{0.075000,0.575000,0.400000,0.400000},{0.125000,0.575000,0.200000,0.200000},{0.125000,0.575000,0.400000,0.400000},{0.175000,0.575000,0.200000,0.200000},{0.175000,0.575000,0.400000,0.400000},{0.225000,0.575000,0.200000,0.200000},{0.225000,0.575000,0.400000,0.400000},{0.275000,0.575000,0.200000,0.200000},{0.275000,0.575000,0.400000,0.400000},{0.325000,0.575000,0.200000,0.200000},{0.325000,0.575000,0.400000,0.400000},{0.375000,0.575000,0.200000,0.200000},{0.375000,0.575000,0.400000,0.400000},{0.425000,0.575000,0.200000,0.200000},{0.425000,0.575000,0.400000,0.400000},{0.475000,0.575000,0.200000,0.200000},{0.475000,0.575000,0.400000,0.400000},{0.525000,0.575000,0.200000,0.200000},{0.525000,0.575000,0.400000,0.400000},{0.575000,0.575000,0.200000,0.200000},{0.575000,0.575000,0.400000,0.400000},{0.625000,0.575000,0.200000,0.200000},{0.625000,0.575000,0.400000,0.400000},{0.675000,0.575000,0.200000,0.200000},{0.675000,0.575000,0.400000,0.400000},{0.725000,0.575000,0.200000,0.200000},{0.725000,0.575000,0.400000,0.400000},{0.775000,0.575000,0.200000,0.200000},{0.775000,0.575000,0.400000,0.400000},{0.825000,0.575000,0.200000,0.200000},{0.825000,0.575000,0.400000,0.400000},{0.875000,0.575000,0.200000,0.200000},{0.875000,0.575000,0.400000,0.400000},{0.925000,0.575000,0.200000,0.200000},{0.925000,0.575000,0.400000,0.400000},{0.975000,0.575000,0.200000,0.200000},{0.975000,0.575000,0.400000,0.400000},{0.025000,0.625000,0.200000,0.200000},{0.025000,0.625000,0.400000,0.400000},{0.075000,0.625000,0.200000,0.200000},{0.075000,0.625000,0.400000,0.400000},{0.125000,0.625000,0.200000,0.200000},{0.125000,0.625000,0.400000,0.400000},{0.175000,0.625000,0.200000,0.200000},{0.175000,0.625000,0.400000,0.400000},{0.225000,0.625000,0.200000,0.200000},{0.225000,0.625000,0.400000,0.400000},{0.275000,0.625000,0.200000,0.200000},{0.275000,0.625000,0.400000,0.400000},{0.325000,0.625000,0.200000,0.200000},{0.325000,0.625000,0.400000,0.400000},{0.375000,0.625000,0.200000,0.200000},{0.375000,0.625000,0.400000,0.400000},{0.425000,0.625000,0.200000,0.200000},{0.425000,0.625000,0.400000,0.400000},{0.475000,0.625000,0.200000,0.200000},{0.475000,0.625000,0.400000,0.400000},{0.525000,0.625000,0.200000,0.200000},{0.525000,0.625000,0.400000,0.400000},{0.575000,0.625000,0.200000,0.200000},{0.575000,0.625000,0.400000,0.400000},{0.625000,0.625000,0.200000,0.200000},{0.625000,0.625000,0.400000,0.400000},{0.675000,0.625000,0.200000,0.200000},{0.675000,0.625000,0.400000,0.400000},{0.725000,0.625000,0.200000,0.200000},{0.725000,0.625000,0.400000,0.400000},{0.775000,0.625000,0.200000,0.200000},{0.775000,0.625000,0.400000,0.400000},{0.825000,0.625000,0.200000,0.200000},{0.825000,0.625000,0.400000,0.400000},{0.875000,0.625000,0.200000,0.200000},{0.875000,0.625000,0.400000,0.400000},{0.925000,0.625000,0.200000,0.200000},{0.925000,0.625000,0.400000,0.400000},{0.975000,0.625000,0.200000,0.200000},{0.975000,0.625000,0.400000,0.400000},{0.025000,0.675000,0.200000,0.200000},{0.025000,0.675000,0.400000,0.400000},{0.075000,0.675000,0.200000,0.200000},{0.075000,0.675000,0.400000,0.400000},{0.125000,0.675000,0.200000,0.200000},{0.125000,0.675000,0.400000,0.400000},{0.175000,0.675000,0.200000,0.200000},{0.175000,0.675000,0.400000,0.400000},{0.225000,0.675000,0.200000,0.200000},{0.225000,0.675000,0.400000,0.400000},{0.275000,0.675000,0.200000,0.200000},{0.275000,0.675000,0.400000,0.400000},{0.325000,0.675000,0.200000,0.200000},{0.325000,0.675000,0.400000,0.400000},{0.375000,0.675000,0.200000,0.200000},{0.375000,0.675000,0.400000,0.400000},{0.425000,0.675000,0.200000,0.200000},{0.425000,0.675000,0.400000,0.400000},{0.475000,0.675000,0.200000,0.200000},{0.475000,0.675000,0.400000,0.400000},{0.525000,0.675000,0.200000,0.200000},{0.525000,0.675000,0.400000,0.400000},{0.575000,0.675000,0.200000,0.200000},{0.575000,0.675000,0.400000,0.400000},{0.625000,0.675000,0.200000,0.200000},{0.625000,0.675000,0.400000,0.400000},{0.675000,0.675000,0.200000,0.200000},{0.675000,0.675000,0.400000,0.400000},{0.725000,0.675000,0.200000,0.200000},{0.725000,0.675000,0.400000,0.400000},{0.775000,0.675000,0.200000,0.200000},{0.775000,0.675000,0.400000,0.400000},{0.825000,0.675000,0.200000,0.200000},{0.825000,0.675000,0.400000,0.400000},{0.875000,0.675000,0.200000,0.200000},{0.875000,0.675000,0.400000,0.400000},{0.925000,0.675000,0.200000,0.200000},{0.925000,0.675000,0.400000,0.400000},{0.975000,0.675000,0.200000,0.200000},{0.975000,0.675000,0.400000,0.400000},{0.025000,0.725000,0.200000,0.200000},{0.025000,0.725000,0.400000,0.400000},{0.075000,0.725000,0.200000,0.200000},{0.075000,0.725000,0.400000,0.400000},{0.125000,0.725000,0.200000,0.200000},{0.125000,0.725000,0.400000,0.400000},{0.175000,0.725000,0.200000,0.200000},{0.175000,0.725000,0.400000,0.400000},{0.225000,0.725000,0.200000,0.200000},{0.225000,0.725000,0.400000,0.400000},{0.275000,0.725000,0.200000,0.200000},{0.275000,0.725000,0.400000,0.400000},{0.325000,0.725000,0.200000,0.200000},{0.325000,0.725000,0.400000,0.400000},{0.375000,0.725000,0.200000,0.200000},{0.375000,0.725000,0.400000,0.400000},{0.425000,0.725000,0.200000,0.200000},{0.425000,0.725000,0.400000,0.400000},{0.475000,0.725000,0.200000,0.200000},{0.475000,0.725000,0.400000,0.400000},{0.525000,0.725000,0.200000,0.200000},{0.525000,0.725000,0.400000,0.400000},{0.575000,0.725000,0.200000,0.200000},{0.575000,0.725000,0.400000,0.400000},{0.625000,0.725000,0.200000,0.200000},{0.625000,0.725000,0.400000,0.400000},{0.675000,0.725000,0.200000,0.200000},{0.675000,0.725000,0.400000,0.400000},{0.725000,0.725000,0.200000,0.200000},{0.725000,0.725000,0.400000,0.400000},{0.775000,0.725000,0.200000,0.200000},{0.775000,0.725000,0.400000,0.400000},{0.825000,0.725000,0.200000,0.200000},{0.825000,0.725000,0.400000,0.400000},{0.875000,0.725000,0.200000,0.200000},{0.875000,0.725000,0.400000,0.400000},{0.925000,0.725000,0.200000,0.200000},{0.925000,0.725000,0.400000,0.400000},{0.975000,0.725000,0.200000,0.200000},{0.975000,0.725000,0.400000,0.400000},{0.025000,0.775000,0.200000,0.200000},{0.025000,0.775000,0.400000,0.400000},{0.075000,0.775000,0.200000,0.200000},{0.075000,0.775000,0.400000,0.400000},{0.125000,0.775000,0.200000,0.200000},{0.125000,0.775000,0.400000,0.400000},{0.175000,0.775000,0.200000,0.200000},{0.175000,0.775000,0.400000,0.400000},{0.225000,0.775000,0.200000,0.200000},{0.225000,0.775000,0.400000,0.400000},{0.275000,0.775000,0.200000,0.200000},{0.275000,0.775000,0.400000,0.400000},{0.325000,0.775000,0.200000,0.200000},{0.325000,0.775000,0.400000,0.400000},{0.375000,0.775000,0.200000,0.200000},{0.375000,0.775000,0.400000,0.400000},{0.425000,0.775000,0.200000,0.200000},{0.425000,0.775000,0.400000,0.400000},{0.475000,0.775000,0.200000,0.200000},{0.475000,0.775000,0.400000,0.400000},{0.525000,0.775000,0.200000,0.200000},{0.525000,0.775000,0.400000,0.400000},{0.575000,0.775000,0.200000,0.200000},{0.575000,0.775000,0.400000,0.400000},{0.625000,0.775000,0.200000,0.200000},{0.625000,0.775000,0.400000,0.400000},{0.675000,0.775000,0.200000,0.200000},{0.675000,0.775000,0.400000,0.400000},{0.725000,0.775000,0.200000,0.200000},{0.725000,0.775000,0.400000,0.400000},{0.775000,0.775000,0.200000,0.200000},{0.775000,0.775000,0.400000,0.400000},{0.825000,0.775000,0.200000,0.200000},{0.825000,0.775000,0.400000,0.400000},{0.875000,0.775000,0.200000,0.200000},{0.875000,0.775000,0.400000,0.400000},{0.925000,0.775000,0.200000,0.200000},{0.925000,0.775000,0.400000,0.400000},{0.975000,0.775000,0.200000,0.200000},{0.975000,0.775000,0.400000,0.400000},{0.025000,0.825000,0.200000,0.200000},{0.025000,0.825000,0.400000,0.400000},{0.075000,0.825000,0.200000,0.200000},{0.075000,0.825000,0.400000,0.400000},{0.125000,0.825000,0.200000,0.200000},{0.125000,0.825000,0.400000,0.400000},{0.175000,0.825000,0.200000,0.200000},{0.175000,0.825000,0.400000,0.400000},{0.225000,0.825000,0.200000,0.200000},{0.225000,0.825000,0.400000,0.400000},{0.275000,0.825000,0.200000,0.200000},{0.275000,0.825000,0.400000,0.400000},{0.325000,0.825000,0.200000,0.200000},{0.325000,0.825000,0.400000,0.400000},{0.375000,0.825000,0.200000,0.200000},{0.375000,0.825000,0.400000,0.400000},{0.425000,0.825000,0.200000,0.200000},{0.425000,0.825000,0.400000,0.400000},{0.475000,0.825000,0.200000,0.200000},{0.475000,0.825000,0.400000,0.400000},{0.525000,0.825000,0.200000,0.200000},{0.525000,0.825000,0.400000,0.400000},{0.575000,0.825000,0.200000,0.200000},{0.575000,0.825000,0.400000,0.400000},{0.625000,0.825000,0.200000,0.200000},{0.625000,0.825000,0.400000,0.400000},{0.675000,0.825000,0.200000,0.200000},{0.675000,0.825000,0.400000,0.400000},{0.725000,0.825000,0.200000,0.200000},{0.725000,0.825000,0.400000,0.400000},{0.775000,0.825000,0.200000,0.200000},{0.775000,0.825000,0.400000,0.400000},{0.825000,0.825000,0.200000,0.200000},{0.825000,0.825000,0.400000,0.400000},{0.875000,0.825000,0.200000,0.200000},{0.875000,0.825000,0.400000,0.400000},{0.925000,0.825000,0.200000,0.200000},{0.925000,0.825000,0.400000,0.400000},{0.975000,0.825000,0.200000,0.200000},{0.975000,0.825000,0.400000,0.400000},{0.025000,0.875000,0.200000,0.200000},{0.025000,0.875000,0.400000,0.400000},{0.075000,0.875000,0.200000,0.200000},{0.075000,0.875000,0.400000,0.400000},{0.125000,0.875000,0.200000,0.200000},{0.125000,0.875000,0.400000,0.400000},{0.175000,0.875000,0.200000,0.200000},{0.175000,0.875000,0.400000,0.400000},{0.225000,0.875000,0.200000,0.200000},{0.225000,0.875000,0.400000,0.400000},{0.275000,0.875000,0.200000,0.200000},{0.275000,0.875000,0.400000,0.400000},{0.325000,0.875000,0.200000,0.200000},{0.325000,0.875000,0.400000,0.400000},{0.375000,0.875000,0.200000,0.200000},{0.375000,0.875000,0.400000,0.400000},{0.425000,0.875000,0.200000,0.200000},{0.425000,0.875000,0.400000,0.400000},{0.475000,0.875000,0.200000,0.200000},{0.475000,0.875000,0.400000,0.400000},{0.525000,0.875000,0.200000,0.200000},{0.525000,0.875000,0.400000,0.400000},{0.575000,0.875000,0.200000,0.200000},{0.575000,0.875000,0.400000,0.400000},{0.625000,0.875000,0.200000,0.200000},{0.625000,0.875000,0.400000,0.400000},{0.675000,0.875000,0.200000,0.200000},{0.675000,0.875000,0.400000,0.400000},{0.725000,0.875000,0.200000,0.200000},{0.725000,0.875000,0.400000,0.400000},{0.775000,0.875000,0.200000,0.200000},{0.775000,0.875000,0.400000,0.400000},{0.825000,0.875000,0.200000,0.200000},{0.825000,0.875000,0.400000,0.400000},{0.875000,0.875000,0.200000,0.200000},{0.875000,0.875000,0.400000,0.400000},{0.925000,0.875000,0.200000,0.200000},{0.925000,0.875000,0.400000,0.400000},{0.975000,0.875000,0.200000,0.200000},{0.975000,0.875000,0.400000,0.400000},{0.025000,0.925000,0.200000,0.200000},{0.025000,0.925000,0.400000,0.400000},{0.075000,0.925000,0.200000,0.200000},{0.075000,0.925000,0.400000,0.400000},{0.125000,0.925000,0.200000,0.200000},{0.125000,0.925000,0.400000,0.400000},{0.175000,0.925000,0.200000,0.200000},{0.175000,0.925000,0.400000,0.400000},{0.225000,0.925000,0.200000,0.200000},{0.225000,0.925000,0.400000,0.400000},{0.275000,0.925000,0.200000,0.200000},{0.275000,0.925000,0.400000,0.400000},{0.325000,0.925000,0.200000,0.200000},{0.325000,0.925000,0.400000,0.400000},{0.375000,0.925000,0.200000,0.200000},{0.375000,0.925000,0.400000,0.400000},{0.425000,0.925000,0.200000,0.200000},{0.425000,0.925000,0.400000,0.400000},{0.475000,0.925000,0.200000,0.200000},{0.475000,0.925000,0.400000,0.400000},{0.525000,0.925000,0.200000,0.200000},{0.525000,0.925000,0.400000,0.400000},{0.575000,0.925000,0.200000,0.200000},{0.575000,0.925000,0.400000,0.400000},{0.625000,0.925000,0.200000,0.200000},{0.625000,0.925000,0.400000,0.400000},{0.675000,0.925000,0.200000,0.200000},{0.675000,0.925000,0.400000,0.400000},{0.725000,0.925000,0.200000,0.200000},{0.725000,0.925000,0.400000,0.400000},{0.775000,0.925000,0.200000,0.200000},{0.775000,0.925000,0.400000,0.400000},{0.825000,0.925000,0.200000,0.200000},{0.825000,0.925000,0.400000,0.400000},{0.875000,0.925000,0.200000,0.200000},{0.875000,0.925000,0.400000,0.400000},{0.925000,0.925000,0.200000,0.200000},{0.925000,0.925000,0.400000,0.400000},{0.975000,0.925000,0.200000,0.200000},{0.975000,0.925000,0.400000,0.400000},{0.025000,0.975000,0.200000,0.200000},{0.025000,0.975000,0.400000,0.400000},{0.075000,0.975000,0.200000,0.200000},{0.075000,0.975000,0.400000,0.400000},{0.125000,0.975000,0.200000,0.200000},{0.125000,0.975000,0.400000,0.400000},{0.175000,0.975000,0.200000,0.200000},{0.175000,0.975000,0.400000,0.400000},{0.225000,0.975000,0.200000,0.200000},{0.225000,0.975000,0.400000,0.400000},{0.275000,0.975000,0.200000,0.200000},{0.275000,0.975000,0.400000,0.400000},{0.325000,0.975000,0.200000,0.200000},{0.325000,0.975000,0.400000,0.400000},{0.375000,0.975000,0.200000,0.200000},{0.375000,0.975000,0.400000,0.400000},{0.425000,0.975000,0.200000,0.200000},{0.425000,0.975000,0.400000,0.400000},{0.475000,0.975000,0.200000,0.200000},{0.475000,0.975000,0.400000,0.400000},{0.525000,0.975000,0.200000,0.200000},{0.525000,0.975000,0.400000,0.400000},{0.575000,0.975000,0.200000,0.200000},{0.575000,0.975000,0.400000,0.400000},{0.625000,0.975000,0.200000,0.200000},{0.625000,0.975000,0.400000,0.400000},{0.675000,0.975000,0.200000,0.200000},{0.675000,0.975000,0.400000,0.400000},{0.725000,0.975000,0.200000,0.200000},{0.725000,0.975000,0.400000,0.400000},{0.775000,0.975000,0.200000,0.200000},{0.775000,0.975000,0.400000,0.400000},{0.825000,0.975000,0.200000,0.200000},{0.825000,0.975000,0.400000,0.400000},{0.875000,0.975000,0.200000,0.200000},{0.875000,0.975000,0.400000,0.400000},{0.925000,0.975000,0.200000,0.200000},{0.925000,0.975000,0.400000,0.400000},{0.975000,0.975000,0.200000,0.200000},{0.975000,0.975000,0.400000,0.400000},{0.050000,0.050000,0.800000,0.800000},{0.050000,0.050000,1.600000,1.600000},{0.150000,0.050000,0.800000,0.800000},{0.150000,0.050000,1.600000,1.600000},{0.250000,0.050000,0.800000,0.800000},{0.250000,0.050000,1.600000,1.600000},{0.350000,0.050000,0.800000,0.800000},{0.350000,0.050000,1.600000,1.600000},{0.450000,0.050000,0.800000,0.800000},{0.450000,0.050000,1.600000,1.600000},{0.550000,0.050000,0.800000,0.800000},{0.550000,0.050000,1.600000,1.600000},{0.650000,0.050000,0.800000,0.800000},{0.650000,0.050000,1.600000,1.600000},{0.750000,0.050000,0.800000,0.800000},{0.750000,0.050000,1.600000,1.600000},{0.850000,0.050000,0.800000,0.800000},{0.850000,0.050000,1.600000,1.600000},{0.950000,0.050000,0.800000,0.800000},{0.950000,0.050000,1.600000,1.600000},{0.050000,0.150000,0.800000,0.800000},{0.050000,0.150000,1.600000,1.600000},{0.150000,0.150000,0.800000,0.800000},{0.150000,0.150000,1.600000,1.600000},{0.250000,0.150000,0.800000,0.800000},{0.250000,0.150000,1.600000,1.600000},{0.350000,0.150000,0.800000,0.800000},{0.350000,0.150000,1.600000,1.600000},{0.450000,0.150000,0.800000,0.800000},{0.450000,0.150000,1.600000,1.600000},{0.550000,0.150000,0.800000,0.800000},{0.550000,0.150000,1.600000,1.600000},{0.650000,0.150000,0.800000,0.800000},{0.650000,0.150000,1.600000,1.600000},{0.750000,0.150000,0.800000,0.800000},{0.750000,0.150000,1.600000,1.600000},{0.850000,0.150000,0.800000,0.800000},{0.850000,0.150000,1.600000,1.600000},{0.950000,0.150000,0.800000,0.800000},{0.950000,0.150000,1.600000,1.600000},{0.050000,0.250000,0.800000,0.800000},{0.050000,0.250000,1.600000,1.600000},{0.150000,0.250000,0.800000,0.800000},{0.150000,0.250000,1.600000,1.600000},{0.250000,0.250000,0.800000,0.800000},{0.250000,0.250000,1.600000,1.600000},{0.350000,0.250000,0.800000,0.800000},{0.350000,0.250000,1.600000,1.600000},{0.450000,0.250000,0.800000,0.800000},{0.450000,0.250000,1.600000,1.600000},{0.550000,0.250000,0.800000,0.800000},{0.550000,0.250000,1.600000,1.600000},{0.650000,0.250000,0.800000,0.800000},{0.650000,0.250000,1.600000,1.600000},{0.750000,0.250000,0.800000,0.800000},{0.750000,0.250000,1.600000,1.600000},{0.850000,0.250000,0.800000,0.800000},{0.850000,0.250000,1.600000,1.600000},{0.950000,0.250000,0.800000,0.800000},{0.950000,0.250000,1.600000,1.600000},{0.050000,0.350000,0.800000,0.800000},{0.050000,0.350000,1.600000,1.600000},{0.150000,0.350000,0.800000,0.800000},{0.150000,0.350000,1.600000,1.600000},{0.250000,0.350000,0.800000,0.800000},{0.250000,0.350000,1.600000,1.600000},{0.350000,0.350000,0.800000,0.800000},{0.350000,0.350000,1.600000,1.600000},{0.450000,0.350000,0.800000,0.800000},{0.450000,0.350000,1.600000,1.600000},{0.550000,0.350000,0.800000,0.800000},{0.550000,0.350000,1.600000,1.600000},{0.650000,0.350000,0.800000,0.800000},{0.650000,0.350000,1.600000,1.600000},{0.750000,0.350000,0.800000,0.800000},{0.750000,0.350000,1.600000,1.600000},{0.850000,0.350000,0.800000,0.800000},{0.850000,0.350000,1.600000,1.600000},{0.950000,0.350000,0.800000,0.800000},{0.950000,0.350000,1.600000,1.600000},{0.050000,0.450000,0.800000,0.800000},{0.050000,0.450000,1.600000,1.600000},{0.150000,0.450000,0.800000,0.800000},{0.150000,0.450000,1.600000,1.600000},{0.250000,0.450000,0.800000,0.800000},{0.250000,0.450000,1.600000,1.600000},{0.350000,0.450000,0.800000,0.800000},{0.350000,0.450000,1.600000,1.600000},{0.450000,0.450000,0.800000,0.800000},{0.450000,0.450000,1.600000,1.600000},{0.550000,0.450000,0.800000,0.800000},{0.550000,0.450000,1.600000,1.600000},{0.650000,0.450000,0.800000,0.800000},{0.650000,0.450000,1.600000,1.600000},{0.750000,0.450000,0.800000,0.800000},{0.750000,0.450000,1.600000,1.600000},{0.850000,0.450000,0.800000,0.800000},{0.850000,0.450000,1.600000,1.600000},{0.950000,0.450000,0.800000,0.800000},{0.950000,0.450000,1.600000,1.600000},{0.050000,0.550000,0.800000,0.800000},{0.050000,0.550000,1.600000,1.600000},{0.150000,0.550000,0.800000,0.800000},{0.150000,0.550000,1.600000,1.600000},{0.250000,0.550000,0.800000,0.800000},{0.250000,0.550000,1.600000,1.600000},{0.350000,0.550000,0.800000,0.800000},{0.350000,0.550000,1.600000,1.600000},{0.450000,0.550000,0.800000,0.800000},{0.450000,0.550000,1.600000,1.600000},{0.550000,0.550000,0.800000,0.800000},{0.550000,0.550000,1.600000,1.600000},{0.650000,0.550000,0.800000,0.800000},{0.650000,0.550000,1.600000,1.600000},{0.750000,0.550000,0.800000,0.800000},{0.750000,0.550000,1.600000,1.600000},{0.850000,0.550000,0.800000,0.800000},{0.850000,0.550000,1.600000,1.600000},{0.950000,0.550000,0.800000,0.800000},{0.950000,0.550000,1.600000,1.600000},{0.050000,0.650000,0.800000,0.800000},{0.050000,0.650000,1.600000,1.600000},{0.150000,0.650000,0.800000,0.800000},{0.150000,0.650000,1.600000,1.600000},{0.250000,0.650000,0.800000,0.800000},{0.250000,0.650000,1.600000,1.600000},{0.350000,0.650000,0.800000,0.800000},{0.350000,0.650000,1.600000,1.600000},{0.450000,0.650000,0.800000,0.800000},{0.450000,0.650000,1.600000,1.600000},{0.550000,0.650000,0.800000,0.800000},{0.550000,0.650000,1.600000,1.600000},{0.650000,0.650000,0.800000,0.800000},{0.650000,0.650000,1.600000,1.600000},{0.750000,0.650000,0.800000,0.800000},{0.750000,0.650000,1.600000,1.600000},{0.850000,0.650000,0.800000,0.800000},{0.850000,0.650000,1.600000,1.600000},{0.950000,0.650000,0.800000,0.800000},{0.950000,0.650000,1.600000,1.600000},{0.050000,0.750000,0.800000,0.800000},{0.050000,0.750000,1.600000,1.600000},{0.150000,0.750000,0.800000,0.800000},{0.150000,0.750000,1.600000,1.600000},{0.250000,0.750000,0.800000,0.800000},{0.250000,0.750000,1.600000,1.600000},{0.350000,0.750000,0.800000,0.800000},{0.350000,0.750000,1.600000,1.600000},{0.450000,0.750000,0.800000,0.800000},{0.450000,0.750000,1.600000,1.600000},{0.550000,0.750000,0.800000,0.800000},{0.550000,0.750000,1.600000,1.600000},{0.650000,0.750000,0.800000,0.800000},{0.650000,0.750000,1.600000,1.600000},{0.750000,0.750000,0.800000,0.800000},{0.750000,0.750000,1.600000,1.600000},{0.850000,0.750000,0.800000,0.800000},{0.850000,0.750000,1.600000,1.600000},{0.950000,0.750000,0.800000,0.800000},{0.950000,0.750000,1.600000,1.600000},{0.050000,0.850000,0.800000,0.800000},{0.050000,0.850000,1.600000,1.600000},{0.150000,0.850000,0.800000,0.800000},{0.150000,0.850000,1.600000,1.600000},{0.250000,0.850000,0.800000,0.800000},{0.250000,0.850000,1.600000,1.600000},{0.350000,0.850000,0.800000,0.800000},{0.350000,0.850000,1.600000,1.600000},{0.450000,0.850000,0.800000,0.800000},{0.450000,0.850000,1.600000,1.600000},{0.550000,0.850000,0.800000,0.800000},{0.550000,0.850000,1.600000,1.600000},{0.650000,0.850000,0.800000,0.800000},{0.650000,0.850000,1.600000,1.600000},{0.750000,0.850000,0.800000,0.800000},{0.750000,0.850000,1.600000,1.600000},{0.850000,0.850000,0.800000,0.800000},{0.850000,0.850000,1.600000,1.600000},{0.950000,0.850000,0.800000,0.800000},{0.950000,0.850000,1.600000,1.600000},{0.050000,0.950000,0.800000,0.800000},{0.050000,0.950000,1.600000,1.600000},{0.150000,0.950000,0.800000,0.800000},{0.150000,0.950000,1.600000,1.600000},{0.250000,0.950000,0.800000,0.800000},{0.250000,0.950000,1.600000,1.600000},{0.350000,0.950000,0.800000,0.800000},{0.350000,0.950000,1.600000,1.600000},{0.450000,0.950000,0.800000,0.800000},{0.450000,0.950000,1.600000,1.600000},{0.550000,0.950000,0.800000,0.800000},{0.550000,0.950000,1.600000,1.600000},{0.650000,0.950000,0.800000,0.800000},{0.650000,0.950000,1.600000,1.600000},{0.750000,0.950000,0.800000,0.800000},{0.750000,0.950000,1.600000,1.600000},{0.850000,0.950000,0.800000,0.800000},{0.850000,0.950000,1.600000,1.600000},{0.950000,0.950000,0.800000,0.800000},{0.950000,0.950000,1.600000,1.600000}};
