--- package/ai/code/shell/object_detect_v4l2_bf16.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-devmem 0x970E00fc 32 0x0fffff00
-devmem 0x970E0100 32 0x000000ff
-devmem 0x970E00f4 32 0x00550000
-
-cd ../exe && ./object_detect_v4l2 ../kmodel/kmodel_compile/object_detect/od_bf16.kmodel 0
